% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R, R/NMFSet-class.R
\name{advanced-NMF}
\alias{advanced-NMF}
\alias{which.best}
\title{Advanced Usage of the Package NMF}
\usage{
which.best(object, FUN = deviance, ...)
}
\arguments{
\item{object}{an NMF model fitted by multiple runs.}

\item{FUN}{the function that computes the quantitative measure.}

\item{...}{extra arguments passed to \code{FUN}.}
}
\description{
The functions documented here provide advanced functionalities useful when
developing within the framework implemented in the NMF package.
}
