% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.measure}
\alias{aggregate.measure}
\title{Utility function to aggregate numerical quality measures from \code{NMFfitXn} objects.}
\usage{
\method{aggregate}{measure}(x, method = c("best", "mean"), decreasing = FALSE, ...)
}
\arguments{
\item{x}{a numerical vector}

\item{method}{the method to aggregate values. This argument can take two values :
- mean: the mean of the measures
- best: the best measure according to the specified sorting order (decreasing or not)}

\item{decreasing}{logical that specified the sorting order}

\item{...}{extra arguments to allow extension}
}
\description{
Given a numerical vector, this function computes an aggregated value using one of the following methods:
best or mean
}
