% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{algorithm,NMFfit-method}
\alias{algorithm,NMFfit-method}
\title{Returns the name of the algorithm that fitted the NMF model \code{object}.}
\usage{
\S4method{algorithm}{NMFfit}(object)
}
\description{
Returns the name of the algorithm that fitted the NMF model \code{object}.
}
