% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSeed-class.R
\docType{methods}
\name{algorithm,NMFSeed-method}
\alias{algorithm,NMFSeed-method}
\title{Returns the workhorse function of the seeding method described by \code{object}.}
\usage{
\S4method{algorithm}{NMFSeed}(object)
}
\description{
Returns the workhorse function of the seeding method described by \code{object}.
}
