% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategyFunction-class.R
\docType{methods}
\name{algorithm,NMFStrategyFunction-method}
\alias{algorithm,NMFStrategyFunction-method}
\title{Returns the single R function that implements the NMF algorithm -- as stored in 
slot \code{algorithm}.}
\usage{
\S4method{algorithm}{NMFStrategyFunction}(object)
}
\description{
Returns the single R function that implements the NMF algorithm -- as stored in 
slot \code{algorithm}.
}
