% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSeed-class.R
\docType{methods}
\name{algorithm<-,NMFSeed,function-method}
\alias{algorithm<-,NMFSeed,function-method}
\title{Sets the workhorse function of the seeding method described by \code{object}.}
\usage{
\S4method{algorithm}{NMFSeed,`function`}(object) <- value
}
\description{
Sets the workhorse function of the seeding method described by \code{object}.
}
