% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategyFunction-class.R
\docType{methods}
\name{algorithm<-,NMFStrategyFunction,function-method}
\alias{algorithm<-,NMFStrategyFunction,function-method}
\title{Sets the function that implements the NMF algorithm, stored in slot \code{algorithm}.}
\usage{
\S4method{algorithm}{NMFStrategyFunction,`function`}(object) <- value
}
\description{
Sets the function that implements the NMF algorithm, stored in slot \code{algorithm}.
}
