% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategyOctave-class.R
\docType{methods}
\name{algorithm<-,NMFStrategyOctave,character-method}
\alias{algorithm<-,NMFStrategyOctave,character-method}
\title{Sets the name of the Octave/Matlab function that implements the NMF algorithm.
It is stored in slot \code{algorithm}.}
\usage{
\S4method{algorithm}{NMFStrategyOctave,character}(object) <- value
}
\description{
Sets the name of the Octave/Matlab function that implements the NMF algorithm.
It is stored in slot \code{algorithm}.
}
