% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary-NMF}
\alias{summary,NMF-method}
\title{Assessing and Comparing NMF Models}
\usage{
summary(object, ...)

\S4method{summary}{NMF}(object, class, target)
}
\arguments{
\item{object}{an NMF object. See available methods in section \emph{Methods}.}

\item{...}{extra arguments passed to the next \code{summary} method.}

\item{class}{known classes/cluster of samples specified in one of the formats
that is supported by the functions \code{\link{entropy}} and \code{\link{purity}}.}

\item{target}{target matrix specified in one of the formats supported by the
functions \code{\link{rss}} and \code{\link{evar}}}
}
\description{
The NMF package defines \code{summary} methods for different classes of objects,
which helps assessing and comparing the quality of NMF models by computing a set
of quantitative measures, e.g. with respect to their ability to recover known
classes and/or the original target matrix.

The most useful methods are for classes \code{\linkS4class{NMF}}, \code{\linkS4class{NMFfit}},
\code{\linkS4class{NMFfitX}} and \code{\linkS4class{NMFList}}, which compute summary measures
for, respectively, a single NMF model, a single fit, a multiple-run fit and a list of heterogenous
fits performed with the function \code{\link{nmf}}.

The following measures are computed:
}
\details{
Due to the somehow hierarchical structure of the classes mentionned in \emph{Description},
their respective \code{summary} methods call each other in chain, each super-class adding some
extra measures, only relevant for objects of a specific class.

\describe{
\item{sparseness}{Sparseness of the factorization computed by the
function \code{\link{sparseness}}.}
\item{entropy}{Purity of the clustering, with respect to known classes,
computed by the function \code{\link{purity}}.}
\item{entropy}{Entropy of the clustering, with respect to known classes,
computed by the function \code{\link{entropy}}.}
\item{RSS}{Residual Sum of Squares computed by the function \code{\link{rss}}.}
\item{evar}{Explained variance computed by the function \code{\link{evar}}.}
}
}
\examples{

# random NMF model
x <- rnmf(3, 20, 12)
summary(x)
summary(x, gl(3, 4))
summary(x, target=rmatrix(x))
summary(x, gl(3,4), target=rmatrix(x))

}
