% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{basis,ANY-method}
\alias{basis,ANY-method}
\title{Default method returns the value of S3 slot or attribute \code{'basis'}.
It returns \code{NULL} if none of these are set.}
\usage{
\S4method{basis}{ANY}(object, ...)
}
\description{
Arguments \code{...} are not used by this method.
}
