% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\name{basis}
\alias{basis}
\alias{basis,ANY-method}
\alias{.basis}
\alias{basis<-}
\alias{.basis<-}
\alias{loadings,NMF-method}
\alias{.coef}
\alias{coef<-}
\alias{.coef<-}
\alias{coefficients}
\alias{scoef}
\alias{scoef,matrix-method}
\title{Accessing NMF Factors}
\usage{
basis(object, ...)

\S4method{basis}{ANY}(object, ...)

.basis(object, ...)

basis(object, ...) <- value

.basis(object, ...) <- value

\S4method{loadings}{NMF}(x)

.coef(object, ...)

coef(object, ...) <- value

.coef(object, ...) <- value

coefficients(object, ...)

scoef(object, ...)

\S4method{scoef}{matrix}(object, scale = 1)
}
\arguments{
\item{object}{an object from which to extract the factor matrices, typically an
object of class \code{\linkS4class{NMF}}.}

\item{...}{extra arguments to allow extension and passed to the low-level
access functions \code{.coef} and \code{.basis}.

Note that these throw an error if used in replacement functions \code{}.}

\item{value}{replacement value}

\item{x}{an object of class \linkS4class{NMF}}

\item{scale}{scaling factor, which indicates to the value the columns of the
coefficient matrix should sum up to.}
}
\description{
\code{basis} and \code{basis<-} are S4 generic functions which respectively
extract and set the matrix of basis components of an NMF model
(i.e. the first matrix factor).

The methods \code{.basis}, \code{.coef} and their replacement versions
are implemented as pure virtual methods for the interface class
\code{NMF}, meaning that concrete NMF models must provide a
definition for their corresponding class (i.e. sub-classes of
class \code{NMF}).
See \code{\linkS4class{NMF}} for more details.

The method \code{loadings} is identical to \code{basis}, but do
not accept any extra argument.

Methods \code{coefficients} and \code{coefficients<-} are
simple aliases for methods \code{coef} and \code{coef<-} respectively.

\code{scoef} is similar to \code{coef}, but returns the mixture
coefficient matrix of an NMF model, with the columns scaled so that they
sum up to a given value (1 by default).
}
\details{
For example, in the case of the standard NMF model \eqn{V \equiv W H}{V ~ W H},
the method \code{basis} will return the matrix \eqn{W}.

\code{basis} and \code{basis<-} are defined for the top
virtual class \code{\linkS4class{NMF}} only, and rely internally on the low-level
S4 generics \code{.basis} and \code{.basis<-} respectively that effectively
extract/set the coefficient data.
These data are post/pre-processed, e.g., to extract/set only their
non-fixed terms or check dimension compatibility.

The method \code{loadings} is provided to standardise the NMF interface
against the one defined in the \code{\link{stats}} package,
and emphasises the similarities between NMF and PCA or factorial analysis
(see \code{\link{loadings}}).
}
\examples{

# Scaled coefficient matrix
x <- rnmf(3, 10, 5)
scoef(x)
scoef(x, 100)

}
\seealso{
Other NMF-interface: 
\code{\link{NMF-class}},
\code{\link{nmfModel}()},
\code{\link{rnmf}()}
}
\concept{NMF-interface}
\section{Methods (by generic)}{
\bold{.basis:}

\itemize{
\item \code{.basis(object = NMF)}: Pure virtual method for objects of class \code{\linkS4class{NMF}}, that
should be overloaded by sub-classes, and throws an error if called.

\item \code{.basis(object = NMFstd)}: Get the basis matrix in standard NMF models

This function returns slot \code{W} of \code{object}.

\item \code{.basis(object = NMFfit)}: Returns the basis matrix from an NMF model fitted with
function \code{\link{nmf}}.

It is a shortcut for \code{.basis(fit(object), ...)}, dispatching the call to
the \code{.basis} method of the actual NMF model.
}


\bold{.basis<-:}

\itemize{
\item \code{.basis<-(object = NMF,value = mMatrix)}: Pure virtual method for objects of class \code{\linkS4class{NMF}}, that
should be overloaded by sub-classes, and throws an error if called.

\item \code{.basis<-(object = NMFstd,value = array)}: Set the basis matrix in standard NMF models

This function sets slot \code{W} of \code{object}.

\item \code{.basis<-(object = NMFstd,value = mMatrix)}: Replaces a slice of the basis array.

\item \code{.basis<-(object = NMFfit,value = matrix)}: Sets the the basis matrix of an NMF model fitted with
function \code{\link{nmf}}.

It is a shortcut for \code{.basis(fit(object)) <- value}, dispatching the call to
the \code{.basis<-} method of the actual NMF model.
It is not meant to be used by the user, except when developing
NMF algorithms, to update the basis matrix of the seed object before
returning it.
}


\bold{.coef:}

\itemize{
\item \code{.coef(object = NMF)}: Pure virtual method for objects of class \code{\linkS4class{NMF}}, that
should be overloaded by sub-classes, and throws an error if called.

\item \code{.coef(object = NMFstd)}: Get the mixture coefficient matrix in standard NMF models, which
is stored in slot \code{H} of \code{object}.

\item \code{.coef(object = NMFfit)}: Returns the the coefficient matrix from an NMF model fitted with
function \code{\link{nmf}}.

It is a shortcut for \code{.coef(fit(object), ...)}, dispatching the call to
the \code{.coef} method of the actual NMF model.
}


\bold{.coef<-:}

\itemize{
\item \code{.coef<-(object = NMF,value = mMatrix)}: Pure virtual method for objects of class \code{\linkS4class{NMF}}, that
should be overloaded by sub-classes, and throws an error if called.

\item \code{.coef<-(object = NMFstd,value = array)}: Set the mixture coefficient matrix in standard NMF models

This function sets slot \code{H} of \code{object}.

\item \code{.coef<-(object = NMFstd,value = mMatrix)}: Replaces a slice of the coefficent array.

\item \code{.coef<-(object = NMFfit,value = matrix)}: Sets the the coefficient matrix of an NMF model fitted with
function \code{\link{nmf}}.

It is a shortcut for \code{.coef(fit(object)) <- value}, dispatching the call to
the \code{.coef<-} method of the actual NMF model.
It is not meant to be used by the user, except when developing
NMF algorithms, to update the coefficient matrix in the seed object before
returning it.
}


\bold{basis:}

\itemize{
\item \code{basis(object = NMFfitXn)}: Returns the basis matrix of the best fit amongst all the fits stored in
\code{object}.
It is a shortcut for \code{basis(fit(object))}.

\item \code{basis(object = ANY)}: Default method returns the value of S3 slot or attribute \code{'basis'}.
It returns \code{NULL} if none of these are set.

Arguments \code{...} are not used by this method.
}


\bold{basis<-:}

\itemize{
\item \code{basis<-(object = NMF)}: Default methods that calls \code{.basis<-} and check the validity of the
updated object.
}


\bold{coef<-:}

\itemize{
\item \code{coef<-(object = NMF)}: Default methods that calls \code{.coef<-} and check the validity of the
updated object.
}


\bold{coefficients:}

\itemize{
\item \code{coefficients(object = NMF)}: Alias to \code{coef,NMF}, therefore also pure virtual.
}

}

