% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{basis,NMFfitXn-method}
\alias{basis,NMFfitXn-method}
\title{Returns the basis matrix of the best fit amongst all the fits stored in 
\code{object}.
It is a shortcut for \code{basis(fit(object))}.}
\usage{
\S4method{basis}{NMFfitXn}(object, ...)
}
\description{
Returns the basis matrix of the best fit amongst all the fits stored in 
\code{object}.
It is a shortcut for \code{basis(fit(object))}.
}
