% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{basis<-,NMF-method}
\alias{basis<-,NMF-method}
\title{Default methods that calls \code{.basis<-} and check the validity of the
updated object.}
\usage{
\S4method{basis}{NMF}(object, ..., use.dimnames = TRUE) <- value
}
\arguments{
\item{use.dimnames}{logical that indicates if the object's dim names should be
set using those from the new value, or left unchanged -- after truncating
them to fit new dimensions if necessary.
This is useful to only set the entries of a factor.}
}
\description{
Default methods that calls \code{.basis<-} and check the validity of the
updated object.
}
