% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{basiscor,matrix,NMF-method}
\alias{basiscor,matrix,NMF-method}
\title{Computes the correlations between the columns of \code{x}
and the the basis vectors of \code{y}.}
\usage{
\S4method{basiscor}{matrix,NMF}(x, y, ...)
}
\description{
Computes the correlations between the columns of \code{x}
and the the basis vectors of \code{y}.
}
