% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{basiscor,NMF,missing-method}
\alias{basiscor,NMF,missing-method}
\title{Computes the correlations between the basis vectors of \code{x}.}
\usage{
\S4method{basiscor}{NMF,missing}(x, y, ...)
}
\description{
Computes the correlations between the basis vectors of \code{x}.
}
