% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{basiscor,NMF,NMF-method}
\alias{basiscor,NMF,NMF-method}
\title{Computes the correlations between the basis vectors of \code{x} and \code{y}.}
\usage{
\S4method{basiscor}{NMF,NMF}(x, y, ...)
}
\description{
Computes the correlations between the basis vectors of \code{x} and \code{y}.
}
