% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\name{basiscor}
\alias{basiscor}
\alias{profcor}
\title{Correlations in NMF Models}
\usage{
basiscor(x, y, ...)

profcor(x, y, ...)
}
\arguments{
\item{x}{a matrix or an object with suitable methods \code{\link{basis}}
or \code{\link{coef}}.}

\item{y}{a matrix or an object with suitable methods \code{\link{basis}}
or \code{\link{coef}}, and dimensions compatible with \code{x}.
If missing the correlations are computed between \code{x} and \code{y=x}.}

\item{...}{extra arguments passed to \code{\link{cor}}.}
}
\description{
\code{basiscor} computes the correlation matrix between basis vectors, i.e.
the \emph{columns} of its basis matrix -- which is the model's first matrix factor.

\code{profcor} computes the correlation matrix between basis profiles,
i.e. the \emph{rows} of the coefficient matrix -- which is the model's second
matrix factor.
}
\details{
Each generic has methods defined for computing correlations between NMF models
and/or compatible matrices.
The computation is performed by the base function \code{\link{cor}}.
}
\examples{

# generate two random NMF models
a <- rnmf(3, 100, 20)
b <- rnmf(3, 100, 20)

# Compute auto-correlations
basiscor(a)
profcor(a)
# Compute correlations with b
basiscor(a, b)
profcor(a, b)

# try to recover the underlying NMF model 'a' from noisy data 
res <- nmf(fitted(a) + rmatrix(a), 3)

# Compute correlations with the true model
basiscor(a, res)
profcor(a, res)

# Compute correlations with a random compatible matrix
W <- rmatrix(basis(a))
basiscor(a, W)
identical(basiscor(a, W), basiscor(W, a))

H <- rmatrix(coef(a))
profcor(a, H)
identical(profcor(a, H), profcor(H, a))

}
\concept{NMFplots Plotting functions for NMF objects}
\section{Methods (by generic)}{
\bold{basiscor:}

\itemize{
\item \code{basiscor(x = NMF,y = matrix)}: Computes the correlations between the basis vectors of \code{x} and
the columns of \code{y}.

\item \code{basiscor(x = matrix,y = NMF)}: Computes the correlations between the columns of \code{x}
and the the basis vectors of \code{y}.

\item \code{basiscor(x = NMF,y = NMF)}: Computes the correlations between the basis vectors of \code{x} and \code{y}.

\item \code{basiscor(x = NMF,y = missing)}: Computes the correlations between the basis vectors of \code{x}.
}


\bold{profcor:}

\itemize{
\item \code{profcor(x = NMF,y = matrix)}: Computes the correlations between the basis profiles of \code{x} and
the rows of \code{y}.

\item \code{profcor(x = matrix,y = NMF)}: Computes the correlations between the rows of \code{x} and the basis
profiles of \code{y}.

\item \code{profcor(x = NMF,y = NMF)}: Computes the correlations between the basis profiles of \code{x} and \code{y}.

\item \code{profcor(x = NMF,y = missing)}: Computes the correlations between the basis profiles of \code{x}.
}

}

