% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\docType{methods}
\name{basismap,NMF-method}
\alias{basismap,NMF-method}
\title{Plots a heatmap of the basis matrix of the NMF model \code{object}.
This method also works for fitted NMF models (i.e. \code{NMFfit} objects).}
\usage{
\S4method{basismap}{NMF}(object, color = "YlOrRd:50", scale = "r1",
  Rowv = TRUE, Colv = NA, subsetRow = FALSE, annRow = NA, annCol = NA,
  tracks = "basis", main = "Basis components", info = FALSE, ...)
}
\arguments{
\item{color}{colour specification for the heatmap. Default to palette
'-RdYlBu2:100', i.e. reversed palette 'RdYlBu2' (a slight modification of
RColorBrewer's palette 'RdYlBu') with 100 colors.
Possible values are:
\itemize{
\item a character/integer vector of length greater than 1 that is directly used
and assumed to contain valid R color specifications.
\item a single color/integer (between 0 and 8)/other numeric value
that gives the dominant colors. Numeric values are converted into a pallete
by \code{rev(sequential_hcl(2, h = x, l = c(50, 95)))}. Other values are
concatenated with the grey colour '#F1F1F1'.
\item one of RColorBrewer's palette name (see \code{\link[RColorBrewer]{display.brewer.all}})
, or one of 'RdYlBu2', 'rainbow', 'heat', 'topo', 'terrain', 'cm'.
}
When the colour palette is specified with a single value, and is negative or
preceded a minus ('-'), the reversed palette is used.
The number of breaks can also be specified after a colon (':'). For example,
the default colour palette is specified as '-RdYlBu2:100'.}

\item{scale}{character indicating how the values should scaled in
either the row direction or the column direction. Note that the scaling is
performed after row/column clustering, so that it has no effect on the
row/column ordering.
Possible values are:
\itemize{
\item \code{"row"}: center and standardize each row separately to row Z-scores
\item \code{"column"}: center and standardize each column separately to column Z-scores
\item \code{"r1"}: scale each row to sum up to one
\item \code{"c1"}: scale each column to sum up to one
\item \code{"none"}: no scaling
}}

\item{Rowv}{clustering specification(s) for the rows. It allows to specify
the distance/clustering/ordering/display parameters to be used for the
\emph{rows only}.

See section \emph{Row/column ordering and display} for details on all supported values.}

\item{Colv}{clustering specification(s) for the columns. It accepts the same
values as argument \code{Rowv} (modulo the expected length for vector specifications),
and allow specifying the distance/clustering/ordering/display parameters to
be used for the \emph{columns only}.

\code{Colv} may also be set to \code{"Rowv"}, in which case the dendrogram
or ordering specifications applied to the rows are also applied to the
columns. Note that this is allowed only for square matrices,
and that the row ordering is in this case by default reversed
(\code{revC=TRUE}) to obtain the diagonal in the standard way
(from top-left to bottom-right).

See section \emph{Row/column ordering and display} for details on all supported values.}

\item{subsetRow}{Argument that specifies how to filter the rows that
will appear in the heatmap.  
When \code{FALSE} (default), all rows are used. 
Besides the values supported by argument \code{subsetRow} of 
\code{\link{aheatmap}}, other possible values are:

\itemize{ 
\item \code{TRUE}: only the rows that are basis-specific are used.  
The default selection method is from \cite{KimH2007}.
This is equivalent to \code{subsetRow='kim'}.

\item a single \code{character} string or numeric value that specifies 
the method to use to select the basis-specific rows, that should appear in the
heatmap (cf. argument \code{method} for function \code{\link{extractFeatures}}).

Note \code{\link{extractFeatures}} is called with argument \code{nodups=TRUE}, 
so that features that are selected for multiple components only appear once.
}}

\item{annRow}{specifications of row annotation tracks displayed as coloured
columns on the left of the heatmaps. The annotation tracks are drawn from
left to right. The same conversion, renaming and colouring rules as for argument
\code{annCol} apply.}

\item{annCol}{specifications of column annotation tracks displayed as coloured
rows on top of the heatmaps. The annotation tracks are drawn from bottom to top.
A single annotation track can be specified as a single vector; multiple tracks
are specified as a list, a data frame, or an
\code{\link[Biobase:ExpressionSet-class]{ExpressionSet}} object, in
which case the phenotypic data is used (\code{pData(eset)}).
Character or integer vectors are converted and displayed as factors.
Unnamed tracks are internally renamed into \code{Xi}, with i being incremented for
each unamed track, across both column and row annotation tracks.
For each track, if no corresponding colour is specified in argument
\code{annColors}, a palette or a ramp is automatically computed and named
after the track's name.}

\item{tracks}{Special additional annotation tracks to highlight associations between 
basis components and sample clusters:
\describe{
\item{basis}{matches each row (resp. column) to the most contributing basis component
in \code{basismap} (resp. \code{coefmap}).
In \code{basismap} (resp. \code{coefmap}), adding a track \code{':basis'} to 
\code{annCol} (resp. \code{annRow}) makes the column (resp. row) corresponding to 
the component being also highlited using the mathcing colours.}
}}

\item{main}{Main title as a character string or a grob.}

\item{info}{if \code{TRUE} then the name of the algorithm that fitted the NMF 
model is displayed at the bottom of the plot, if available.
Other wise it is passed as is to \code{aheatmap}.}
}
\description{
Plots a heatmap of the basis matrix of the NMF model \code{object}.
This method also works for fitted NMF models (i.e. \code{NMFfit} objects).
}
