% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{basismap,NMFfitX-method}
\alias{basismap,NMFfitX-method}
\title{Plots a heatmap of the basis matrix of the best fit in \code{object}.}
\usage{
\S4method{basismap}{NMFfitX}(object, ...)
}
\description{
Plots a heatmap of the basis matrix of the best fit in \code{object}.
}
