% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{basisnames,ANY-method}
\alias{basisnames,ANY-method}
\title{Default method which returns the column names of the basis matrix extracted from
\code{x}, using the \code{basis} method.}
\usage{
\S4method{basisnames}{ANY}(x)
}
\description{
For NMF objects these also correspond to the row names of the coefficient matrix.
}
