% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{basisnames<-,ANY-method}
\alias{basisnames<-,ANY-method}
\title{Default method which sets, respectively, the row and the column names of the basis
matrix and coefficient matrix of \code{x} to \code{value}.}
\usage{
\S4method{basisnames}{ANY}(x, ...) <- value
}
\description{
Default method which sets, respectively, the row and the column names of the basis
matrix and coefficient matrix of \code{x} to \code{value}.
}
