% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry-algorithms.R
\docType{methods}
\name{canFit,character,ANY-method}
\alias{canFit,character,ANY-method}
\title{Tells if a registered NMF algorithm can fit a given NMF model}
\usage{
\S4method{canFit}{character,ANY}(x, y, ...)
}
\description{
Tells if a registered NMF algorithm can fit a given NMF model
}
