% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry-algorithms.R
\docType{methods}
\name{canFit,NMFStrategy,character-method}
\alias{canFit,NMFStrategy,character-method}
\title{Tells if an NMF algorithm can fit a given class of NMF models}
\usage{
\S4method{canFit}{NMFStrategy,character}(x, y, exact = FALSE)
}
\arguments{
\item{exact}{for logical that indicates if an algorithm is considered able to fit 
only the models that it explicitly declares (\code{TRUE}), or if it should be
considered able to also fit models that extend models that it explicitly fits.}
}
\description{
Tells if an NMF algorithm can fit a given class of NMF models
}
