% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry-algorithms.R
\docType{methods}
\name{canFit,NMFStrategy,NMF-method}
\alias{canFit,NMFStrategy,NMF-method}
\title{Tells if an NMF algorithm can fit the same class of models as \code{y}}
\usage{
\S4method{canFit}{NMFStrategy,NMF}(x, y, ...)
}
\description{
Tells if an NMF algorithm can fit the same class of models as \code{y}
}
