% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorcode.R
\name{ccBreaks}
\alias{ccBreaks}
\title{Generate Break Intervals from Numeric Variables}
\usage{
ccBreaks(x, breaks, center = NULL)
}
\description{
Implementation is borrowed from the R core function \code{\link{cut.default}}.
}
\keyword{internal}
