% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorcode.R
\name{ccPalette}
\alias{ccPalette}
\title{Builds a Color Palette from Compact Color Specification}
\usage{
ccPalette(x, n = NA, verbose = FALSE)
}
\description{
Builds a Color Palette from Compact Color Specification
}
\keyword{internal}
