% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorcode.R
\name{ccSpec}
\alias{ccSpec}
\title{Extract Colour Palette Specification}
\usage{
ccSpec(x)
}
\arguments{
\item{a}{character string that specify a colour palette.}
}
\value{
a list with elements: palette, n and rev
}
\description{
Extract Colour Palette Specification
}
\keyword{internal}
