% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{coef,NMFfitXn-method}
\alias{coef,NMFfitXn-method}
\title{Returns the coefficient matrix of the best fit amongst all the fits stored in 
\code{object}.
It is a shortcut for \code{coef(fit(object))}.}
\usage{
\S4method{coef}{NMFfitXn}(object, ...)
}
\description{
Returns the coefficient matrix of the best fit amongst all the fits stored in 
\code{object}.
It is a shortcut for \code{coef(fit(object))}.
}
