% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{coef<-,NMF-method}
\alias{coef<-,NMF-method}
\title{Default methods that calls \code{.coef<-} and check the validity of the
updated object.}
\usage{
\S4method{coef}{NMF}(object, ..., use.dimnames = TRUE) <- value
}
\description{
Default methods that calls \code{.coef<-} and check the validity of the
updated object.
}
