% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{coefficients,NMF-method}
\alias{coefficients,NMF-method}
\title{Alias to \code{coef,NMF}, therefore also pure virtual.}
\usage{
\S4method{coefficients}{NMF}(object, ..., all = TRUE)
}
\description{
Alias to \code{coef,NMF}, therefore also pure virtual.
}
