% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\docType{methods}
\name{coefmap,NMF-method}
\alias{coefmap,NMF-method}
\title{The default method for NMF objects has special default values for 
some arguments of \code{\link{aheatmap}} (see argument description).}
\usage{
\S4method{coefmap}{NMF}(object, color = "YlOrRd:50", scale = "c1",
  Rowv = NA, Colv = TRUE, annRow = NA, annCol = NA, tracks = "basis",
  main = "Mixture coefficients", info = FALSE, ...)
}
\description{
The default method for NMF objects has special default values for 
some arguments of \code{\link{aheatmap}} (see argument description).
}
