% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{coefmap,NMFfitX-method}
\alias{coefmap,NMFfitX-method}
\title{Plots a heatmap of the coefficient matrix of the best fit in \code{object}.}
\usage{
\S4method{coefmap}{NMFfitX}(object, Colv = TRUE, annRow = NA, annCol = NA,
  tracks = c("basis", "consensus:"), ...)
}
\description{
This method adds:
\itemize{
\item an extra special column annotation track for multi-run NMF fits,
\code{'consensus:'}, that shows the consensus cluster associated to each sample.
\item a column sorting schema \code{'consensus'} that can be passed
to argument \code{Colv} and orders the columns using the hierarchical clustering of the 
consensus matrix with average linkage, as returned by \code{\link{consensushc}(object)}.
This is also the ordering that is used by default for the heatmap of the consensus matrix 
as ploted by \code{\link{consensusmap}}. 
}
}
