% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{connectivity,ANY-method}
\alias{connectivity,ANY-method}
\title{Default method which computes the connectivity matrix
using the result of \code{predict(x, ...)} as cluster membership index.}
\usage{
\S4method{connectivity}{ANY}(object, ...)
}
\description{
Default method which computes the connectivity matrix
using the result of \code{predict(x, ...)} as cluster membership index.
}
\examples{

# clustering of random data
h <- hclust(dist(rmatrix(10,20)))
connectivity(stats::cutree(h, 2))

}
