% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{connectivity,factor-method}
\alias{connectivity,factor-method}
\title{Computes the connectivity matrix using \code{x} as cluster membership index.}
\usage{
\S4method{connectivity}{factor}(object, ...)
}
\description{
Computes the connectivity matrix using \code{x} as cluster membership index.
}
\examples{
connectivity(gl(2, 4))

}
