% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{connectivity,NMF-method}
\alias{connectivity,NMF-method}
\title{Computes the connectivity matrix for an NMF model, for which cluster
membership is given by the most contributing basis component in each sample.
See \code{\link{predict,NMF-method}}.}
\usage{
\S4method{connectivity}{NMF}(object, no.attrib = FALSE)
}
\arguments{
\item{no.attrib}{a logical that indicates if attributes containing information
about the NMF model should be attached to the result (\code{TRUE}) or not
(\code{FALSE}).}
}
\description{
Computes the connectivity matrix for an NMF model, for which cluster
membership is given by the most contributing basis component in each sample.
See \code{\link{predict,NMF-method}}.
}
