% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{connectivity,numeric-method}
\alias{connectivity,numeric-method}
\title{Equivalent to \code{connectivity(as.factor(x))}.}
\usage{
\S4method{connectivity}{numeric}(object, ...)
}
\description{
Equivalent to \code{connectivity(as.factor(x))}.
}
