% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R, R/NMFSet-class.R
\name{connectivity}
\alias{connectivity}
\alias{connectivity,ANY-method}
\alias{connectivity,factor-method}
\alias{connectivity,numeric-method}
\alias{consensus}
\title{Clustering Connectivity and Consensus Matrices}
\usage{
connectivity(object, ...)

\S4method{connectivity}{ANY}(object, ...)

\S4method{connectivity}{factor}(object, ...)

\S4method{connectivity}{numeric}(object, ...)

consensus(object, ...)
}
\arguments{
\item{object}{an object with a suitable \code{\link{predict}} method.}

\item{...}{extra arguments to allow extension.
They are passed to \code{\link{predict}}, except for the \code{vector} and
\code{factor} methods.}
}
\value{
a square matrix of dimension the number of samples in the model, full
of 0s or 1s.
}
\description{
\code{connectivity} is an S4 generic that computes the connectivity matrix
based on the clustering of samples obtained from a model's \code{\link{predict}}
method.

The consensus matrix has been proposed by Brunet et al. (2004) to help
visualising and measuring the stability of the clusters obtained by
NMF approaches.
For objects of class \code{NMF} (e.g. results of a single NMF run, or NMF
models), the consensus matrix reduces to the connectivity matrix.
}
\details{
The connectivity matrix of a given partition of a set of samples (e.g. given
as  a cluster membership index) is the matrix \eqn{C} containing only 0 or 1
entries such that:
\deqn{C_{ij} = \left\{\begin{array}{l}
1\mbox{ if sample }i\mbox{ belongs to the same cluster as sample }j\\
0\mbox{ otherwise}
\end{array}\right..}{
C_{ij} = 1 if sample i belongs to the same cluster as sample j, 0 otherwise}
}
\examples{

# clustering of random data
h <- hclust(dist(rmatrix(10,20)))
connectivity(stats::cutree(h, 2))

connectivity(gl(2, 4))

}
\references{
Brunet J, Tamayo P, Golub TR, Mesirov JP (2004). “Metagenes and molecular pattern discovery using matrix factorization.” _Proceedings of
the National Academy of Sciences of the United States of America_, *101*(12), 4164-9. ISSN 0027-8424, doi: 10.1073/pnas.0308531101 (URL:
https://doi.org/10.1073/pnas.0308531101).
}
\seealso{
\code{\link{predict}}
}
\section{Methods (by generic)}{
\bold{connectivity:}

\itemize{
\item \code{connectivity(object = NMF)}: Computes the connectivity matrix for an NMF model, for which cluster
membership is given by the most contributing basis component in each sample.
See \code{\link{predict,NMF-method}}.

\item \code{connectivity(object = ANY)}: Default method which computes the connectivity matrix
using the result of \code{predict(x, ...)} as cluster membership index.

\item \code{connectivity(object = factor)}: Computes the connectivity matrix using \code{x} as cluster membership index.

\item \code{connectivity(object = numeric)}: Equivalent to \code{connectivity(as.factor(x))}.
}


\bold{consensus:}

\itemize{
\item \code{consensus(object = NMFfitX)}: Pure virtual method defined to ensure \code{consensus} is defined for sub-classes of \code{NMFfitX}.
It throws an error if called.

\item \code{consensus(object = NMF)}: This method is provided for completeness and is identical to
\code{\link{connectivity}}, and returns the connectivity matrix,
which, in the case of a single NMF model, is also the consensus matrix.

\item \code{consensus(object = NMFfit)}: Shorcut for \code{consensus(fit(object), ...)}

\item \code{consensus(object = NMFfitX1)}: Returns the consensus matrix computed while performing all NMF runs,
amongst which \code{object} was selected as the best fit.

The result is the matrix stored in slot \sQuote{consensus}.
This method returns \code{NULL} if the consensus matrix is empty.

\item \code{consensus(object = NMFfitXn)}: Computes the consensus matrix of the set of fits stored in \code{object}, as
the mean connectivity matrix across runs.

This method returns \code{NULL} on an empty object.
The result is a matrix with several attributes attached, that are used by
plotting functions such as \code{\link{consensusmap}} to annotate the plots.
}

}

