% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{consensus,NMF-method}
\alias{consensus,NMF-method}
\title{This method is provided for completeness and is identical to 
\code{\link{connectivity}}, and returns the connectivity matrix, 
which, in the case of a single NMF model, is also the consensus matrix.}
\usage{
\S4method{consensus}{NMF}(object, ...)
}
\description{
This method is provided for completeness and is identical to 
\code{\link{connectivity}}, and returns the connectivity matrix, 
which, in the case of a single NMF model, is also the consensus matrix.
}
