% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{consensus,NMFfitX-method}
\alias{consensus,NMFfitX-method}
\title{Pure virtual method defined to ensure \code{consensus} is defined for sub-classes of \code{NMFfitX}.
It throws an error if called.}
\usage{
\S4method{consensus}{NMFfitX}(object, ...)
}
\description{
Pure virtual method defined to ensure \code{consensus} is defined for sub-classes of \code{NMFfitX}.
It throws an error if called.
}
