% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{consensus,NMFfitXn-method}
\alias{consensus,NMFfitXn-method}
\alias{plot.NMF.consensus}
\title{Computes the consensus matrix of the set of fits stored in \code{object}, as
the mean connectivity matrix across runs.}
\usage{
\S4method{consensus}{NMFfitXn}(object, ..., no.attrib = FALSE)
}
\description{
This method returns \code{NULL} on an empty object.
The result is a matrix with several attributes attached, that are used by 
plotting functions such as \code{\link{consensusmap}} to annotate the plots.
}
