% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{consensushc,matrix-method}
\alias{consensushc,matrix-method}
\title{Workhorse method for matrices.}
\usage{
\S4method{consensushc}{matrix}(object, method = "average",
  dendrogram = TRUE)
}
\arguments{
\item{method}{linkage method passed to \code{\link{hclust}}.}

\item{dendrogram}{a logical that specifies if the result of the hierarchical 
clustering (en \code{hclust} object) should be converted into a dendrogram. 
Default value is \code{TRUE}.}
}
\description{
Workhorse method for matrices.
}
