% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{consensushc,NMF-method}
\alias{consensushc,NMF-method}
\title{Compute the hierarchical clustering on the connectivity matrix of \code{object}.}
\usage{
\S4method{consensushc}{NMF}(object, ...)
}
\description{
Compute the hierarchical clustering on the connectivity matrix of \code{object}.
}
