% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{consensushc,NMFfitX-method}
\alias{consensushc,NMFfitX-method}
\title{Compute the hierarchical clustering on the consensus matrix of \code{object}, 
or on the connectivity matrix of the best fit in \code{object}.}
\usage{
\S4method{consensushc}{NMFfitX}(object, what = c("consensus", "fit"), ...)
}
\arguments{
\item{what}{character string that indicates which matrix to use in the 
computation.}
}
\description{
Compute the hierarchical clustering on the consensus matrix of \code{object}, 
or on the connectivity matrix of the best fit in \code{object}.
}
