% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{consensusmap,matrix-method}
\alias{consensusmap,matrix-method}
\title{Main method that redefines default values for arguments of \code{\link{aheatmap}}.}
\usage{
\S4method{consensusmap}{matrix}(object, color = "-RdYlBu",
  distfun = function(x) as.dist(1 - x), hclustfun = "average",
  Rowv = TRUE, Colv = "Rowv", main = if (is.null(nr) || nr > 1)
  "Consensus matrix" else "Connectiviy matrix", info = FALSE, ...)
}
\description{
Main method that redefines default values for arguments of \code{\link{aheatmap}}.
}
