% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{consensusmap,NMF-method}
\alias{consensusmap,NMF-method}
\title{Plots a heatmap of the connectivity matrix of an NMF model.}
\usage{
\S4method{consensusmap}{NMF}(object, ...)
}
\description{
Plots a heatmap of the connectivity matrix of an NMF model.
}
