% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{consensusmap,NMFfitX-method}
\alias{consensusmap,NMFfitX-method}
\title{Plots a heatmap of the consensus matrix obtained when fitting an NMF model with multiple runs.}
\usage{
\S4method{consensusmap}{NMFfitX}(object, annRow = NA, annCol = NA,
  tracks = c("basis:", "consensus:", "silhouette:"),
  main = "Consensus matrix", info = FALSE, ...)
}
\description{
Plots a heatmap of the consensus matrix obtained when fitting an NMF model with multiple runs.
}
