% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{cophcor,matrix-method}
\alias{cophcor,matrix-method}
\title{Workhorse method for matrices.}
\usage{
\S4method{cophcor}{matrix}(object, linkage = "average")
}
\arguments{
\item{linkage}{linkage method used in the hierarchical clustering.
It is passed to \code{\link{hclust}}.}
}
\description{
Workhorse method for matrices.
}
