% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{cophcor,NMFfitX-method}
\alias{cophcor,NMFfitX-method}
\title{Computes the cophenetic correlation coefficient on the consensus matrix
of \code{object}.
All arguments in \code{...} are passed to the method \code{cophcor,matrix}.}
\usage{
\S4method{cophcor}{NMFfitX}(object, ...)
}
\description{
Computes the cophenetic correlation coefficient on the consensus matrix
of \code{object}.
All arguments in \code{...} are passed to the method \code{cophcor,matrix}.
}
