% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophCorCoef.R
\name{cophCorCoef}
\alias{cophCorCoef}
\title{Calculate the cophenetic correlation coefficient}
\usage{
cophCorCoef(clusters, verbose = F)
}
\arguments{
\item{clusters}{A matrix were columns represent samples and rows represent ranks AND iterations.
Rownames must be in the form k_{rank number}.rep_{repeat number}}

\item{verbose}{Show progress messages?}
}
\value{
A numeric vector of cophenetic correlation coefficients
}
\description{
The cophenetic measures the stability of the NMF clusters (i.e. signatures).

The procedure is below (based on the Brunet NMF paper: https://www.pnas.org/content/101/12/4164):
(i) Within each rank, and each iteration, assign each sample a signature/cluster number (=the max
contribution signature from the W matrix).
(ii) Convert this vector of cluster numbers into a 0/1 matrix of n_samples x n_samples, with a 1
indicating 2 samples belong to the sample cluster. This is the 'connectivity matrix' (C in the
Brunet paper)
(iii) With e.g. 100 NMF repeats, there are 100 connectivity matrices. Take the mean of the
connectivity matrices. This is the 'consensus matrix' (C_dash in the paper)
(iv) Take 1 - consensus matrix to get the distance matrix, then hierachical cluster on this and
calculate the cophenetic matrix with `stats::cophenetic()`
(v) The cophenetic coefficient is the pearson correlation of the distance matrix with the
cophenetic matrix

The above procedure is repeated for each rank.
}
