% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-terms.R
\docType{methods}
\name{cterms<-,NMFstd-method}
\alias{cterms<-,NMFstd-method}
\title{Default method tries to coerce \code{value} into a \code{data.frame} with 
\code{\link{as.data.frame}}.}
\usage{
\S4method{cterms}{NMFstd}(object) <- value
}
\description{
Default method tries to coerce \code{value} into a \code{data.frame} with 
\code{\link{as.data.frame}}.
}
