% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmap.R
\name{cutdendro}
\alias{cutdendro}
\title{Fade Out the Upper Branches from a Dendrogram}
\usage{
cutdendro(x, n)
}
\arguments{
\item{x}{a dendrogram}

\item{n}{the number of groups}
}
\description{
Fade Out the Upper Branches from a Dendrogram
}
\keyword{internal}
