% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{deviance,NMF-method}
\alias{deviance,NMF-method}
\title{Computes the distance between a matrix and the estimate of an \code{NMF} model.}
\usage{
\S4method{deviance}{NMF}(object, y, method = c("", "KL", "euclidean"), ...)
}
\arguments{
\item{y}{a matrix compatible with the NMF model \code{object}, i.e. \code{y}
must have the same dimension as \code{fitted(object)}.}

\item{method}{a character string or a function with signature
\code{(x="NMF", y="matrix", ...)} that implements a distance measure between
an NMF model \code{x} and a target matrix \code{y}, i.e. an objective function
to use to compute the deviance.
In \code{deviance}, it is passed to \code{nmfDistance} to get the function
that effectively computes the deviance.}

\item{...}{extra parameters passed to the objective function.}
}
\description{
Computes the distance between a matrix and the estimate of an \code{NMF} model.
}
\seealso{
Other stats: \code{\link{deviance}},
  \code{\link{residuals}}
}
