% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{deviance,NMFfit-method}
\alias{deviance,NMFfit-method}
\title{Returns the deviance of a fitted NMF model.}
\usage{
\S4method{deviance}{NMFfit}(object, y, method, ...)
}
\description{
This method returns the final residual value if the target matrix \code{y} is
not supplied, or the approximation error between the fitted NMF model stored 
in \code{object} and \code{y}.
In this case, the computation is performed using the objective function 
\code{method} if not missing, or the objective of the algorithm that 
fitted the model (stored in slot \code{'distance'}).
}
\details{
If not computed by the NMF algorithm itself, the value is automatically
computed at the end of the fitting process by the function \code{\link{nmf}}, 
using the objective function associated with the NMF algorithm, so that it 
should always be available.
}
