% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{deviance,NMFfitX-method}
\alias{deviance,NMFfitX-method}
\title{Returns the deviance achieved by the best fit object, i.e. the lowest 
deviance achieved across all NMF runs.}
\usage{
\S4method{deviance}{NMFfitX}(object, ...)
}
\description{
Returns the deviance achieved by the best fit object, i.e. the lowest 
deviance achieved across all NMF runs.
}
