% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{methods}
\name{deviance,NMFStrategy-method}
\alias{deviance,NMFStrategy-method}
\title{Computes the value of the objective function between the estimate \code{x}
and the target \code{y}.}
\usage{
\S4method{deviance}{NMFStrategy}(object, x, y, ...)
}
\arguments{
\item{x}{an NMF model that estimates \code{y}.}
}
\description{
Computes the value of the objective function between the estimate \code{x}
and the target \code{y}.
}
