% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\name{deviance}
\alias{deviance}
\alias{nmfDistance}
\title{Distances and Objective Functions}
\usage{
deviance(object, ...)

nmfDistance(method = c("", "KL", "euclidean"))
}
\value{
\code{deviance} returns a nonnegative numerical value

\code{nmfDistance} returns a function with least two arguments:
an NMF model and a matrix.
}
\description{
The NMF package defines methods for the generic \code{deviance} from the package \code{stats},
to compute approximation errors between NMF models and matrices, using a variety of
objective functions.
}
\seealso{
Other stats: 
\code{\link{NMF-class}},
\code{\link{residuals}()}
}
\concept{stats}
\section{Methods (by generic)}{
\itemize{
\item \code{deviance(object = NMF)}: Computes the distance between a matrix and the estimate of an \code{NMF} model.

\item \code{deviance(object = NMFfit)}: Returns the deviance of a fitted NMF model.

This method returns the final residual value if the target matrix \code{y} is
not supplied, or the approximation error between the fitted NMF model stored
in \code{object} and \code{y}.
In this case, the computation is performed using the objective function
\code{method} if not missing, or the objective of the algorithm that
fitted the model (stored in slot \code{'distance'}).

If not computed by the NMF algorithm itself, the value is automatically
computed at the end of the fitting process by the function \code{\link{nmf}},
using the objective function associated with the NMF algorithm, so that it
should always be available.

\item \code{deviance(object = NMFfitX)}: Returns the deviance achieved by the best fit object, i.e. the lowest
deviance achieved across all NMF runs.

\item \code{deviance(object = NMFStrategy)}: Computes the value of the objective function between the estimate \code{x}
and the target \code{y}, using the objective function associated with the given
\code{NMFStrategy} object.
}}

