% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{dispersion,matrix-method}
\alias{dispersion,matrix-method}
\title{Workhorse method that computes the dispersion on a given matrix.}
\usage{
\S4method{dispersion}{matrix}(object, ...)
}
\description{
Workhorse method that computes the dispersion on a given matrix.
}
