% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{dispersion,NMFfitX-method}
\alias{dispersion,NMFfitX-method}
\title{Computes the dispersion on the consensus matrix obtained from multiple NMF
runs.}
\usage{
\S4method{dispersion}{NMFfitX}(object, ...)
}
\description{
Computes the dispersion on the consensus matrix obtained from multiple NMF
runs.
}
