% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\name{dispersion}
\alias{dispersion}
\alias{dispersion,matrix-method}
\title{Dispersion of a Matrix}
\usage{
dispersion(object, ...)

\S4method{dispersion}{matrix}(object, ...)
}
\arguments{
\item{object}{an object from which the dispersion is computed}

\item{...}{extra arguments to allow extension}
}
\description{
Computes the dispersion coefficient of a -- consensus -- matrix
\code{object}, generally obtained from multiple NMF runs.
}
\details{
The dispersion coefficient is based on the consensus matrix (i.e. the
average of connectivity matrices) and was proposed by Kim and Park (2007) to
measure the reproducibility of the clusters obtained from NMF.

It is defined as:
\deqn{\rho = \sum_{i,j=1}^n 4 (C_{ij} - \frac{1}{2})^2 , }
where \eqn{n} is the total number of samples.

By construction, \eqn{0 \leq \rho \leq 1} and \eqn{\rho = 1} only for a perfect
consensus matrix, where all entries 0 or 1.
A perfect consensus matrix is obtained only when all the connectivity matrices
are the same, meaning that the algorithm gave the same clusters at each run.
See Kim and Park (2007).
}
\references{
Kim H, Park H (2007). “Sparse non-negative matrix factorizations via alternating non-negativity-constrained least squares for microarray data analysis.” _Bioinformatics (Oxford, England)_, *23*(12), 1495-502.
ISSN 1460-2059, doi: 10.1093/bioinformatics/btm134 (URL: https://doi.org/10.1093/bioinformatics/btm134).
}
\section{Methods (by generic)}{
\itemize{
\item \code{dispersion(object = NMFfitX)}: Computes the \link{dispersion} on the consensus matrix obtained from multiple NMF
runs.

\item \code{dispersion(object = matrix)}: Workhorse method that computes the dispersion on a given matrix.
}}

