\docType{methods}
\name{.basis,NMFarray-method}
\alias{.basis,NMFarray-method}
\title{Get the basis matrix in NMF array models}
\usage{
  \S4method{.basis}{NMFarray}(object, slice = NULL)
}
\arguments{
  \item{object}{an object from which to extract the factor
  matrices, typically an object of class
  \code{\linkS4class{NMF}}.}
}
\description{
  This function returns slot \code{W} of \code{object}.
}
\keyword{methods}

