\docType{methods}
\name{.coef,NMFarray-method}
\alias{.coef,NMFarray-method}
\title{Get the mixture coefficient array in standard NMF array models}
\usage{
  \S4method{.coef}{NMFarray}(object, slice = NULL)
}
\arguments{
  \item{object}{an object from which to extract the factor
  matrices, typically an object of class
  \code{\linkS4class{NMF}}.}
}
\description{
  This function returns slot \code{H} of \code{object}.
}
\keyword{methods}

