% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{entropy,ANY,ANY-method}
\alias{entropy,ANY,ANY-method}
\title{Default method that should work for results of clustering algorithms, that have a
suitable \code{predict} method that returns the cluster membership vector:
the purity is computed between \code{x} and \code{predict{y}}}
\usage{
\S4method{entropy}{ANY,ANY}(x, y, ...)
}
\description{
Default method that should work for results of clustering algorithms, that have a
suitable \code{predict} method that returns the cluster membership vector:
the purity is computed between \code{x} and \code{predict{y}}
}
