% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{entropy,NMFfitXn,ANY-method}
\alias{entropy,NMFfitXn,ANY-method}
\title{Computes the best or mean entropy across all NMF fits stored in \code{x}.}
\usage{
\S4method{entropy}{NMFfitXn,ANY}(x, y, method = "best", ...)
}
\description{
Computes the best or mean entropy across all NMF fits stored in \code{x}.
}
