% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{entropy,table,missing-method}
\alias{entropy,table,missing-method}
\title{Computes the purity directly from the contingency table \code{x}.}
\usage{
\S4method{entropy}{table,missing}(x, y, ...)
}
\description{
This is the workhorse method that is eventually called by all other methods.
}
