% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{evar,ANY-method}
\alias{evar,ANY-method}
\title{Default method for \code{evar}.}
\usage{
\S4method{evar}{ANY}(object, target, ...)
}
\description{
It requires a suitable \code{rss} method to be defined
for \code{object}, as it internally calls \code{rss(object, target, ...)}.
}
