\name{exploreNmf}
\alias{exploreNmf}
\title{Function to extract information on the metagenes obtained after running \code{\link{nmf}}}
\description{It identifies the main features that contributef to defining each metagene, 
as well as identifying to which metagene the samples belong to, based on the frequency they were assigned to it. This function is mainly called by the function \code{\link{runNmf}}}
\usage{
exploreNmf(original.data, nmf.result, prefix, postfix)
}
\arguments{
  \item{original.data}{Matrix with the original data submited to the \code{\link{nmf}}.}
  \item{nmf.result}{Object returned by \code{\link{nmf}} function, run with a single rank and a single algorithm.}
  \item{prefix}{A string used to name the generated plots and the folder to save them.}
  \item{postfix}{A string that informs the range of ranks tested, to be included as part of the files names.}
}
\value{
It generates files to be saved in the folder named after the \code{prefix}.
\item{*.rankSum.tsv}{Tab-separated table with the \code{summary} information of \code{nmf.result}.}
\item{*.V.HAT.tsv}{Tab-separated table with the \code{fitted} results of \code{nmf.result}.}
\item{*.W.tsv}{Tab-separated table with the \code{basis} results for \code{nmf.result}.}
\item{*.H.tsv}{Tab-separated table with the \code{coef} results for \code{nmf.result}.}
\item{*.metagene.tsv}{Tab-separated table informing to which metagene a row or a column of the original data matrix was assigned based on the \code{\link{nmf}} analysis. It is generated by the \code{exportAnalysis} function.}
\item{*.exploreNMF.rda}{File with R objects that contain relevant information regarding the metagenes.
They are exported to enable the user to review the information in the future.}
}
\author{
Cintia C Palu
}