\name{exportAnalysis}
\alias{exportAnalysis}
\title{Function to find to which metagene a row or column of a matrix were more frequently assigned according to the \code{\link{nmf}} results.}
\description{
It searches for the higher value in each column of a matrix, in the context of \code{\link{nmf}} the higher value indicates the metagene more frequently assigned to the varaible represented by that column during the multiple \code{\link{nmf}} runs. The metagene information is added to the orginald ata matrix and saved as a .tsv file. It is designed for users interested in use \code{\link{nmf}} to find clusters on their data. \code{exportAnalysis} is autmatically called by \code{\link{exploreNmf}} function}
\usage{
exportAnalysis(meta, fileID, data)
}
\arguments{
  \item{meta}{Data matrix generated by the  \code{coef} or \code{basis} functions.}
  \item{fileID}{A string providing the file name prefix to be used to save the reuslts.}
  \item{data}{Original matrix submited to \code{\link{nmf}} analysis.}
}
\value{
It generates files to be saved in the folder named after the \code{prefix}.
\item{*.metagene.tsv}{Tab-separated table informing to which metagene a row or a column of the original data matrix was assigned based on the \code{\link{nmf}} analysis.}
}
\author{
Cintia C Palu
}