% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\docType{methods}
\name{extractFeatures,matrix-method}
\alias{extractFeatures,matrix-method}
\title{Select features on a given matrix, that contains the basis component in columns.}
\usage{
\S4method{extractFeatures}{matrix}(object, method = c("kim", "max"),
  format = c("list", "combine", "subset"), nodups = TRUE)
}
\arguments{
\item{method}{scoring or selection method.
It specifies the name of one of the method described in sections \emph{Feature scores} 
and \emph{Feature selection}. 

Additionally for \code{extractFeatures}, it may be an integer vector that 
indicates the number of top most contributing features to 
extract from each column of \code{object}, when ordered in decreasing order, 
or a numeric value between 0 and 1 that indicates the minimum relative basis 
contribution above which a feature is selected (i.e. basis contribution threshold).
In the case of a single numeric value (integer or percentage), it is used for all columns.

Note that \code{extractFeatures(x, 1)} means relative contribution threshold of
100\%, to select the top contributing features one must explicitly specify 
an integer value as in \code{extractFeatures(x, 1L)}.
However, if all elements in methods are > 1, they are automatically treated as 
if they were integers: \code{extractFeatures(x, 2)} means the top-2 most 
contributing features in each component.}

\item{format}{output format. 
The following values are accepted:
\describe{
\item{\sQuote{list}}{(default) returns a list with one element per column in 
\code{object}, each containing the indexes of the selected features, as an 
integer vector.
If \code{object} has row names, these are used to name each index vector. 
Components for which no feature were selected are assigned a \code{NA} value.}

\item{\sQuote{combine}}{ returns all indexes in a single vector.
Duplicated indexes are made unique if \code{nodups=TRUE} (default).}

\item{\sQuote{subset}}{ returns an object of the same class as \code{object}, 
but subset with the selected indexes, so that it contains data only from 
basis-specific features.}
}}

\item{nodups}{logical that indicates if duplicated indexes, 
i.e. features selected on multiple basis components (which should in 
theory not happen), should be only appear once in the result.
Only used when \code{format='combine'}.}
}
\description{
Select features on a given matrix, that contains the basis component in columns.
}
\examples{

# random NMF model
x <- rnmf(3, 50,20)

# probably no feature is selected
extractFeatures(x)
# extract top 5 for each basis
extractFeatures(x, 5L)
# extract features that have a relative basis contribution above a threshold
extractFeatures(x, 0.5)
# ambiguity?
extractFeatures(x, 1) # means relative contribution above 100\%
extractFeatures(x, 1L) # means top contributing feature in each component

}
