% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\docType{methods}
\name{extractFeatures,NMF-method}
\alias{extractFeatures,NMF-method}
\title{Select basis-specific features from an NMF model, by applying the method 
\code{extractFeatures,matrix} to its basis matrix.}
\usage{
\S4method{extractFeatures}{NMF}(object, ...)
}
\description{
Select basis-specific features from an NMF model, by applying the method 
\code{extractFeatures,matrix} to its basis matrix.
}
