% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms-snmf.R
\docType{methods}
\name{fcnnls,ANY,numeric-method}
\alias{fcnnls,ANY,numeric-method}
\title{Shortcut for \code{fcnnls(x, as.matrix(y), ...)}.}
\usage{
\S4method{fcnnls}{ANY,numeric}(x, y, ...)
}
\description{
Shortcut for \code{fcnnls(x, as.matrix(y), ...)}.
}
