% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms-snmf.R
\docType{methods}
\name{fcnnls,matrix,matrix-method}
\alias{fcnnls,matrix,matrix-method}
\title{This method wraps a call to the internal function \code{.fcnnls}, and
formats the results in a similar way as other lest-squares methods such 
as \code{\link{lm}}.}
\usage{
\S4method{fcnnls}{matrix,matrix}(x, y, verbose = FALSE, pseudo = TRUE, ...)
}
\arguments{
\item{verbose}{toggle verbosity (default is \code{FALSE}).}
}
\description{
This method wraps a call to the internal function \code{.fcnnls}, and
formats the results in a similar way as other lest-squares methods such 
as \code{\link{lm}}.
}
