% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms-snmf.R
\docType{methods}
\name{fcnnls,numeric,matrix-method}
\alias{fcnnls,numeric,matrix-method}
\title{Shortcut for \code{fcnnls(as.matrix(x), y, ...)}.}
\usage{
\S4method{fcnnls}{numeric,matrix}(x, y, ...)
}
\description{
Shortcut for \code{fcnnls(as.matrix(x), y, ...)}.
}
