% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\docType{methods}
\name{featureScore,matrix-method}
\alias{featureScore,matrix-method}
\title{Computes feature scores on a given matrix, that contains the basis component in columns.}
\usage{
\S4method{featureScore}{matrix}(object, method = c("kim", "max"))
}
\description{
Computes feature scores on a given matrix, that contains the basis component in columns.
}
