% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\docType{methods}
\name{featureScore,NMF-method}
\alias{featureScore,NMF-method}
\title{Computes feature scores on the basis matrix of an NMF model.}
\usage{
\S4method{featureScore}{NMF}(object, ...)
}
\description{
Computes feature scores on the basis matrix of an NMF model.
}
