% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{fit,NMFfit-method}
\alias{fit,NMFfit-method}
\title{Returns the NMF model object stored in slot \code{'fit'}.}
\usage{
\S4method{fit}{NMFfit}(object)
}
\description{
Returns the NMF model object stored in slot \code{'fit'}.
}
