% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{fit,NMFfitX-method}
\alias{fit,NMFfitX-method}
\title{Returns the model object that achieves the lowest residual approximation 
error across all the runs.}
\usage{
\S4method{fit}{NMFfitX}(object)
}
\description{
It is a pure virtual method defined to ensure \code{fit} is defined 
for sub-classes of \code{NMFfitX}, which throws an error if called.
}
