% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{fit,NMFfitXn-method}
\alias{fit,NMFfitXn-method}
\title{Returns the best NMF fit object amongst all the fits stored in \code{object},
i.e. the fit that achieves the lowest estimation residuals.}
\usage{
\S4method{fit}{NMFfitXn}(object)
}
\description{
Returns the best NMF fit object amongst all the fits stored in \code{object},
i.e. the fit that achieves the lowest estimation residuals.
}
