% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{fit<-,NMFfit,NMF-method}
\alias{fit<-,NMFfit,NMF-method}
\title{Updates the NMF model object stored in slot \code{'fit'} with a new value.}
\usage{
\S4method{fit}{NMFfit,NMF}(object) <- value
}
\description{
Updates the NMF model object stored in slot \code{'fit'} with a new value.
}
