% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\name{fit}
\alias{fit}
\alias{fit<-}
\alias{minfit}
\title{Extracting Fitted Models}
\usage{
fit(object, ...)

fit(object) <- value

minfit(object, ...)
}
\arguments{
\item{object}{an object fitted by some algorithm, e.g. as returned by the
function \code{\link{nmf}}.}

\item{...}{extra arguments to allow extension}

\item{value}{replacement value}
}
\description{
The functions \code{fit} and \code{minfit} are S4 genetics that extract
the best model object and the best fit object respectively, from a collection
of models or from a wrapper object.
}
\details{
A fit object differs from a model object in that it contains data about the
fit, such as the initial RNG settings, the CPU time used, etc\ldots, while
a model object only contains the actual modelling data such as regression
coefficients, loadings, etc\ldots

That best model is generally defined as the one that achieves the
maximum/minimum some quantitative measure, amongst all models in a collection.

In the case of NMF models, the best model is the one that achieves the best
approximation error, according to the objective function associated with the
algorithm that performed the fit(s).
}
\section{Methods (by generic)}{
\bold{fit:}

\itemize{
\item \code{fit(object = NMFfit)}: Returns the NMF model object stored in slot \code{'fit'}.

\item \code{fit(object = NMFfitX)}: Returns the model object that achieves the lowest residual approximation
error across all the runs.

It is a pure virtual method defined to ensure \code{fit} is defined
for sub-classes of \code{NMFfitX}, which throws an error if called.

\item \code{fit(object = NMFfitX1)}: Returns the model object associated with the best fit, amongst all the
runs performed when fitting \code{object}.

Since \code{NMFfitX1} objects only hold the best fit, this method simply
returns the NMF model fitted by \code{object} -- that is stored in slot
\sQuote{fit}.

\item \code{fit(object = NMFfitXn)}: Returns the best NMF fit object amongst all the fits stored in \code{object},
i.e. the fit that achieves the lowest estimation residuals.
}


\bold{fit<-:}

\itemize{
\item \code{fit<-(object = NMFfit,value = NMF)}: Updates the NMF model object stored in slot \code{'fit'} with a new value.
}


\bold{minfit:}

\itemize{
\item \code{minfit(object = NMFfit)}: Returns the object its self, since there it is the result of a single NMF run.

\item \code{minfit(object = NMFfitX)}: Returns the fit object that achieves the lowest residual approximation
error across all the runs.

It is a pure virtual method defined to ensure \code{minfit} is defined
for sub-classes of \code{NMFfitX}, which throws an error if called.

\item \code{minfit(object = NMFfitX1)}: Returns the fit object associated with the best fit, amongst all the
runs performed when fitting \code{object}.

Since \code{NMFfitX1} objects only hold the best fit, this method simply
returns \code{object} coerced into an \code{NMFfit} object.

\item \code{minfit(object = NMFfitXn)}: Returns the best NMF model in the list, i.e. the run that achieved the lower
estimation residuals.

The model is selected based on its \code{deviance} value.
}

}

