% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{fitted,NMF-method}
\alias{fitted,NMF-method}
\title{Pure virtual method for objects of class \code{\linkS4class{NMF}}, that 
should be overloaded by sub-classes, and throws an error if called.}
\usage{
\S4method{fitted}{NMF}(object, ...)
}
\description{
Pure virtual method for objects of class \code{\linkS4class{NMF}}, that 
should be overloaded by sub-classes, and throws an error if called.
}
