% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFns-class.R
\docType{methods}
\name{fitted,NMFns-method}
\alias{fitted,NMFns-method}
\title{Compute estimate for an NMFns object, according to the Nonsmooth NMF model 
(cf. \code{\link{NMFns-class}}).}
\usage{
\S4method{fitted}{NMFns}(object, W, H, S, ...)
}
\arguments{
\item{S}{smoothing matrix to use instead of \code{smoothing(object)}
It must be a square matrix compatible with the basis and coefficient matrices 
used in the computation.}
}
\description{
Extra arguments in \code{...} are passed to method \code{smoothing}, and are 
typically used to pass a value for \code{theta}, which is used to compute 
the smoothing matrix instead of the one stored in \code{object}.
}
