% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFOffset-class.R
\docType{methods}
\name{fitted,NMFOffset-method}
\alias{fitted,NMFOffset-method}
\title{Computes the target matrix estimate for an NMFOffset object.}
\usage{
\S4method{fitted}{NMFOffset}(object, W, H, offset = object@offset)
}
\arguments{
\item{offset}{offset vector}
}
\description{
The estimate is computed as:
\deqn{ W H + offset }
}
