% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFstd-class.R
\docType{methods}
\name{fitted,NMFstd-method}
\alias{fitted,NMFstd-method}
\title{Compute the target matrix estimate in \emph{standard NMF models}.}
\usage{
\S4method{fitted}{NMFstd}(object, W, H)
}
\arguments{
\item{W}{a matrix to use in the computation as the basis matrix in place of 
\code{basis(object)}. 
It must be compatible with the coefficient matrix used 
in the computation (i.e. number of columns in \code{W} = number of rows in \code{H}).}

\item{H}{a matrix to use in the computation as the coefficient matrix in place of 
\code{coef(object)}. 
It must be compatible with the basis matrix used 
in the computation (i.e. number of rows in \code{H} = number of columns in \code{W}).}
}
\description{
The estimate matrix is computed as the product of the two matrix slots 
\code{W} and \code{H}:
\deqn{\hat{V} = W H}{V ~ W H}
}
\examples{
# random standard NMF model
x <- rnmf(3, 10, 5)
all.equal(fitted(x), basis(x) \%*\% coef(x))


}
