% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\name{fitted}
\alias{fitted}
\title{Fitted Matrix in NMF Models}
\usage{
fitted(object, ...)
}
\arguments{
\item{object}{an object that inherit from class \linkS4class{NMF}}

\item{...}{extra arguments to allow extension}
}
\value{
the target matrix estimate as fitted by the model \code{object}
}
\description{
Computes the estimated target matrix based on a given \emph{NMF} model.
The estimation depends on the underlying NMF model.
For example in the standard model \eqn{V \equiv W H}{V ~ W H}, the target matrix is
estimated by the matrix product \eqn{W H}.
In other models, the estimate may depend on extra parameters/matrix
(cf. Non-smooth NMF in \code{\link{NMFns-class}}).
}
\details{
This function is a S4 generic function imported from \link[stats:fitted]{stats::fitted} in
the package \emph{stats}.
It is implemented as a pure virtual method for objects of class
\code{NMF}, meaning that concrete NMF models must provide a
definition for their corresponding class (i.e. sub-classes of
class \code{NMF}).
See \code{\linkS4class{NMF}} for more details.
}
\section{Methods (by generic)}{
\itemize{
\item \code{fitted(object = NMF)}: Pure virtual method for objects of class \code{\linkS4class{NMF}}, that
should be overloaded by sub-classes, and throws an error if called.

\item \code{fitted(object = NMFstd)}: Compute the target matrix estimate in \emph{standard NMF models}.

The estimate matrix is computed as the product of the two matrix slots
\code{W} and \code{H}:
\deqn{\hat{V} = W H}{V ~ W H}

\item \code{fitted(object = NMFOffset)}: Computes the target matrix estimate for an NMFOffset object.

The estimate is computed as:
\deqn{ W H + offset }

\item \code{fitted(object = NMFns)}: Compute estimate for an NMFns object, according to the Nonsmooth NMF model
(cf. \code{\link{NMFns-class}}).

Extra arguments in \code{...} are passed to method \code{smoothing}, and are
typically used to pass a value for \code{theta}, which is used to compute
the smoothing matrix instead of the one stored in \code{object}.

\item \code{fitted(object = NMFfit)}: Computes and return the estimated target matrix from an NMF model fitted with
function \code{\link{nmf}}.

It is a shortcut for \code{fitted(fit(object), ...)}, dispatching the call to
the \code{fitted} method of the actual NMF model.
}}

