% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{registerDoBackend}
\alias{registerDoBackend}
\alias{getDoBackend}
\alias{setDoBackend}
\alias{register}
\alias{register.foreach_backend}
\alias{ForeachBackend}
\alias{ForeachBackend,ANY-method}
\alias{ForeachBackend,character-method}
\alias{ForeachBackend,missing-method}
\alias{ForeachBackend,NULL-method}
\alias{ForeachBackend,cluster-method}
\alias{ForeachBackend,numeric-method}
\alias{ForeachBackend,doParallel_backend-method}
\alias{ForeachBackend,doParallelMC_backend-method}
\alias{ForeachBackend,doParallelSNOW_backend-method}
\alias{ForeachBackend,doPSOCK_backend-method}
\alias{register.doParallel_backend}
\alias{register.doMPI_backend}
\alias{ForeachBackend,mpicluster-method}
\alias{ForeachBackend,doMPI_backend-method}
\alias{getDoParHosts}
\alias{getDoParHosts,ANY-method}
\alias{getDoParNHosts}
\title{Utilities and Extensions for Foreach Loops}
\usage{
registerDoBackend(object, ...)

getDoBackend()

setDoBackend(data, cleanup = FALSE)

register(x, ...)

\method{register}{foreach_backend}(x, ...)

ForeachBackend(object, ...)

\S4method{ForeachBackend}{ANY}(object, ...)

\S4method{ForeachBackend}{character}(object, ...)

\S4method{ForeachBackend}{missing}(object, ...)

\S4method{ForeachBackend}{`NULL`}(object, ...)

\S4method{ForeachBackend}{cluster}(object, ...)

\S4method{ForeachBackend}{numeric}(object, ...)

\S4method{ForeachBackend}{doParallel_backend}(object, cl, type = NULL)

\S4method{ForeachBackend}{doParallelMC_backend}(object, ...)

\S4method{ForeachBackend}{doParallelSNOW_backend}(object, ...)

\S4method{ForeachBackend}{doPSOCK_backend}(object, cl)

\method{register}{doParallel_backend}(x, ...)

\method{register}{doMPI_backend}(x, ...)

\S4method{ForeachBackend}{mpicluster}(object, ...)

\S4method{ForeachBackend}{doMPI_backend}(object, cl)

getDoParHosts(object, ...)

\S4method{getDoParHosts}{ANY}(object, ...)

getDoParNHosts(object)
}
\arguments{
\item{object}{specification of a foreach backend, e.g. \sQuote{SEQ},
\sQuote{PAR} (for doParallel), \sQuote{MPI}, etc\ldots}

\item{...}{extra arguments passed to the backend own registration function.}

\item{data}{internal data of a foreach \%dopar\% backend.}

\item{cleanup}{logical that indicates if the previous
backend's cleanup procedure should be run, \strong{before}
setting the new backend.}

\item{x}{specification of a foreach backend}

\item{cl}{cluster specification: a cluster object or a numeric that indicates the
number of nodes to use.}

\item{type}{type of cluster, See \code{\link[parallel]{makeCluster}}.}
}
\description{
\code{registerDoBackend} is a unified register function for foreach backends.
}
\keyword{internal}
\section{Methods (by generic)}{
\bold{ForeachBackend:}

\itemize{
\item \code{ForeachBackend(object = ANY)}: Default method defined to throw an informative error message, when no other
method was found.

\item \code{ForeachBackend(object = character)}: Creates a foreach backend object based on its name.

\item \code{ForeachBackend(object = missing)}: Creates a foreach backend object for the currently registered backend.

\item \code{ForeachBackend(object = NULL)}: Dummy method that returns \code{NULL}, defined for correct dispatch.

\item \code{ForeachBackend(object = cluster)}: Creates a doParallel foreach backend that uses the cluster described in
\code{object}.

\item \code{ForeachBackend(object = numeric)}: Creates a doParallel foreach backend with \code{object} processes.

\item \code{ForeachBackend(object = doParallel_backend)}: doParallel-specific backend factory

\item \code{ForeachBackend(object = doParallelMC_backend)}: doParallel-specific backend factory for multicore (fork) clusters

This method is needed since version 1.0.7 of \pkg{doParallel}, which removed
internal function \code{info} and defined separate backend names for mc and snow clusters.

\item \code{ForeachBackend(object = doParallelSNOW_backend)}: doParallel-specific backend factory for SNOW clusters.

This method is needed since version 1.0.7 of \pkg{doParallel}, which removed
internal function \code{info} and defined separate backend names for mc and snow clusters.

\item \code{ForeachBackend(object = doPSOCK_backend)}: doSNOW-specific backend factory

\item \code{ForeachBackend(object = mpicluster)}: Creates a doMPI foreach backend that uses the MPI cluster described in
\code{object}.

\item \code{ForeachBackend(object = doMPI_backend)}: doMPI-specific backend factory
}


\bold{getDoParHosts:}

\itemize{
\item \code{getDoParHosts(object = ANY)}: Default method that tries to heuristaically infer the number of hosts and in last
resort temporarly register the backend and performs a foreach loop, to retrieve the
nodename from each worker.
}

}

