% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\docType{methods}
\name{ForeachBackend,ANY-method}
\alias{ForeachBackend,ANY-method}
\title{Default method defined to throw an informative error message, when no other
method was found.}
\usage{
\S4method{ForeachBackend}{ANY}(object, ...)
}
\description{
Default method defined to throw an informative error message, when no other
method was found.
}
