% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\docType{methods}
\name{ForeachBackend,character-method}
\alias{ForeachBackend,character-method}
\title{Creates a foreach backend object based on its name.}
\usage{
\S4method{ForeachBackend}{character}(object, ...)
}
\description{
Creates a foreach backend object based on its name.
}
