% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\docType{methods}
\name{ForeachBackend,cluster-method}
\alias{ForeachBackend,cluster-method}
\title{Creates a doParallel foreach backend that uses the cluster described in 
\code{object}.}
\usage{
\S4method{ForeachBackend}{cluster}(object, ...)
}
\description{
Creates a doParallel foreach backend that uses the cluster described in 
\code{object}.
}
