% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\docType{methods}
\name{ForeachBackend,doMPI_backend-method}
\alias{ForeachBackend,doMPI_backend-method}
\title{doMPI-specific backend factory}
\usage{
\S4method{ForeachBackend}{doMPI_backend}(object, cl)
}
\description{
doMPI-specific backend factory
}
