% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\docType{methods}
\name{ForeachBackend,doParallelMC_backend-method}
\alias{ForeachBackend,doParallelMC_backend-method}
\title{doParallel-specific backend factory for multicore (fork) clusters}
\usage{
\S4method{ForeachBackend}{doParallelMC_backend}(object, ...)
}
\description{
This method is needed since version 1.0.7 of \pkg{doParallel}, which removed 
internal function \code{info} and defined separate backend names for mc and snow clusters.
}
