% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\docType{methods}
\name{ForeachBackend,doPSOCK_backend-method}
\alias{ForeachBackend,doPSOCK_backend-method}
\title{doSNOW-specific backend factory}
\usage{
\S4method{ForeachBackend}{doPSOCK_backend}(object, cl)
}
\description{
doSNOW-specific backend factory
}
