% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\docType{methods}
\name{ForeachBackend,missing-method}
\alias{ForeachBackend,missing-method}
\title{Creates a foreach backend object for the currently registered backend.}
\usage{
\S4method{ForeachBackend}{missing}(object, ...)
}
\description{
Creates a foreach backend object for the currently registered backend.
}
