% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\docType{methods}
\name{ForeachBackend,mpicluster-method}
\alias{ForeachBackend,mpicluster-method}
\title{Creates a doMPI foreach backend that uses the MPI cluster described in 
\code{object}.}
\usage{
\S4method{ForeachBackend}{mpicluster}(object, ...)
}
\description{
Creates a doMPI foreach backend that uses the MPI cluster described in 
\code{object}.
}
