% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\docType{methods}
\name{ForeachBackend,NULL-method}
\alias{ForeachBackend,NULL-method}
\title{Dummy method that returns \code{NULL}, defined for correct dispatch.}
\usage{
\S4method{ForeachBackend}{`NULL`}(object, ...)
}
\description{
Dummy method that returns \code{NULL}, defined for correct dispatch.
}
