% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\docType{methods}
\name{ForeachBackend,numeric-method}
\alias{ForeachBackend,numeric-method}
\title{Creates a doParallel foreach backend with \code{object} processes.}
\usage{
\S4method{ForeachBackend}{numeric}(object, ...)
}
\description{
Creates a doParallel foreach backend with \code{object} processes.
}
