% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\docType{methods}
\name{getDoParHosts,ANY-method}
\alias{getDoParHosts,ANY-method}
\title{Default method that tries to heuristaically infer the number of hosts and in last 
resort temporarly register the backend and performs a foreach loop, to retrieve the 
nodename from each worker.}
\usage{
\S4method{getDoParHosts}{ANY}(object, ...)
}
\description{
Default method that tries to heuristaically infer the number of hosts and in last 
resort temporarly register the backend and performs a foreach loop, to retrieve the 
nodename from each worker.
}
