% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{getRNG1,NMFfitX-method}
\alias{getRNG1,NMFfitX-method}
\title{Returns the RNG settings used for the first NMF run of multiple NMF runs.}
\usage{
\S4method{getRNG1}{NMFfitX}(object)
}
\description{
Returns the RNG settings used for the first NMF run of multiple NMF runs.
}
\examples{
# For multiple NMF runs, the RNG settings used for the first run is also stored
V <- rmatrix(20,10)
res <- nmf(V, 3, nrun=4)
# RNG used for the best fit
getRNG(res)
# RNG used for the first of all fits
getRNG1(res)
# they may differ if the best fit is not the first one
rng.equal(res, getRNG1(res))

}
