% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{getRNG1,NMFfitX1-method}
\alias{getRNG1,NMFfitX1-method}
\title{Returns the RNG settings used to compute the first of all NMF runs, amongst
which \code{object} was selected as the best fit.}
\usage{
\S4method{getRNG1}{NMFfitX1}(object)
}
\description{
Returns the RNG settings used to compute the first of all NMF runs, amongst
which \code{object} was selected as the best fit.
}
