% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{getRNG1,NMFfitXn-method}
\alias{getRNG1,NMFfitXn-method}
\title{Returns the RNG settings used for the first run.}
\usage{
\S4method{getRNG1}{NMFfitXn}(object)
}
\description{
This method throws an error if the object is empty.
}
