% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{tryViewport}
\alias{tryViewport}
\alias{current.vpPath_patched}
\alias{.use.grid.patch}
\title{Internal Grid Extension}
\usage{
tryViewport(name, verbose = FALSE)

current.vpPath_patched(force.unpatched = FALSE, verbose = FALSE)

.use.grid.patch()
}
\arguments{
\item{name}{viewport name}

\item{verbose}{toggle verbosity}
}
\description{
These functions enable mixing base and grid graphics in \code{\link{aheatmap}},
by avoiding calls to the grid internal function \code{'L_gridDirty'}.
They are not exported (i.e. not tampering core functions) and are only meant for internal
use within the \pkg{NMF} package.
}
\details{
\code{tryViewport} tries to go down to a viewport in the current tree,
given its name.

\code{tryViewport} uses \code{\link[grid]{grid.ls}} and not
\code{\link{seekViewport}} as the latter would reset the graphic device
and break the mix grid/base graphic capability.
}
\keyword{internal}
