% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{ibterms,NMFfit-method}
\alias{ibterms,NMFfit-method}
\title{Method for single NMF fit objects, which returns the indexes of fixed
basis terms from the fitted model.}
\usage{
\S4method{ibterms}{NMFfit}(object)
}
\description{
Method for single NMF fit objects, which returns the indexes of fixed
basis terms from the fitted model.
}
