% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{ibterms,NMFfitX-method}
\alias{ibterms,NMFfitX-method}
\title{Method for multiple NMF fit objects, which returns the indexes of fixed basis 
terms from the best fitted model.}
\usage{
\S4method{ibterms}{NMFfitX}(object)
}
\description{
Method for multiple NMF fit objects, which returns the indexes of fixed basis 
terms from the best fitted model.
}
