% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-terms.R
\docType{methods}
\name{ibterms,NMFstd-method}
\alias{ibterms,NMFstd-method}
\title{Method for standard NMF models, which returns the integer vector that is 
stored in slot \code{ibterms} when a formula-based NMF model is instantiated.}
\usage{
\S4method{ibterms}{NMFstd}(object)
}
\description{
Method for standard NMF models, which returns the integer vector that is 
stored in slot \code{ibterms} when a formula-based NMF model is instantiated.
}
