% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-terms.R
\docType{methods}
\name{icterms,NMF-method}
\alias{icterms,NMF-method}
\title{Default pure virtual method that ensure a method is defined for concrete 
NMF model classes.}
\usage{
\S4method{icterms}{NMF}(object, ...)
}
\description{
Default pure virtual method that ensure a method is defined for concrete 
NMF model classes.
}
