% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R, R/NMFSet-class.R
\docType{methods}
\name{icterms,NMFfit-method}
\alias{icterms,NMFfit-method}
\alias{icterms,NMFfit-method}
\title{Method for single NMF fit objects, which returns the indexes of fixed 
coefficient terms from the fitted model.}
\usage{
\S4method{icterms}{NMFfit}(object)

\S4method{icterms}{NMFfit}(object)
}
\description{
Method for single NMF fit objects, which returns the indexes of fixed 
coefficient terms from the fitted model.

Method for multiple NMF fit objects, which returns the indexes of fixed 
coefficient terms from the best fitted model.
}
