% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-terms.R
\docType{methods}
\name{icterms,NMFstd-method}
\alias{icterms,NMFstd-method}
\title{Method for standard NMF models, which returns the integer vector that is 
stored in slot \code{icterms} when a formula-based NMF model is instantiated.}
\usage{
\S4method{icterms}{NMFstd}(object)
}
\description{
Method for standard NMF models, which returns the integer vector that is 
stored in slot \code{icterms} when a formula-based NMF model is instantiated.
}
