\docType{methods}
\name{initialize,NMFStrategyOctave-method}
\alias{initialize,NMFStrategyOctave-method}
\title{Initialize method for \code{\linkS4class{NMFStrategyOctave}}.}
\usage{
  \S4method{initialize}{NMFStrategyOctave}(.Object, ...,
    mfiles)
}
\arguments{
  \item{mfiles}{.m files specifications.}

  \item{.Object}{ An object: see the Details section.}

  \item{...}{data to include in the new object.  Named
  arguments correspond to slots in the class definition.
  Unnamed arguments must be objects from classes that this
  class extends.}
}
\description{
  Initialize method for
  \code{\linkS4class{NMFStrategyOctave}}.
}
\keyword{methods}

