% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedData-methods.R,
%   R/NMFStrategyIterative-class.R, ./<text>
\name{names,AnnotationData-method}
\alias{names,AnnotationData-method}
\alias{names<-,AnnotationData-method}
\alias{dimnames,AnnotationData-method}
\alias{dimnames<-,AnnotationData,ANY-method}
\alias{[,AnnotationData,ANY,ANY,ANY-method}
\alias{$,AnnotationData-method}
\alias{$<-,AnnotationData-method}
\alias{[[,AnnotationData-method}
\alias{[[<-,AnnotationData-method}
\alias{dimnames,AnnotatedData-method}
\alias{dimnames<-,AnnotatedData,ANY-method}
\alias{dim,AnnotatedData-method}
\alias{[,AnnotatedData,ANY,ANY,ANY-method}
\alias{$,AnnotatedData-method}
\alias{$<-,AnnotatedData-method}
\alias{[[,AnnotatedData-method}
\alias{[[<-,AnnotatedData-method}
\alias{run,NMFStrategyIterativeX,mMatrix,NMFfit-method}
\alias{Internal-objects}
\title{Internal Undocumented Entries
This man page is automatically generated to gather all internal functions, classes and objects that are
not meant to be called outside of the package.}
\usage{
\S4method{names}{AnnotationData}(x)

\S4method{names}{AnnotationData}(x) <- value

\S4method{dimnames}{AnnotationData}(x)

\S4method{dimnames}{AnnotationData,ANY}(x) <- value

\S4method{[}{AnnotationData,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{$}{AnnotationData}(x, name)

\S4method{$}{AnnotationData}(x, name) <- value

\S4method{[[}{AnnotationData}(x, i, j, ...)

\S4method{[[}{AnnotationData}(x, i, j, ...) <- value

\S4method{dimnames}{AnnotatedData}(x)

\S4method{dimnames}{AnnotatedData,ANY}(x) <- value

\S4method{dim}{AnnotatedData}(x)

\S4method{[}{AnnotatedData,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{$}{AnnotatedData}(x, name)

\S4method{$}{AnnotatedData}(x, name) <- value

\S4method{[[}{AnnotatedData}(x, i, j, ...)

\S4method{[[}{AnnotatedData}(x, i, j, ...) <- value

\S4method{run}{NMFStrategyIterativeX,mMatrix,NMFfit}(object, y, x, maxIter, ...)
}
\description{
Internal Undocumented Entries
This man page is automatically generated to gather all internal functions, classes and objects that are
not meant to be called outside of the package.
}
\keyword{internal}
