% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithmic.R
\docType{methods}
\name{logs,ANY-method}
\alias{logs,ANY-method}
\title{Default method that returns the value of attribute/slot \code{'logs'} or, if this latter  
does not exists, the value of element \code{'logs'} if \code{object} is a \code{list}.
It returns \code{NULL} if no logging data was found.}
\usage{
\S4method{logs}{ANY}(object)
}
\description{
Default method that returns the value of attribute/slot \code{'logs'} or, if this latter  
does not exists, the value of element \code{'logs'} if \code{object} is a \code{list}.
It returns \code{NULL} if no logging data was found.
}
