\name{match_atrack}
\alias{match_atrack}
\title{Extending Annotation Vectors}
\usage{
  match_atrack(x, data = NULL)
}
\arguments{
  \item{x}{annotation vector}

  \item{data}{reference data}
}
\value{
  a vector of the same type as \code{x}
}
\description{
  Extends a vector used as an annotation track to match the
  number of rows and the row names of a given data.
}

