% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\docType{methods}
\name{metaHeatmap,matrix-method}
\alias{metaHeatmap,matrix-method}
\title{Defunct method substituted by \code{\link{aheatmap}}.}
\usage{
\S4method{metaHeatmap}{matrix}(object, ...)
}
\description{
Defunct method substituted by \code{\link{aheatmap}}.
}
