% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\docType{methods}
\name{metaHeatmap,NMF-method}
\alias{metaHeatmap,NMF-method}
\title{Deprecated method that is substituted by \code{\link{coefmap}} and \code{\link{basismap}}.}
\usage{
\S4method{metaHeatmap}{NMF}(object, ...)
}
\description{
Deprecated method that is substituted by \code{\link{coefmap}} and \code{\link{basismap}}.
}
