% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{metaHeatmap,NMFfitX-method}
\alias{metaHeatmap,NMFfitX-method}
\title{Deprecated method subsituted by \code{\link{consensusmap}}.}
\usage{
\S4method{metaHeatmap}{NMFfitX}(object, ...)
}
\description{
Deprecated method subsituted by \code{\link{consensusmap}}.
}
