% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{minfit,NMFfit-method}
\alias{minfit,NMFfit-method}
\title{Returns the object its self, since there it is the result of a single NMF run.}
\usage{
\S4method{minfit}{NMFfit}(object)
}
\description{
Returns the object its self, since there it is the result of a single NMF run.
}
