% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{minfit,NMFfitXn-method}
\alias{minfit,NMFfitXn-method}
\title{Returns the best NMF model in the list, i.e. the run that achieved the lower 
estimation residuals.}
\usage{
\S4method{minfit}{NMFfitXn}(object)
}
\description{
The model is selected based on its \code{deviance} value.
}
