% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithmic.R
\docType{methods}
\name{modelname,ANY-method}
\alias{modelname,ANY-method}
\title{Default method which returns the class name(s) of \code{object}.
This should work for objects representing models on their own.}
\usage{
\S4method{modelname}{ANY}(object)
}
\description{
For NMF objects, this is the type of NMF model, that corresponds to the 
name of the S4 sub-class of \code{\linkS4class{NMF}}, inherited by \code{object}.
}
\examples{
# get the type of an NMF model
modelname(nmfModel(3))
modelname(nmfModel(3, model='NMFns'))
modelname(nmfModel(3, model='NMFOffset'))

}
