% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{modelname,NMFfit-method}
\alias{modelname,NMFfit-method}
\title{Returns the type of a fitted NMF model.
It is a shortcut for \code{modelname(fit(object)}.}
\usage{
\S4method{modelname}{NMFfit}(object)
}
\description{
Returns the type of a fitted NMF model.
It is a shortcut for \code{modelname(fit(object)}.
}
