% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{modelname,NMFfitXn-method}
\alias{modelname,NMFfitXn-method}
\title{Returns the common type NMF model of all fits stored in \code{object}}
\usage{
\S4method{modelname}{NMFfitXn}(object)
}
\description{
Since all fits are from the same NMF model, this method returns the 
model type of the first fit.
It returns \code{NULL} if the object is empty.
}
