% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{methods}
\name{modelname,NMFStrategy-method}
\alias{modelname,NMFStrategy-method}
\title{Returns the model(s) that an NMF algorithm can fit.}
\usage{
\S4method{modelname}{NMFStrategy}(object)
}
\description{
Returns the model(s) that an NMF algorithm can fit.
}
\examples{
# get the type of model(s) associated with an NMF algorithm
modelname( nmfAlgorithm('brunet') )
modelname( nmfAlgorithm('nsNMF') )
modelname( nmfAlgorithm('offset') )

}
