% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{methods}
\name{name<-,Strategy,character-method}
\alias{name<-,Strategy,character-method}
\title{Sets the name(s) of an NMF algorithm}
\usage{
\S4method{name}{Strategy,character}(object) <- value
}
\description{
Sets the name(s) of an NMF algorithm
}
