% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{methods}
\name{name,Strategy-method}
\alias{name,Strategy-method}
\title{Returns the name of an algorithm}
\usage{
\S4method{name}{Strategy}(object, all = FALSE)
}
\arguments{
\item{all}{a logical that indicates if all the names associated with a strategy 
should be returned (\code{TRUE}), or only the first (primary) one (\code{FALSE}).}
}
\description{
Returns the name of an algorithm
}
