% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{nbasis,ANY-method}
\alias{nbasis,ANY-method}
\title{Default method which returns the number of columns of the basis matrix extracted
from \code{x} using a suitable method \code{basis}, or, if the latter is \code{NULL},
the value of attributes \code{'nbasis'}.}
\usage{
\S4method{nbasis}{ANY}(x, ...)
}
\description{
For NMF models, this also corresponds to the number of rows in the coefficient
matrix.
}
