% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{nbasis,NMFfitXn-method}
\alias{nbasis,NMFfitXn-method}
\title{Returns the number of basis components common to all fits.}
\usage{
\S4method{nbasis}{NMFfitXn}(x, ...)
}
\description{
Since all fits have been computed using the same rank, it returns the 
factorization rank of the first fit.
This method returns \code{NULL} if the object is empty.
}
