% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{niter,NMFfit-method}
\alias{niter,NMFfit-method}
\title{Returns the number of iteration performed to fit an NMF model, typically 
with function \code{\link{nmf}}.}
\usage{
\S4method{niter}{NMFfit}(object, ...)
}
\description{
Currently this data is stored in slot \code{'extra'}, but this might change 
in the future.
}
