% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{niter<-,NMFfit,numeric-method}
\alias{niter<-,NMFfit,numeric-method}
\title{Sets the number of iteration performed to fit an NMF model.}
\usage{
\S4method{niter}{NMFfit,numeric}(object) <- value
}
\description{
This function is used internally by the function \code{\link{nmf}}.
It is not meant to be called by the user, except when developing 
new NMF algorithms implemented as single function, to set the number 
of iterations performed by the algorithm on the seed, before returning it 
(see \code{\linkS4class{NMFStrategyFunction}}).
}
