% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{nmf,data.frame,ANY,ANY-method}
\alias{nmf,data.frame,ANY,ANY-method}
\title{Fits an NMF model on a \code{data.frame}.}
\usage{
\S4method{nmf}{data.frame,ANY,ANY}(x, rank, method, ...)
}
\description{
The target \code{data.frame} is coerced into a matrix with \code{\link{as.matrix}}.
}
