% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R, R/heatmaps.R
\name{NMF-defunct}
\alias{NMF-defunct}
\alias{metaHeatmap}
\alias{metaHeatmap,matrix-method}
\title{Defunct Functions and Classes in the NMF Package}
\usage{
metaHeatmap(object, ...)

\S4method{metaHeatmap}{matrix}(object, ...)
}
\arguments{
\item{object}{an R object}

\item{...}{other arguments}
}
\description{
Defunct Functions and Classes in the NMF Package
}
\section{Methods (by generic)}{
\itemize{
\item \code{metaHeatmap(object = NMF)}: Deprecated method that is substituted by \code{\link{coefmap}} and \code{\link{basismap}}.

\item \code{metaHeatmap(object = NMFfitX)}: Deprecated method subsituted by \code{\link{consensusmap}}.

\item \code{metaHeatmap(object = matrix)}: Defunct method substituted by \code{\link{aheatmap}}.
}}

