% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\name{NMF-deprecated}
\alias{NMF-deprecated}
\title{Deprecated Functions in the Package NMF}
\arguments{
\item{object}{an R object}

\item{...}{extra arguments}
}
\description{
Deprecated Functions in the Package NMF
}
