% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{nmf,formula,ANY,ANY-method}
\alias{nmf,formula,ANY,ANY-method}
\title{This method implements the interface for fitting formula-based NMF models.
See \code{\link{nmfModel}}.}
\usage{
\S4method{nmf}{formula,ANY,ANY}(x, rank, method, ..., model = NULL)
}
\description{
Argument \code{rank} target matrix or formula environment.
If not missing, \code{model} must be a \code{list}, a \code{data.frame} or 
an \code{environment} in which formula variables are searched for.
}
