\name{Internals}
\Rdversion{1.1}
\alias{NMF-internals}
\alias{nmfAlgorithm}
\alias{nmfSeed}
\alias{algorithm,NMFStrategyFunction-method}
\alias{algorithm<-,NMFStrategyFunction,character-method}
\alias{algorithm<-,NMFStrategyFunction,function-method}
\alias{show,NMFStrategy-method}
\alias{show,NMFStrategyIterative-method}
\alias{name,NMFStrategy-method}
\alias{name<-,NMFStrategy,character-method}
%\alias{nmf,matrix,numeric,NMFStrategy-method}
\alias{objective,NMFStrategy-method}
\alias{objective<-,NMFStrategy,character-method}
\alias{objective<-,NMFStrategy,function-method}

\alias{name,NMFSeed-method}
\alias{name<-,NMFSeed,character-method}
\alias{show,NMFSeed-method}

\alias{$,NMF-method}
\alias{$<-,NMF-method}

\title{Undocumented or Internal methods and functions for package NMF}

\description{
Dummy Rd file that gathers undocumented or internal methods and functions for package NMF.
}

\seealso{
	\code{\linkS4class{NMF}}, \code{\link{nmf-methods}}, package's vignette
}

\author{ Renaud Gaujoux \email{renaud@cbio.uct.ac.za} }
