% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{nmf,matrix,data.frame,ANY-method}
\alias{nmf,matrix,data.frame,ANY-method}
\title{Shortcut for \code{nmf(x, as.matrix(rank), method, ...)}.}
\usage{
\S4method{nmf}{matrix,data.frame,ANY}(x, rank, method, ...)
}
\description{
Shortcut for \code{nmf(x, as.matrix(rank), method, ...)}.
}
