% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{nmf,matrix,matrix,ANY-method}
\alias{nmf,matrix,matrix,ANY-method}
\title{Fits an NMF model partially seeding the computation with a given matrix passed 
in \code{rank}.}
\usage{
\S4method{nmf}{matrix,matrix,ANY}(x, rank, method, seed, model = list(), ...)
}
\description{
The matrix \code{rank} is used either as initial value for the basis or mixture 
coefficient matrix, depending on its dimension.
}
\details{
Currently, such partial NMF model is directly used as a seed, meaning that 
the remaining part is left uninitialised, which is not accepted by all NMF algorithm.
This should change in the future, where the missing part of the model will be 
drawn from some random distribution.

Amongst built-in algorithms, only \sQuote{snmf/l} and \sQuote{snmf/r} support 
partial seeds, with only the coefficient or basis matrix initialised 
respectively.
}
