% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{nmf,matrix,missing,ANY-method}
\alias{nmf,matrix,missing,ANY-method}
\title{Method defined to ensure the correct dispatch to workhorse methods in case
of argument \code{rank} is missing.}
\usage{
\S4method{nmf}{matrix,missing,ANY}(x, rank, method, ...)
}
\description{
Method defined to ensure the correct dispatch to workhorse methods in case
of argument \code{rank} is missing.
}
