% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{nmf,matrix,NMF,ANY-method}
\alias{nmf,matrix,NMF,ANY-method}
\title{Fits an NMF model using the NMF model \code{rank} to seed the computation, 
i.e. as a starting point.}
\usage{
\S4method{nmf}{matrix,NMF,ANY}(x, rank, method, seed, ...)
}
\description{
This method is provided for convenience as a shortcut for 
\code{nmf(x, nbasis(object), method, seed=object, ...)} 
It discards any value passed in argument \code{seed} and uses the NMF model passed 
in \code{rank} instead.
It throws a warning if argument \code{seed} not missing.
}
\details{
If \code{method} is missing, this method will call the method 
\code{nmf,matrix,numeric,NULL}, which will infer an algorithm suitable for fitting an 
NMF model of the class of \code{rank}.
}
