% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{nmf,matrix,NULL,ANY-method}
\alias{nmf,matrix,NULL,ANY-method}
\title{Fits an NMF model using the NMF model supplied in \code{seed}, to seed the computation,
i.e. as a starting point.}
\usage{
\S4method{nmf}{matrix,`NULL`,ANY}(x, rank, method, seed, ...)
}
\description{
This method is provided for completeness and is equivalent to 
\code{nmf(x, seed, method, ...)}.
}
