% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{nmf,matrix,numeric,character-method}
\alias{nmf,matrix,numeric,character-method}
\title{Fits an NMF model on \code{x} using an algorithm registered with access key 
\code{method}.}
\usage{
\S4method{nmf}{matrix,numeric,character}(x, rank, method, ...)
}
\description{
Argument \code{method} is partially match against the access keys of all 
registered algorithms (case insensitive).
Available algorithms are listed in section \emph{Algorithms} below or the 
introduction vignette. 
A vector of their names may be retrieved via \code{nmfAlgorithm()}.
}
\section{Algorithms}{

All algorithms are accessible by their respective access key as listed below.
The following algorithms are available:
\describe{

\item{\sQuote{brunet}}{ Standard NMF, based on the Kullback-Leibler divergence, 
from \cite{Brunet2004}.
It uses simple multiplicative updates from \cite{Lee2001}, enhanced to avoid 
numerical underflow.

Default stopping criterion: invariance of the connectivity matrix
(see \code{\link{nmf.stop.connectivity}}).
}

\item{\sQuote{lee}}{ Standard NMF based on the Euclidean distance from \cite{Lee2001}.
It uses simple multiplicative updates.

Default stopping criterion: invariance of the connectivity matrix
(see \code{\link{nmf.stop.connectivity}}).
}

\item{ls-nmf}{ Least-Square NMF from \cite{Wang2006}.
It uses modified versions of Lee and Seung's multiplicative updates for the 
Euclidean distance, which incorporates weights on each entry of the target 
matrix, e.g. to reflect measurement uncertainty.

Default stopping criterion: stationarity of the objective function
(see \code{\link{nmf.stop.stationary}}).
}

\item{\sQuote{nsNMF}}{ Nonsmooth NMF from \cite{Pascual-Montano2006}. 
It uses a modified version of Lee and Seung's multiplicative updates for the 
Kullback-Leibler divergence \cite{Lee2001}, to fit a extension of the standard 
NMF model, that includes an intermediate smoothing matrix, meant meant to produce 
sparser factors.

Default stopping criterion: invariance of the connectivity matrix
(see \code{\link{nmf.stop.connectivity}}).
}

\item{\sQuote{offset}}{ NMF with offset from \cite{Badea2008}.
It uses a modified version of Lee and Seung's multiplicative 
updates for Euclidean distance \cite{Lee2001}, to fit an NMF model that includes 
an intercept, meant to capture a common baseline and shared patterns, in 
order to produce cleaner basis components.

Default stopping criterion: invariance of the connectivity matrix
(see \code{\link{nmf.stop.connectivity}}).
}

\item{\sQuote{pe-nmf}}{ Pattern-Expression NMF from \emph{Zhang2008}. 
It uses multiplicative updates to minimize an objective function based on the 
Euclidean distance, that is regularized for effective expression of patterns 
with basis vectors.

Default stopping criterion: stationarity of the objective function
(see \code{\link{nmf.stop.stationary}}).
}

\item{\sQuote{snmf/r}, \sQuote{snmf/l}}{ Alternating Least Square (ALS) approach 
from \cite{KimH2007}.
It applies the nonnegative least-squares algorithm from \cite{VanBenthem2004} 
(i.e. fast combinatorial nonnegative least-squares for multiple right-hand), 
to estimate the basis and coefficient matrices alternatively 
(see \code{\link{fcnnls}}).
It minimises an Euclidean-based objective function, that is regularized to 
favour sparse basis matrices (for \sQuote{snmf/l}) or sparse coefficient matrices 
(for \sQuote{snmf/r}).

Stopping criterion: built-in within the internal workhorse function \code{nmf_snmf}, 
based on the KKT optimality conditions.
}

}
}

\section{Seeding methods}{

The purpose of seeding methods is to compute initial values for the factor 
matrices in a given NMF model. 
This initial guess will be used as a starting point by the chosen NMF algorithm.

The seeding method to use in combination with the algorithm can be passed 
to interface \code{nmf} through argument \code{seed}.
The seeding seeding methods available in registry are listed by the function 
\code{\link{nmfSeed}} (see list therein). 

Detailed examples of how to specify the seeding method and its parameters can 
be found in the \emph{Examples} section of this man page and in the package's 
vignette.
}

\seealso{
\code{\link{nmfAlgorithm}}
}
