% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{nmf,matrix,numeric,function-method}
\alias{nmf,matrix,numeric,function-method}
\title{Fits an NMF model on \code{x} using a custom algorithm defined the function 
\code{method}.}
\usage{
\S4method{nmf}{matrix,numeric,`function`}(x, rank, method, seed,
  model = "NMFstd", ..., name, objective = "euclidean", mixed = FALSE)
}
\arguments{
\item{name}{name associated with the NMF algorithm implemented by the function
\code{method} [only used when \code{method} is a function].}

\item{objective}{specification of the objective function associated with the 
algorithm implemented by the function \code{method} 
[only used when \code{method} is a function].

It may be either \code{'euclidean'} or \code{'KL'} for specifying the euclidean 
distance (Frobenius norm) or the Kullback-Leibler divergence respectively, 
or a function with signature \code{(x="NMF", y="matrix", ...)} that computes
the objective value for an NMF model \code{x} on a target matrix \code{y}, 
i.e. the residuals between the target matrix and its NMF estimate.
Any extra argument may be specified, e.g. \code{function(x, y, alpha, beta=2, ...)}.}

\item{mixed}{a logical that indicates if the algorithm implemented by the function
\code{method} support mixed-sign target matrices, i.e. that may contain negative 
values [only used when \code{method} is a function].}
}
\description{
The supplied function must have signature \code{(x=matrix, start=NMF, ...)} 
and return an object that inherits from class \code{\linkS4class{NMF}}. 
It will be called internally by the workhorse \code{nmf} method, with an NMF model 
to be used as a starting point passed in its argument \code{start}.
}
\details{
Extra arguments in \code{...} are passed to \code{method} from the top 
\code{nmf} call.
Extra arguments that have no default value in the definition of the function 
\code{method} are required to run the algorithm (e.g. see argument \code{alpha} 
of \code{myfun} in the examples).

If the algorithm requires a specific type of NMF model, this can be specified 
in argument \code{model} that is handled as in the workhorse \code{nmf} 
method (see description for this argument).
}
