% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{nmf,matrix,numeric,list-method}
\alias{nmf,matrix,numeric,list-method}
\title{Fits multiple NMF models on a common matrix using a list of algorithms.}
\usage{
\S4method{nmf}{matrix,numeric,list}(x, rank, method, ...,
  .parameters = list())
}
\arguments{
\item{.parameters}{list of method-specific parameters.
Its elements must have names matching a single method listed in \code{method},
and be lists of named values that are passed to the corresponding method.}
}
\description{
The models are fitted sequentially with \code{nmf} using the same options 
and parameters for all algorithms.
In particular, irrespective of the way the computation is seeded, this method 
ensures that all fits are performed using the same initial RNG settings.
}
\details{
This method returns an object of class \code{\linkS4class{NMFList}}, that is  
essentially a list containing each fit.
}
