% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{nmf,matrix,numeric,missing-method}
\alias{nmf,matrix,numeric,missing-method}
\title{Method defined to ensure the correct dispatch to workhorse methods in case
of argument \code{method} is missing.}
\usage{
\S4method{nmf}{matrix,numeric,missing}(x, rank, method, ...)
}
\description{
Method defined to ensure the correct dispatch to workhorse methods in case
of argument \code{method} is missing.
}
