% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{nmf,matrix,numeric,NULL-method}
\alias{nmf,matrix,numeric,NULL-method}
\title{Fits an NMF model using an appropriate algorithm when \code{method} is not supplied.}
\usage{
\S4method{nmf}{matrix,numeric,`NULL`}(x, rank, method, seed = NULL,
  model = NULL, ...)
}
\description{
This method tries to select an appropriate algorithm amongst the NMF algorithms 
stored in the internal algorithm registry, which contains the type of NMF models 
each algorithm can fit.
This is possible when the type of NMF model to fit is available from argument \code{seed}, 
i.e. if it is an NMF model itself.
Otherwise the algorithm to use is obtained from \code{nmf.getOption('default.algorithm')}.
}
\details{
This method is provided for internal usage, when called from other \code{nmf} methods 
with argument \code{method} missing in the top call (e.g. \code{nmf,matrix,numeric,missing}).
}
