% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\docType{package}
\name{NMF-package}
\alias{NMF-package}
\alias{_PACKAGE}
\alias{NMF}
\title{NMF: Algorithms and Framework for Nonnegative Matrix Factorization (NMF)}
\description{
This package provides a framework to perform Non-negative Matrix
    Factorization (NMF). It implements a set of already published algorithms
    and seeding methods, and provides a framework to test, develop and plug
    new/custom algorithms. Most of the built-in algorithms have been optimized
    in C++, and the main interface function provides an easy way of performing
    parallel computations on multicore machines.
}
\examples{
# generate a synthetic dataset with known classes
n <- 50; counts <- c(5, 5, 8);
V <- syntheticNMF(n, counts)

# perform a 3-rank NMF using the default algorithm
res <- nmf(V, 3)

basismap(res)
coefmap(res)

}
\references{
\url{https://cran.r-project.org/}

\url{https://renozao.github.io/NMF}
}
\seealso{
\code{\link{nmf}}
}
\keyword{package}
