% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{nmf.equal,list,list-method}
\alias{nmf.equal,list,list-method}
\title{Compares the results of multiple NMF runs.}
\usage{
\S4method{nmf.equal}{list,list}(x, y, ..., all = FALSE, vector = FALSE)
}
\arguments{
\item{all}{a logical that indicates if all fits should be compared separately
or only the best fits}

\item{vector}{a logical, only used when \code{all=TRUE}, that indicates if 
all fits must be equal for \code{x} and \code{y} to be declared equal, or 
if one wants to return the result of each comparison in a vector.}
}
\description{
This method either compare the two best fit, or all fits separately.
All extra arguments in \code{...} are passed to each internal call to 
\code{nmf.equal}.
}
