% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{nmf.equal,list,missing-method}
\alias{nmf.equal,list,missing-method}
\title{Compare all elements in \code{x} to \code{x[[1]]}.}
\usage{
\S4method{nmf.equal}{list,missing}(x, y, ...)
}
\description{
Compare all elements in \code{x} to \code{x[[1]]}.
}
