% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{nmf.equal,NMF,NMF-method}
\alias{nmf.equal,NMF,NMF-method}
\title{Compares two NMF models.}
\usage{
\S4method{nmf.equal}{NMF,NMF}(x, y, identical = TRUE, ...)
}
\description{
Arguments in \code{...} are used only when \code{identical=FALSE} and are
passed to \code{all.equal}.
}
