% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{nmf.equal,NMF,NMFfitX-method}
\alias{nmf.equal,NMF,NMFfitX-method}
\title{Compares two NMF models when at least one comes from multiple NMF runs.}
\usage{
\S4method{nmf.equal}{NMF,NMFfitX}(x, y, ...)
}
\description{
Compares two NMF models when at least one comes from multiple NMF runs.
}
