% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{nmf.equal,NMFfit,NMF-method}
\alias{nmf.equal,NMFfit,NMF-method}
\title{Compares two NMF models when at least one comes from a NMFfit object, 
i.e. an object returned by a single run of \code{\link{nmf}}.}
\usage{
\S4method{nmf.equal}{NMFfit,NMF}(x, y, ...)
}
\description{
Compares two NMF models when at least one comes from a NMFfit object, 
i.e. an object returned by a single run of \code{\link{nmf}}.
}
