% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{nmf.equal,NMFfit,NMFfit-method}
\alias{nmf.equal,NMFfit,NMFfit-method}
\title{Compares two fitted NMF models, i.e. objects returned by single runs of 
\code{\link{nmf}}.}
\usage{
\S4method{nmf.equal}{NMFfit,NMFfit}(x, y, ...)
}
\description{
Compares two fitted NMF models, i.e. objects returned by single runs of 
\code{\link{nmf}}.
}
