% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{nmf.equal,NMFfitX1,NMFfitX1-method}
\alias{nmf.equal,NMFfitX1,NMFfitX1-method}
\title{Compares the NMF models fitted by multiple runs, that only kept the best fits.}
\usage{
\S4method{nmf.equal}{NMFfitX1,NMFfitX1}(x, y, ...)
}
\description{
Compares the NMF models fitted by multiple runs, that only kept the best fits.
}
