% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{runNmf}
\alias{runNmf}
\title{Run NNLM::nnmf() and calculate MSE}
\usage{
runNmf(
  A,
  k = 1,
  repeat.num = 1,
  max.samples = NULL,
  perf.metrics = c("mse_imputed", "mse_perm"),
  impute.prop = 0.1,
  return.perf = 1,
  seed = 1,
  verbose = F
)
}
\description{
Run NNLM::nnmf() and calculate MSE
}
