% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{NMFfitX,list-method}
\alias{NMFfitX,list-method}
\title{Create an \code{NMFfitX} object from a list of fits.}
\usage{
\S4method{NMFfitX}{list}(object, ..., .merge = FALSE)
}
\arguments{
\item{.merge}{a logical that indicates if the fits should be aggregated, only 
keeping the best fit, and return an \code{NMFfitX1} object.
If \code{FALSE}, an \code{NMFfitXn} object containing the data of all the fits
is returned.}
}
\description{
Create an \code{NMFfitX} object from a list of fits.
}
