% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{NMFfitX,NMFfit-method}
\alias{NMFfitX,NMFfit-method}
\title{Creates an \code{NMFfitX1} object from a single fit.
This is used in \code{\link{nmf}} when only the best fit is kept in memory or 
on disk.}
\usage{
\S4method{NMFfitX}{NMFfit}(object, ...)
}
\description{
Creates an \code{NMFfitX1} object from a single fit.
This is used in \code{\link{nmf}} when only the best fit is kept in memory or 
on disk.
}
