% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{NMFfitX,NMFfitX-method}
\alias{NMFfitX,NMFfitX-method}
\title{Provides a way to aggregate \code{NMFfitXn} objects into an \code{NMFfitX1} 
object.}
\usage{
\S4method{NMFfitX}{NMFfitX}(object, ...)
}
\description{
Provides a way to aggregate \code{NMFfitXn} objects into an \code{NMFfitX1} 
object.
}
