% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R,
%   R/NMFStrategyFunction-class.R, R/NMFStrategyIterative-class.R
\name{nmfFormals}
\alias{nmfFormals}
\alias{nmfFormals.character}
\alias{nmfFormals.NMFStrategy}
\alias{nmfArgs}
\alias{nmfFormals.NMFStrategyFunction}
\alias{nmfFormals.NMFStrategyIterative}
\title{Showing Arguments of NMF Algorithms}
\usage{
nmfFormals(x, ...)

\method{nmfFormals}{character}(x, ...)

\method{nmfFormals}{NMFStrategy}(x, ...)

nmfArgs(x)

\method{nmfFormals}{NMFStrategyFunction}(x, ...)

\method{nmfFormals}{NMFStrategyIterative}(x, runtime = FALSE, ...)
}
\arguments{
\item{x}{algorithm specification}

\item{...}{extra argument to allow extension}
}
\description{
This function returns the extra arguments that can be passed
to a given NMF algorithm in call to \code{\link{nmf}}.
}
\examples{

# show arguments of an NMF algorithm
nmfArgs('brunet')
nmfArgs('snmf/r')
}
