% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfModel.R
\docType{methods}
\name{nmfModel,data.frame,data.frame-method}
\alias{nmfModel,data.frame,data.frame-method}
\title{Same as \code{nmfModel('matrix', 'matrix')} but for \code{data.frame} objects,
which are generally produced by \code{\link{read.delim}}-like functions.}
\usage{
\S4method{nmfModel}{data.frame,data.frame}(rank, target = 0L, ...)
}
\description{
The input \code{data.frame} objects are converted into matrices with 
\code{\link{as.matrix}}.
}
