% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfModel.R
\docType{methods}
\name{nmfModel,formula,ANY-method}
\alias{nmfModel,formula,ANY-method}
\title{Build a formula-based NMF model, that can incorporate fixed basis or 
coefficient terms.}
\usage{
\S4method{nmfModel}{formula,ANY}(rank, target, ..., data = NULL,
  no.attrib = FALSE)
}
\arguments{
\item{data}{Optional argument where to look for the variables used in the 
formula.}

\item{no.attrib}{logical that indicate if attributes containing data related
to the formula should be attached as attributes. 
If \code{FALSE} attributes \code{'target'} and \code{'formula'} contain the 
target matrix, and a list describing each formula part (response, regressors, 
etc.).}
}
\description{
Build a formula-based NMF model, that can incorporate fixed basis or 
coefficient terms.
}
\examples{

# empty 3-rank model
nmfModel(~ 3)

# 3-rank model that fits a given data matrix
x <- rmatrix(20,10)
nmfModel(x ~ 3)

# add fixed coefficient term defined by a factor
gr <- gl(2, 5)
nmfModel(x ~ 3 + gr)

# add fixed coefficient term defined by a numeric covariate
nmfModel(x ~ 3 + gr + b, data=list(b=runif(10)))

# 3-rank model that fits a given ExpressionSet (with fixed coef terms)
e <- ExpressionSet(x)
pData(e) <- data.frame(a=runif(10))
nmfModel(e ~ 3 + gr + a) # `a` is looked up in the phenotypic data of x pData(x)

}
