% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfModel.R
\docType{methods}
\name{nmfModel,matrix,matrix-method}
\alias{nmfModel,matrix,matrix-method}
\title{Creates an NMF model based on two existing factors.}
\usage{
\S4method{nmfModel}{matrix,matrix}(rank, target = 0L, ...)
}
\description{
This method is equivalent to \code{nmfModel(0, 0, W=rank, H=target..., force.dim=FALSE)}.
This allows for a natural shortcut for wrapping existing \strong{compatible} 
matrices into NMF models:
\samp{nmfModel(w, h)}
}
\details{
Note that an error is thrown if their dimensions are not compatible.
}
\examples{
## From two existing factors

# allows a convenient call without argument names
w <- rmatrix(n, 3); h <- rmatrix(3, p)
nmfModel(w, h)

# Specify the type of NMF model (e.g. 'NMFns' for non-smooth NMF)
mod <- nmfModel(w, h, model='NMFns')
mod

# One can use such an NMF model as a seed when fitting a target matrix with nmf()
V <- rmatrix(mod)
res <- nmf(V, mod)
nmf.equal(res, nmf(V, mod))

# NB: when called only with such a seed, the rank and the NMF algorithm 
# are selected based on the input NMF model. 
# e.g. here rank was 3 and the algorithm "nsNMF" is used, because it is the default 
# algorithm to fit "NMFns" models (See ?nmf). 

}
