% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfModel.R
\docType{methods}
\name{nmfModel,missing,ANY-method}
\alias{nmfModel,missing,ANY-method}
\title{Creates an empty NMF model of null rank and a given dimension.}
\usage{
\S4method{nmfModel}{missing,ANY}(rank, target = 0L, ...)
}
\description{
This call is equivalent to \code{nmfModel(0, target, ...)}.
}
\examples{
nmfModel(target=10) #square
nmfModel(target=c(10, 5))

}
