% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfModel.R
\docType{methods}
\name{nmfModel,missing,missing-method}
\alias{nmfModel,missing,missing-method}
\title{Creates an empty NMF model or from existing factors}
\usage{
\S4method{nmfModel}{missing,missing}(rank, target = 0L, ...)
}
\description{
This method is equivalent to \code{nmfModel(0, 0, ..., force.dim=FALSE)}.	
This means that the dimensions of the NMF model will be taken from the optional 
basis and mixture coefficient arguments \code{W} and \code{H}.
An error is thrown if their dimensions are not compatible.
}
\details{
Hence, this method may be used to generate an NMF model from existing factor 
matrices, by providing the named arguments \code{W} and/or \code{H}:

\code{nmfModel(W=w)} or \code{nmfModel(H=h)} or \code{nmfModel(W=w, H=h)}

Note that this may be achieved using the more convenient interface is 
provided by the method \code{nmfModel,matrix,matrix} (see its dedicated description).

See the description of the appropriate method below.
}
\examples{

# Build an empty NMF model 
nmfModel()

# create a NMF object based on one random matrix: the missing matrix is deduced
# Note this only works when using factory method NMF 
n <- 50; r <- 3; 
w <- rmatrix(n, r) 
nmfModel(W=w)

# create a NMF object based on random (compatible) matrices
p <- 20
h <- rmatrix(r, p)
nmfModel(H=h)

# specifies two compatible matrices
nmfModel(W=w, H=h)
# error if not compatible
try( nmfModel(W=w, H=h[-1,]) ) 

}
