% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfModel.R
\docType{methods}
\name{nmfModel,NULL,ANY-method}
\alias{nmfModel,NULL,ANY-method}
\title{Creates an empty NMF model of null rank and given dimension.}
\usage{
\S4method{nmfModel}{`NULL`,ANY}(rank, target = 0L, ...)
}
\description{
This call is equivalent to \code{nmfModel(0, target, ...)}, and is meant for 
internal usage only.
}
