% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfModel.R
\docType{methods}
\name{nmfModel,numeric,matrix-method}
\alias{nmfModel,numeric,matrix-method}
\title{Creates an NMF model compatible with a target matrix.}
\usage{
\S4method{nmfModel}{numeric,matrix}(rank, target, ..., use.names = TRUE)
}
\arguments{
\item{use.names}{a logical that indicates whether the dimension names of the 
target matrix should be set on the returned NMF model.}
}
\description{
This call is equivalent to \code{nmfModel(rank, dim(target), ...)}.
That is that the returned NMF object fits a target matrix of the same 
dimension as \code{target}.
}
\details{
Only the dimensions of \code{target} are used to construct the \code{NMF} object. 
The matrix slots are filled with \code{NA} values if these are not specified 
in arguments \code{W} and/or \code{H}.
However, dimension names are set on the return NMF model if present in 
\code{target} and argument \code{use.names=TRUE}.
}
\examples{
 
# create a r-ranked NMF model compatible with a given target matrix
obj <- nmfModel(r, V)
all(is.na(basis(obj)))

}
