% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfModel.R
\docType{methods}
\name{nmfModel,numeric,missing-method}
\alias{nmfModel,numeric,missing-method}
\title{Creates an empty NMF model of a given rank.}
\usage{
\S4method{nmfModel}{numeric,missing}(rank, target = 0L, ...)
}
\description{
This call is equivalent to \code{nmfModel(rank, 0L, ...)}, which 
creates \emph{empty} \code{NMF} object with a basis and mixture coefficient matrix  
of dimension 0 x \code{rank} and \code{rank} x 0 respectively.
}
\examples{
## Empty model of given rank
nmfModel(3)
 
}
\seealso{
\code{\link{is.empty.nmf}}
}
