% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSeed-class.R
\docType{methods}
\name{NMFSeed,character-method}
\alias{NMFSeed,character-method}
\title{Default method simply calls \code{\link{new}} with the same arguments.}
\usage{
\S4method{NMFSeed}{character}(key, method, ...)
}
\description{
Default method simply calls \code{\link{new}} with the same arguments.
}
