% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSeed-class.R
\docType{class}
\name{NMFSeed-class}
\alias{NMFSeed-class}
\alias{show,NMFSeed-method}
\alias{algorithm,NMFSeed-method}
\alias{algorithm<-,NMFSeed,function-method}
\title{Base class that defines the interface for NMF seeding methods.}
\usage{
\S4method{show}{NMFSeed}(object)

\S4method{algorithm}{NMFSeed}(object)

\S4method{algorithm}{NMFSeed,`function`}(object) <- value
}
\arguments{
\item{object}{an object of class \code{NMFSeed}.}

\item{value}{object whose value is used to modify properties of the given object.
See details for each method.}
}
\description{
This class implements a simple wrapper strategy object that defines a unified
interface to seeding methods, that are used to initialise NMF models before
fitting them with any NMF algorithm.
}
\section{Slots}{

\describe{
\item{\code{name}}{character string giving the name of the seeding strategy}

\item{\code{method}}{workhorse function that implements the seeding strategy.
It must have signature \code{(object="NMF", x="matrix", ...)} and initialise
the NMF model \code{object} with suitable values for fitting the target
matrix \code{x}.}
}}

\section{Methods (by generic)}{
\bold{algorithm:}

\itemize{
\item \code{algorithm(object = NMFSeed)}: Returns the workhorse function of the seeding method described by \code{object}.
}


\bold{algorithm<-:}

\itemize{
\item \code{algorithm<-(object = NMFSeed,value = function)}: Sets the workhorse function of the seeding method described by \code{object}.
}


\bold{NMFSeed:}

\itemize{
\item \code{NMFSeed(key = NMFSeed)}: Creates an \code{NMFSeed} based on a template object (Constructor-Copy),
in particular it uses the \strong{same} name.
}


\bold{seed:}

\itemize{
\item \code{seed(x = mMatrix,model = NMF,method = NMFSeed)}: This is the workhorse method that seeds an NMF model object using a given
seeding strategy defined by an \code{NMFSeed} object, to fit a given
target matrix.

\item \code{seed(x = ANY,model = list,method = NMFSeed)}: Seed a model using the elements in \code{model} to instantiate it with
\code{\link{nmfModel}}.

\item \code{seed(x = ANY,model = numeric,method = NMFSeed)}: Seeds a standard NMF model (i.e. of class \code{\linkS4class{NMFstd}}) of rank
\code{model}.
}


\bold{show:}

\itemize{
\item \code{show(object = NMFSeed)}: Show method for objects of class \code{NMFSeed}
}

}

