% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSeed-class.R
\docType{methods}
\name{NMFSeed,NMFSeed-method}
\alias{NMFSeed,NMFSeed-method}
\title{Creates an \code{NMFSeed} based on a template object (Constructor-Copy), 
in particular it uses the \strong{same} name.}
\usage{
\S4method{NMFSeed}{NMFSeed}(key, method, ...)
}
\description{
Creates an \code{NMFSeed} based on a template object (Constructor-Copy), 
in particular it uses the \strong{same} name.
}
