% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{methods}
\name{NMFStrategy,character,character-method}
\alias{NMFStrategy,character,character-method}
\title{Creates an \code{NMFStrategy} based on a registered NMF algorithm that is used 
as a template.}
\usage{
\S4method{NMFStrategy}{character,character}(name, method, ...)
}
\description{
Creates an \code{NMFStrategy} based on a registered NMF algorithm that is used 
as a template.
}
