% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{methods}
\name{NMFStrategy,character,function-method}
\alias{NMFStrategy,character,function-method}
\title{Creates an \code{NMFStrategyFunction} object that wraps the function \code{method}
into a unified interface.}
\usage{
\S4method{NMFStrategy}{character,`function`}(name, method, ...)
}
\description{
\code{method} must be a function with signature \code{(y="matrix", x="NMFfit", ...)}, 
and return an object of class \code{\linkS4class{NMFfit}}.
}
