% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{methods}
\name{NMFStrategy,character,missing-method}
\alias{NMFStrategy,character,missing-method}
\title{Creates an NMFStrategy, determining its type from the extra arguments passed 
in \code{...}: if there is an argument named \code{Update} then an 
\code{NMFStrategyIterative} is created, or if there is an argument 
named \code{algorithm} then an \code{NMFStrategyFunction} is created.
Calls other than these generates an error.}
\usage{
\S4method{NMFStrategy}{character,missing}(name, method, ...)
}
\description{
Creates an NMFStrategy, determining its type from the extra arguments passed 
in \code{...}: if there is an argument named \code{Update} then an 
\code{NMFStrategyIterative} is created, or if there is an argument 
named \code{algorithm} then an \code{NMFStrategyFunction} is created.
Calls other than these generates an error.
}
