% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{methods}
\name{NMFStrategy,character,NMFStrategy-method}
\alias{NMFStrategy,character,NMFStrategy-method}
\title{Creates an \code{NMFStrategy} object based on a template object (Constructor-Copy).}
\usage{
\S4method{NMFStrategy}{character,NMFStrategy}(name, method, ...)
}
\description{
Creates an \code{NMFStrategy} object based on a template object (Constructor-Copy).
}
