% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{class}
\name{NMFStrategy-class}
\alias{NMFStrategy-class}
\alias{show,NMFStrategy-method}
\alias{objective,NMFStrategy-method}
\alias{objective<-,NMFStrategy,character-method}
\alias{objective<-,NMFStrategy,function-method}
\alias{is.mixed}
\title{Virtual Interface for NMF Algorithms}
\usage{
\S4method{show}{NMFStrategy}(object)

\S4method{objective}{NMFStrategy}(object)

\S4method{objective}{NMFStrategy,character}(object) <- value

\S4method{objective}{NMFStrategy,`function`}(object) <- value

is.mixed(object)
}
\description{
This class partially implements the generic interface defined for general
algorithms defined in the \pkg{NMF} package (see \code{\link{algorithmic-NMF}}).
}
\section{Functions}{
\itemize{
\item \code{is.mixed}: \code{is.mixed} tells if an NMF algorithm works on mixed-sign data.
}}

\section{Slots}{

\describe{
\item{\code{objective}}{the objective function associated with the algorithm (Frobenius, Kullback-Leibler, etc...).
It is either an access key of a registered objective function or a function definition.
In the latter case, the given function must have the following signature \code{(x="NMF", y="matrix")}
and return a nonnegative real value.}

\item{\code{model}}{a character string giving either the (sub)class name of the NMF-class instance used
and returned by the strategy, or a function name.}

\item{\code{mixed}}{a logical that indicates if the algorithm works on mixed-sign data.}
}}

\keyword{internal}
\section{Methods (by generic)}{
\bold{canFit:}

\itemize{
\item \code{canFit(x = NMFStrategy,y = character)}: Tells if an NMF algorithm can fit a given class of NMF models

\item \code{canFit(x = NMFStrategy,y = NMF)}: Tells if an NMF algorithm can fit the same class of models as \code{y}
}


\bold{deviance:}

\itemize{
\item \code{deviance(object = NMFStrategy)}: Computes the value of the objective function between the estimate \code{x}
and the target \code{y}.
}


\bold{modelname:}

\itemize{
\item \code{modelname(object = NMFStrategy)}: Returns the model(s) that an NMF algorithm can fit.
}


\bold{NMFStrategy:}

\itemize{
\item \code{NMFStrategy(name = character,method = NMFStrategy)}: Creates an \code{NMFStrategy} object based on a template object (Constructor-Copy).

\item \code{NMFStrategy(name = NMFStrategy,method = missing)}: Creates an \code{NMFStrategy} based on a template object (Constructor-Copy),
in particular it uses the \strong{same} name.

\item \code{NMFStrategy(name = NULL,method = NMFStrategy)}: Creates an \code{NMFStrategy} based on a template object (Constructor-Copy)
but using a randomly generated name.
}

}

