% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{methods}
\name{NMFStrategy,missing,character-method}
\alias{NMFStrategy,missing,character-method}
\title{Creates an \code{NMFStrategy} based on a registered NMF algorithm that is used 
as a template (Constructor-Copy), in particular it uses the \strong{same} name.}
\usage{
\S4method{NMFStrategy}{missing,character}(name, method, ...)
}
\description{
It is a shortcut for \code{NMFStrategy(nmfAlgorithm(method, exact=TRUE), ...)}.
}
