% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{methods}
\name{NMFStrategy,NMFStrategy,missing-method}
\alias{NMFStrategy,NMFStrategy,missing-method}
\title{Creates an \code{NMFStrategy} based on a template object (Constructor-Copy), 
in particular it uses the \strong{same} name.}
\usage{
\S4method{NMFStrategy}{NMFStrategy,missing}(name, method, ...)
}
\description{
Creates an \code{NMFStrategy} based on a template object (Constructor-Copy), 
in particular it uses the \strong{same} name.
}
