% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{methods}
\name{NMFStrategy,NULL,character-method}
\alias{NMFStrategy,NULL,character-method}
\title{Creates an \code{NMFStrategy} based on a registered NMF algorithm (Constructor-Copy) 
using a randomly generated name.}
\usage{
\S4method{NMFStrategy}{`NULL`,character}(name, method, ...)
}
\description{
It is a shortcut for \code{NMFStrategy(NULL, nmfAlgorithm(method), ...)}.
}
