% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{methods}
\name{NMFStrategy,NULL,NMFStrategy-method}
\alias{NMFStrategy,NULL,NMFStrategy-method}
\title{Creates an \code{NMFStrategy} based on a template object (Constructor-Copy) 
but using a randomly generated name.}
\usage{
\S4method{NMFStrategy}{`NULL`,NMFStrategy}(name, method, ...)
}
\description{
Creates an \code{NMFStrategy} based on a template object (Constructor-Copy) 
but using a randomly generated name.
}
