% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategyFunction-class.R
\docType{class}
\name{NMFStrategyFunction-class}
\alias{NMFStrategyFunction-class}
\title{Interface for Single Function NMF Strategies}
\description{
This class implements the virtual interface \code{\link{NMFStrategy}} for
NMF algorithms that are implemented by a single workhorse R function.
}
\section{Slots}{

\describe{
\item{\code{algorithm}}{a function that implements an NMF algorithm.
It must have signature \code{(y='matrix', x='NMFfit')}, where \code{y} is the
target matrix to approximate and \code{x} is the NMF model assumed to be
seeded with an appropriate initial value -- as it is done internally by
function \code{\link{nmf}}.

Note that argument names currently do not matter, but it is recommended to
name them as specified above.}
}}

\section{Methods (by generic)}{
\itemize{
\item \code{algorithm(object = NMFStrategyFunction)}: Returns the single R function that implements the NMF algorithm -- as stored in
slot \code{algorithm}.

\item \code{algorithm<-(object = NMFStrategyFunction,value = function)}: Sets the function that implements the NMF algorithm, stored in slot \code{algorithm}.
}}

