% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\docType{methods}
\name{nneg,matrix-method}
\alias{nneg,matrix-method}
\title{Transforms a mixed-sign matrix into a nonnegative matrix, optionally apply a
lower threshold. 
This is the workhorse method, that is eventually called by all other 
methods defined in the \code{\link{NMF}} package.}
\usage{
\S4method{nneg}{matrix}(object, method = c("pmax", "posneg", "absolute",
  "min"), threshold = 0, shift = TRUE)
}
\arguments{
\item{method}{Name of the transformation method to use, that is partially 
matched against the following possible methods:
\describe{
\item{pmax}{Each entry is constrained to be above threshold \code{threshold}.}

\item{posneg}{The matrix is split into its "positive" and "negative" parts, 
with the entries of each part constrained to be above threshold \code{threshold}.
The result consists in these two parts stacked in rows (i.e. \code{\link{rbind}}-ed)
into a single matrix, which has double the number of rows of the input 
matrix \code{object}.}

\item{absolute}{The absolute value of each entry is constrained to be above 
threshold \code{threshold}.}

\item{min}{Global shift by adding the minimum entry to each entry, only if 
it is negative, and then apply threshold.
}

}}

\item{threshold}{Nonnegative lower threshold value (single numeric). 
See argument \code{shit} for details on how the threshold is used and affects
the result.}

\item{shift}{a logical indicating whether the entries below the threshold 
value \code{threshold} should be forced (shifted) to 0 (default) or to 
the threshold value itself. 
In other words, if \code{shift=TRUE} (default) all entries in 
the result matrix are either 0 or strictly greater than \code{threshold}.
They are all greater or equal than \code{threshold} otherwise.}
}
\description{
Transforms a mixed-sign matrix into a nonnegative matrix, optionally apply a
lower threshold. 
This is the workhorse method, that is eventually called by all other 
methods defined in the \code{\link{NMF}} package.
}
\examples{

# random mixed sign data (normal distribution)
set.seed(1)
x <- rmatrix(5,5, rnorm, mean=0, sd=5)
x

# pmax (default)
nneg(x)
# using a threshold
nneg(x, threshold=2)
# without shifting the entries lower than threshold
nneg(x, threshold=2, shift=FALSE)

# posneg: split positive and negative part
nneg(x, method='posneg')
nneg(x, method='pos', threshold=2)

# absolute
nneg(x, method='absolute')
nneg(x, method='abs', threshold=2)

# min
nneg(x, method='min')
nneg(x, method='min', threshold=2)

}
\seealso{
\code{\link{pmax}}
}
