% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\docType{methods}
\name{nneg,NMF-method}
\alias{nneg,NMF-method}
\title{Apply \code{nneg} to the basis matrix of an \code{\link{NMF}} 
object (i.e. \code{basis(object)}).
All extra arguments in \code{...} are passed to the method \code{nneg,matrix}.}
\usage{
\S4method{nneg}{NMF}(object, ...)
}
\description{
Apply \code{nneg} to the basis matrix of an \code{\link{NMF}} 
object (i.e. \code{basis(object)}).
All extra arguments in \code{...} are passed to the method \code{nneg,matrix}.
}
\examples{

# random 
M <- nmfModel(x, rmatrix(ncol(x), 3))
nnM <- nneg(M) 
basis(nnM)
# mixture coefficients are not affected
identical( coef(M), coef(nnM) )

}
