% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{nneg}
\alias{nneg}
\alias{posneg}
\alias{rposneg}
\title{Transforming from Mixed-sign to Nonnegative Data}
\usage{
nneg(object, ...)

posneg(...)

rposneg(object, ...)
}
\arguments{
\item{object}{The data object to transform}

\item{...}{extra arguments to allow extension or passed down to \code{nneg,matrix}
or \code{rposneg,matrix} in subsequent calls.}
}
\value{
an object of the same class as argument \code{object}.

an object of the same type of \code{object}
}
\description{
\code{nneg}  is a generic function to transform a data objects that 
contains negative values into a similar object that only contains 
values that are nonnegative or greater than a given threshold.

\code{posneg} is a shortcut for \code{nneg(..., method='posneg')}, to split 
mixed-sign data into its positive and negative part. 
See description for method \code{"posneg"}, in \code{\link{nneg}}.

\code{rposneg} performs the "reverse" transformation of the \code{\link{posneg}} function.
}
\examples{
# shortcut for the "posneg" transformation
posneg(x)
posneg(x, 2)

}
\seealso{
Other transforms: \code{\link{t.NMF}}
}
