% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithmic.R
\docType{methods}
\name{nrun,ANY-method}
\alias{nrun,ANY-method}
\title{Default method that returns the value of attribute \sQuote{nrun}.}
\usage{
\S4method{nrun}{ANY}(object)
}
\description{
Such an attribute my be attached to objects to keep track of data about 
the parent fit object (e.g. by method \code{\link{consensus}}), which 
can be used by subsequent function calls such as plot functions 
(e.g. see \code{\link{consensusmap}}). 
This method returns \code{NULL} if no suitable data was found.
}
