% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{nrun,NMFfit-method}
\alias{nrun,NMFfit-method}
\title{This method always returns 1, since an \code{NMFfit} object is obtained 
from a single NMF run.}
\usage{
\S4method{nrun}{NMFfit}(object)
}
\description{
This method always returns 1, since an \code{NMFfit} object is obtained 
from a single NMF run.
}
