% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{nrun,NMFfitX-method}
\alias{nrun,NMFfitX-method}
\title{Returns the number of NMF runs performed to create \code{object}.}
\usage{
\S4method{nrun}{NMFfitX}(object)
}
\description{
It is a pure virtual method defined to ensure \code{nrun} is defined 
for sub-classes of \code{NMFfitX}, which throws an error if called.
}
\details{
Note that because the \code{\link{nmf}} function allows to run the NMF 
computation keeping only the best fit, \code{nrun} may return a value 
greater than one, while only the result of the best run is stored in 
the object (cf. option \code{'k'} in method \code{\link{nmf}}).
}
