% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{nrun,NMFfitX1-method}
\alias{nrun,NMFfitX1-method}
\title{Returns the number of NMF runs performed, amongst which \code{object} was 
selected as the best fit.}
\usage{
\S4method{nrun}{NMFfitX1}(object)
}
\description{
Returns the number of NMF runs performed, amongst which \code{object} was 
selected as the best fit.
}
