% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{nrun,NMFfitXn-method}
\alias{nrun,NMFfitXn-method}
\title{Returns the number of runs performed to compute the fits stored in the list 
(i.e. the length of the list itself).}
\usage{
\S4method{nrun}{NMFfitXn}(object)
}
\description{
Returns the number of runs performed to compute the fits stored in the list 
(i.e. the length of the list itself).
}
