% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{objective,NMFfit-method}
\alias{objective,NMFfit-method}
\title{Returns the objective function associated with the algorithm that computed the 
fitted NMF model \code{object}, or the objective value with respect to a given 
target matrix \code{y} if it is supplied.}
\usage{
\S4method{objective}{NMFfit}(object, y)
}
\arguments{
\item{y}{optional target matrix used to compute the objective value.}
}
\description{
Returns the objective function associated with the algorithm that computed the 
fitted NMF model \code{object}, or the objective value with respect to a given 
target matrix \code{y} if it is supplied.
}
