\docType{methods}
\name{offset,NMFfit-method}
\alias{offset,NMFfit-method}
\title{Returns the offset from the fitted model.}
\usage{
  \S4method{offset}{NMFfit}(object)
}
\arguments{
  \item{object}{An offset to be included in a model frame}
}
\description{
  Returns the offset from the fitted model.
}
\keyword{methods}

