% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{offset,NMFfit-method}
\alias{offset,NMFfit-method}
\title{Returns the offset from the fitted model.}
\usage{
\S4method{offset}{NMFfit}(object)
}
\description{
Returns the offset from the fitted model.
}
