% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFOffset-class.R
\docType{methods}
\name{offset,NMFOffset-method}
\alias{offset,NMFOffset-method}
\title{Offsets in NMF Models with Offset}
\usage{
\S4method{offset}{NMFOffset}(object)
}
\arguments{
\item{object}{an instance of class \code{NMFOffset}.}
}
\description{
The function \code{offset} returns the offset vector from an NMF model 
that has an offset, e.g. an \code{NMFOffset} model.
}
