% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{predict,NMF-method}
\alias{predict,NMF-method}
\title{Default method for NMF models}
\usage{
\S4method{predict}{NMF}(object, what = c("columns", "rows", "samples",
  "features"), prob = FALSE, dmatrix = FALSE)
}
\arguments{
\item{what}{a character string that indicates the type of cluster membership should
be returned: \sQuote{columns} or \sQuote{rows} for clustering the colmuns or the
rows of the target matrix respectively.
The values \sQuote{samples} and \sQuote{features} are aliases for \sQuote{colmuns}
and \sQuote{rows} respectively.}

\item{prob}{logical that indicates if the relative contributions of/to the dominant
basis component should be computed and returned. See \emph{Details}.}

\item{dmatrix}{logical that indicates if a dissimiliarity matrix should be
attached to the result.
This is notably used internally when computing NMF clustering silhouettes.}
}
\description{
Default method for NMF models
}
\examples{

# random target matrix
v <- rmatrix(20, 10)
# fit an NMF model
x <- nmf(v, 5)

# predicted column and row clusters
predict(x)
predict(x, 'rows')

# with relative contributions of each basis component
predict(x, prob=TRUE)
predict(x, 'rows', prob=TRUE)

}
