% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{predict,NMFfitX-method}
\alias{predict,NMFfitX-method}
\title{Returns the cluster membership index from an NMF model fitted with multiple 
runs.}
\usage{
\S4method{predict}{NMFfitX}(object, what = c("columns", "rows", "samples",
  "features", "consensus", "chc"), dmatrix = FALSE, ...)
}
\description{
Besides the type of clustering available for any NMF models 
(\code{'columns', 'rows', 'samples', 'features'}), this method can return 
the cluster membership index based on the consensus matrix, computed from 
the multiple NMF runs.
}
\details{
Argument \code{what} accepts the following extra types:
\describe{
\item{\code{'chc'}}{ returns the cluster membership based on the 
hierarchical clustering of the consensus matrix, as performed by 
\code{\link{consensushc}}.}
\item{\code{'consensus'}}{ same as \code{'chc'} but the levels of the membership 
index are re-labeled to match the order of the clusters as they would be displayed on the 
associated dendrogram, as re-ordered on the default annotation track in consensus 
heatmap produced by \code{\link{consensusmap}}.}
}
}
