% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\name{predict}
\alias{predict}
\title{Clustering and Prediction}
\usage{
predict(object, ...)
}
\arguments{
\item{object}{an NMF model}
}
\description{
The methods \code{predict} for NMF models return the cluster membership
of each sample or each feature.
Currently the classification/prediction of new data is not implemented.
}
\details{
The cluster membership is computed as the index of the dominant basis
component for each sample (\code{what='samples' or 'columns'}) or each feature
(\code{what='features' or 'rows'}), based on their corresponding
entries in the coefficient matrix or basis matrix respectively.

For example, if \code{what='samples'}, then the dominant basis component
is computed for each column of the coefficient matrix as the row index
of the maximum within the column.

If argument \code{prob=FALSE} (default), the result is a \code{factor}.
Otherwise a list with two elements is returned: element \code{predict}
contains the cluster membership index (as a \code{factor}) and element
\code{prob} contains the relative contribution of
the dominant component to each sample (resp. the relative contribution of
each feature to the dominant basis component):

\itemize{
\item Samples: \deqn{p_j = x_{k_0} / \sum_k x_k}{p(j) = x(k0) / sum_k x(k)},
for each sample \eqn{1\leq j \leq p}, where \eqn{x_k}{x(k)} is the contribution
of the \eqn{k}-th basis component to \eqn{j}-th sample (i.e. \code{H[k ,j]}), and
\eqn{x_{k_0}}{x(k0)} is the maximum of these contributions.

\item Features: \deqn{p_i = y_{k_0} / \sum_k y_k}{p(i) = y(k0) / sum_k y(k)},
for each feature \eqn{1\leq i \leq p}, where \eqn{y_k}{y(k)} is the contribution
of the \eqn{k}-th basis component to \eqn{i}-th feature (i.e. \code{W[i, k]}), and
\eqn{y_{k_0}}{y(k0)} is the maximum of these contributions.

}
}
