% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{profcor,matrix,NMF-method}
\alias{profcor,matrix,NMF-method}
\title{Computes the correlations between the rows of \code{x} and the basis
profiles of \code{y}.}
\usage{
\S4method{profcor}{matrix,NMF}(x, y, ...)
}
\description{
Computes the correlations between the rows of \code{x} and the basis
profiles of \code{y}.
}
