% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{profcor,NMF,matrix-method}
\alias{profcor,NMF,matrix-method}
\title{Computes the correlations between the basis profiles of \code{x} and
the rows of \code{y}.}
\usage{
\S4method{profcor}{NMF,matrix}(x, y, ...)
}
\description{
Computes the correlations between the basis profiles of \code{x} and
the rows of \code{y}.
}
