% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{profcor,NMF,missing-method}
\alias{profcor,NMF,missing-method}
\title{Computes the correlations between the basis profiles of \code{x}.}
\usage{
\S4method{profcor}{NMF,missing}(x, y, ...)
}
\description{
Computes the correlations between the basis profiles of \code{x}.
}
