% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{profcor,NMF,NMF-method}
\alias{profcor,NMF,NMF-method}
\title{Computes the correlations between the basis profiles of \code{x} and \code{y}.}
\usage{
\S4method{profcor}{NMF,NMF}(x, y, ...)
}
\description{
Computes the correlations between the basis profiles of \code{x} and \code{y}.
}
