% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{purity,factor,ANY-method}
\alias{purity,factor,ANY-method}
\title{Computes the purity on the contingency table of \code{x} and \code{y}, that is
coerced into a factor if necessary.}
\usage{
\S4method{purity}{factor,ANY}(x, y, ...)
}
\description{
Computes the purity on the contingency table of \code{x} and \code{y}, that is
coerced into a factor if necessary.
}
