% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{purity,NMFfitXn,ANY-method}
\alias{purity,NMFfitXn,ANY-method}
\title{Computes the best or mean purity across all NMF fits stored in \code{x}.}
\usage{
\S4method{purity}{NMFfitXn,ANY}(x, y, method = "best", ...)
}
\arguments{
\item{method}{a character string that specifies how the value is computed.
It may be either \code{'best'} or \code{'mean'} to compute the best or mean 
purity respectively.}
}
\description{
Computes the best or mean purity across all NMF fits stored in \code{x}.
}
