% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{purity,table,missing-method}
\alias{purity,table,missing-method}
\title{Computes the purity directly from the contingency table \code{x}}
\usage{
\S4method{purity}{table,missing}(x, y)
}
\description{
Computes the purity directly from the contingency table \code{x}
}
