% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{residuals,NMFfit-method}
\alias{residuals,NMFfit-method}
\title{Returns the residuals -- track -- between the target matrix and the NMF 
fit \code{object}.}
\usage{
\S4method{residuals}{NMFfit}(object, track = FALSE, niter = NULL, ...)
}
\arguments{
\item{track}{a logical that indicates if the complete track of residuals 
should be returned (if it has been computed during the fit), or only the last
value.}

\item{niter}{specifies the iteration number for which one wants 
to get/set/test a residual value. This argument is used only if not \code{NULL}}
}
\description{
Returns the residuals -- track -- between the target matrix and the NMF 
fit \code{object}.
}
