% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{residuals,NMFfitX-method}
\alias{residuals,NMFfitX-method}
\title{Returns the residuals achieved by the best fit object, i.e. the lowest 
residual approximation error achieved across all NMF runs.}
\usage{
\S4method{residuals}{NMFfitX}(object, ...)
}
\description{
Returns the residuals achieved by the best fit object, i.e. the lowest 
residual approximation error achieved across all NMF runs.
}
