% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmatrix.R
\docType{methods}
\name{rmatrix,ANY-method}
\alias{rmatrix,ANY-method}
\title{Default method which calls \code{rmatrix,vector} on the dimensions of \code{x}
that is assumed to be returned by a suitable \code{dim} method:
it is equivalent to \code{rmatrix(dim(x), y=NULL, ...)}.}
\usage{
\S4method{rmatrix}{ANY}(x, ...)
}
\description{
Default method which calls \code{rmatrix,vector} on the dimensions of \code{x}
that is assumed to be returned by a suitable \code{dim} method:
it is equivalent to \code{rmatrix(dim(x), y=NULL, ...)}.
}
\examples{

# random matrix of the same dimension as another matrix
x <- matrix(3,4)
dim(rmatrix(x))

}
