% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{rmatrix,NMF-method}
\alias{rmatrix,NMF-method}
\title{Returns the target matrix estimate of the NMF model \code{x}, perturbated by
adding a random matrix generated using the default method of \code{rmatrix}:
it is a equivalent to \code{fitted(x) + rmatrix(fitted(x), ...)}.}
\usage{
\S4method{rmatrix}{NMF}(x, ...)
}
\description{
This method can be used to generate random target matrices that depart from
a known NMF model to a controlled extend.
This is useful to test the robustness of NMF algorithms to the presence of
certain types of noise in the data.
}
\examples{
# generate noisy fitted target from an NMF model (the true model)
gr <- as.numeric(mapply(rep, 1:3, 3))
h <- outer(1:3, gr, '==') + 0
x <- rnmf(10, H=h)
y <- rmatrix(x)
\dontrun{
# show heatmap of the noisy target matrix: block patterns should be clear
aheatmap(y)
}
\dontshow{ stopifnot( identical(dim(y), dim(x)[1:2]) ) }

# test NMF algorithm on noisy data
# add some noise to the true model (drawn from uniform [0,1])
res <- nmf(rmatrix(x), 3)
summary(res)

# add more noise to the true model (drawn from uniform [0,10])
res <- nmf(rmatrix(x, max=10), 3)
summary(res)

}
