% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmatrix.R
\docType{methods}
\name{rmatrix,numeric-method}
\alias{rmatrix,numeric-method}
\title{Generates a random matrix of given dimensions, whose entries 
are drawn using the distribution function \code{dist}.}
\usage{
\S4method{rmatrix}{numeric}(x, y = NULL, dist = runif, byrow = FALSE,
  dimnames = NULL, ..., .rng = NULL)
}
\arguments{
\item{y}{optional specification of number of columns}

\item{dist}{a random distribution function or a numeric seed (see details of method 
\code{rmatrix,numeric})}

\item{byrow}{a logical passed in the internal call to the function 
\code{\link{matrix}}}

\item{dimnames}{\code{NULL} or a \code{list} passed in the internal call to 
the function \code{\link{matrix}}}

\item{...}{extra arguments passed to the distribution function \code{dist}, or to 
\code{rmatrix} in the case of \code{rmatrix_int}.}

\item{.rng}{any RNG seed specification supported by \code{\link[rngtools]{setRNG}}.}
}
\description{
This is the workhorse method that is eventually called by all other methods.
It returns a matrix with:
\itemize{
\item \code{x} rows and \code{y} columns if \code{y} is not missing and 
not \code{NULL};
\item dimension \code{x[1]} x \code{x[2]} if \code{x} has at least two elements;
\item dimension \code{x} (i.e. a square matrix) otherwise.
}
}
\details{
The default is to draw its entries from the standard uniform distribution using
the base function \code{\link{runif}}, but any other function that generates 
random numeric vectors of a given length may be specified in argument \code{dist}.
All arguments in \code{...} are passed to the function specified in \code{dist}.

The only requirement is that the function in \code{dist} is of the following form:

\samp{
function(n, ...){
# return vector of length n
...
}}

This is the case of all base random draw function such as \code{\link{rnorm}}, 
\code{\link{rgamma}}, etc\ldots
}
\examples{
## Generate a random matrix of a given size
rmatrix(5, 3)
\dontshow{ stopifnot( identical(dim(rmatrix(5, 3)), c(5L,3L)) ) }

## Generate a random matrix of the same dimension of a template matrix
a <- matrix(1, 3, 4)
rmatrix(a)
\dontshow{ stopifnot( identical(dim(rmatrix(a)), c(3L,4L)) ) }

## Specificy the distribution to use

# the default is uniform
a <- rmatrix(1000, 50)
\dontrun{ hist(a) }

# use normal ditribution
a <- rmatrix(1000, 50, rnorm)
\dontrun{ hist(a) }

# extra arguments can be passed to the random variate generation function 
a <- rmatrix(1000, 50, rnorm, mean=2, sd=0.5)
\dontrun{ hist(a) }

}
