% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmatrix.R
\name{rmatrix}
\alias{rmatrix}
\alias{rmatrix_int}
\title{Generating Random Matrices}
\usage{
rmatrix(x, ...)

rmatrix_int(...)
}
\arguments{
\item{x}{object from which to generate a random matrix}
}
\description{
The S4 generic \code{rmatrix} generates a random matrix from a given object.  
Methods are provided to generate matrices with entries drawn from any 
given random distribution function, e.g. \code{\link{runif}} or 
\code{\link{rnorm}}.

\code{rmatrix_int} generate an integer matrix, by changing the storage mode of a matrix
generated by \code{rmatrix}. All its arguments are directly passed to \code{rmatrix}.
}
