% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnmf.R
\docType{methods}
\name{rnmf,ANY,data.frame-method}
\alias{rnmf,ANY,data.frame-method}
\title{Shortcut for \code{rnmf(x, as.matrix(target))}.}
\usage{
\S4method{rnmf}{ANY,data.frame}(x, target, ...)
}
\description{
Shortcut for \code{rnmf(x, as.matrix(target))}.
}
