% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnmf.R
\docType{methods}
\name{rnmf,ANY,matrix-method}
\alias{rnmf,ANY,matrix-method}
\title{Generates a random NMF model compatible and consistent with a target matrix.}
\usage{
\S4method{rnmf}{ANY,matrix}(x, target, ..., dist = list(max = max(max(target,
  na.rm = TRUE), 1)), use.dimnames = TRUE)
}
\arguments{
\item{use.dimnames}{a logical that indicates whether the dimnames of the 
target matrix should be set on the returned NMF model.}
}
\description{
The entries are uniformly drawn between \code{0} and \code{max(target)}.
It is more or less a shortcut for:
\samp{ rnmf(x, dim(target), max=max(target), ...)}
}
\details{
It returns an NMF model of the same class as \code{x}.
}
\examples{
# random NMF compatible with a target matrix
x <- nmfModel(3, 10, 5)
y <- rmatrix(20, 13)
rnmf(x, y) # rank of x
rnmf(2, y) # rank 2

}
