% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnmf.R
\docType{methods}
\name{rnmf,formula,ANY-method}
\alias{rnmf,formula,ANY-method}
\title{Generate a random formula-based NMF model, using the method 
\code{\link{nmfModel,formula,ANY-method}}.}
\usage{
\S4method{rnmf}{formula,ANY}(x, target, ..., dist = runif)
}
\description{
Generate a random formula-based NMF model, using the method 
\code{\link{nmfModel,formula,ANY-method}}.
}
