% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnmf.R
\docType{methods}
\name{rnmf,NMF,missing-method}
\alias{rnmf,NMF,missing-method}
\title{Generates a random NMF model of the same dimension as another NMF model.}
\usage{
\S4method{rnmf}{NMF,missing}(x, target, ...)
}
\description{
It is a shortcut for \code{rnmf(x, nrow(x), ncol(x), ...)}, which returns
a random NMF model of the same class and dimensions as \code{x}.
}
\examples{
## random NMF from another model

a <- nmfModel(3, 100, 20)
b <- rnmf(a)
\dontshow{ stopifnot( !nmf.equal(a,b) ) }

}
