% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnmf.R
\docType{methods}
\name{rnmf,NMF,numeric-method}
\alias{rnmf,NMF,numeric-method}
\title{Generates a random NMF model of the same class and rank as another NMF model.}
\usage{
\S4method{rnmf}{NMF,numeric}(x, target, ncol = NULL, keep.names = TRUE,
  dist = runif)
}
\arguments{
\item{ncol}{single numeric value that specifies the number of columns of the
coefficient matrix. Only used when \code{target} is a single numeric value.}

\item{keep.names}{a logical that indicates if the dimension names of the 
original NMF object \code{x} should be conserved (\code{TRUE}) or discarded
(\code{FALSE}).}

\item{dist}{specification of the random distribution to use to draw the entries 
of the basis and coefficient matrices.
It may be specified as:
\itemize{

\item a \code{function} which must be a distribution function such as e.g. 
\code{\link{runif}} that is used to draw the entries of both the basis and 
coefficient matrices. It is passed in the \code{dist} argument of 
\code{\link{rmatrix}}.

\item a \code{list} of arguments that are passed internally to \code{\link{rmatrix}}, 
via \code{do.call('rmatrix', dist)}.   

\item a \code{character} string that is partially matched to \sQuote{basis} or 
\sQuote{coef}, that specifies which matrix in should be drawn randomly, the 
other remaining as in \code{x} -- unchanged.

\item a \code{list} with elements \sQuote{basis} and/or \sQuote{coef}, which 
specify the \code{dist} argument separately for the basis and coefficient 
matrix respectively.

These elements may be either a distribution function, or a list of arguments that  
are passed internally to \code{\link{rmatrix}}, via 
\code{do.call('rmatrix', dist$basis)} 
or \code{do.call('rmatrix', dist$coef)}.   
}}
}
\description{
This is the workhorse method that is eventually called by all other methods.
It generates an NMF model of the same class and rank as \code{x}, compatible with the 
dimensions specified in \code{target}, that can be a single or 2-length 
numeric vector, to specify a square or rectangular target matrix respectively.
}
\details{
The second dimension can also be passed via argument \code{ncol}, so that 
calling \code{rnmf(x, 20, 10, ...)} is equivalent to \code{rnmf(x, c(20, 10), ...)}, 
but easier to write.

The entries are uniformly drawn between \code{0} and \code{max} 
(optionally specified in \code{...}) that defaults to 1.

By default the dimnames of \code{x} are set on the returned NMF model. 
This behaviour is disabled with argument \code{keep.names=FALSE}. 
See \code{\link{nmfModel}}.
}
\examples{

## random NMF of same class and rank as another model

x <- nmfModel(3, 10, 5)
x
rnmf(x, 20) # square
rnmf(x, 20, 13)
rnmf(x, c(20, 13))

# using another distribution
rnmf(x, 20, dist=rnorm) 

# other than standard model
y <- rnmf(3, 50, 10, model='NMFns')
y
\dontshow{ stopifnot( identical(dim(y), c(50L,10L,3L)) ) }
\dontshow{ stopifnot( is(y, 'NMFns') ) }

}
