% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFOffset-class.R
\docType{methods}
\name{rnmf,NMFOffset,numeric-method}
\alias{rnmf,NMFOffset,numeric-method}
\title{Generates a random NMF model with offset, from class \code{NMFOffset}.}
\usage{
\S4method{rnmf}{NMFOffset,numeric}(x, target, ...)
}
\description{
The offset values are drawn from a uniform distribution between 0 and 
the maximum entry of the basis and coefficient matrices, which are drawn 
by the next suitable \code{\link{rnmf}} method, which is the workhorse 
method \code{rnmf,NMF,numeric}.
}
\examples{

# random NMF model with offset
x <- rnmf(2, 3, model='NMFOffset')
x
offset(x)
# from a matrix
x <- rnmf(2, rmatrix(5,3, max=10), model='NMFOffset')
offset(x)

}
