% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnmf.R
\docType{methods}
\name{rnmf,numeric,missing-method}
\alias{rnmf,numeric,missing-method}
\title{Generates a random NMF model of a given rank, with known basis and/or 
coefficient matrices.}
\usage{
\S4method{rnmf}{numeric,missing}(x, target, ..., W, H, dist = runif)
}
\description{
This methods allow to easily generate partially random NMF model, where one 
or both factors are known.
Although the later case might seems strange, it makes sense for NMF models that 
have fit extra data, other than the basis and coefficient matrices, that 
are drawn by an \code{rnmf} method defined for their own class, which should 
internally call \code{rnmf,NMF,numeric} and let it draw the basis and 
coefficient matrices.
(e.g. see \code{\linkS4class{NMFOffset}} and \code{\link{rnmf,NMFOffset,numeric-method}}).
}
\details{
Depending on whether arguments \code{W} and/or \code{H} are missing, 
this method interprets \code{x} differently:
\itemize{

\item \code{W} provided, \code{H} missing: \code{x} is taken as the number of 
columns that must be drawn to build a random coefficient matrix 
(i.e. the number of columns in the target matrix).

\item \code{W} is missing, \code{H} is provided: \code{x} is taken as the number of 
rows that must be drawn to build a random basis matrix 
(i.e. the number of rows in the target matrix).

\item both \code{W} and \code{H} are provided: \code{x} is taken as the target 
rank of the model to generate.
\item Having both \code{W} and \code{H} missing produces an error, as the 
dimension of the model cannot be determined in this case. 
}

The matrices \code{W} and \code{H} are reduced if necessary and possible 
to be consistent with this value of the rank, by the internal call to 
\code{\link{nmfModel}}.
 
All arguments in \code{...} are passed to the function \code{\link{nmfModel}} 
which is used to build an initial NMF model, that is in turn passed to 
\code{rnmf,NMF,numeric} with \code{dist=list(coef=dist)} or 
\code{dist=list(basis=dist)} when suitable.
The type of NMF model to generate can therefore be specified in argument 
\code{model} (see \code{\link{nmfModel}} for other possible arguments).

The returned NMF model, has a basis matrix equal to \code{W} (if not missing) 
and a coefficient matrix equal to \code{H} (if not missing), or drawn 
according to the specification provided in argument \code{dist} 
(see method \code{rnmf,NMF,numeric} for details on the supported values for \code{dist}).
}
\examples{
# random NMF model with known basis matrix
x <- rnmf(5, W=matrix(1:18, 6)) # 6 x 5 model with rank=3
basis(x) # fixed 
coef(x) # random

# random NMF model with known coefficient matrix
x <- rnmf(5, H=matrix(1:18, 3)) # 5 x 6 model with rank=3 
basis(x) # random
coef(x) # fixed

# random model other than standard NMF
x <- rnmf(5, H=matrix(1:18, 3), model='NMFOffset')
basis(x) # random
coef(x) # fixed
offset(x) # random

}
