% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnmf.R
\docType{methods}
\name{rnmf,numeric,numeric-method}
\alias{rnmf,numeric,numeric-method}
\title{Generates a random standard NMF model of given dimensions.}
\usage{
\S4method{rnmf}{numeric,numeric}(x, target, ncol = NULL, ..., dist = runif)
}
\arguments{
\item{ncol}{a numeric value that specifies the number 
of columns of the target matrix, fitted the NMF model.
It is used only if not missing and when argument \code{target} is a single 
numeric value.}
}
\description{
This is a shortcut for \code{rnmf(nmfModel(x, target, ncol, ...)), dist=dist)}.
It generates a standard NMF model compatible with the dimensions passed in 
\code{target}, that can be a single or 2-length numeric vector, to specify 
a square or rectangular target matrix respectively. 
See \code{\link{nmfModel}}.
}
\examples{

## random standard NMF of given dimensions

# generate a random NMF model with rank 3 that fits a 100x20 matrix  
rnmf(3, 100, 20)
\dontshow{ stopifnot( identical(dim(rnmf(3, 100, 20)), c(100L,20L,3L)) ) }
# generate a random NMF model with rank 3 that fits a 100x100 matrix
rnmf(3, 100)
\dontshow{ stopifnot( identical(dim(rnmf(3, 100)), c(100L,100L,3L)) ) }

}
