% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\docType{methods}
\name{rposneg,NMF-method}
\alias{rposneg,NMF-method}
\title{Apply \code{rposneg} to the basis matrix of an \code{\link{NMF}} object.}
\usage{
\S4method{rposneg}{NMF}(object, ...)
}
\description{
Apply \code{rposneg} to the basis matrix of an \code{\link{NMF}} object.
}
\examples{

# random mixed signed NMF model 
M <- nmfModel(rmatrix(10, 3, rnorm), rmatrix(3, 4))
# split positive and negative part
nnM <- posneg(M)
M2 <- rposneg(nnM)
identical(M, M2)
}
