% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{rss,matrix-method}
\alias{rss,matrix-method}
\title{Computes the RSS between a target matrix and its estimate \code{object},
which must be a matrix of the same dimensions as \code{target}.}
\usage{
\S4method{rss}{matrix}(object, target)
}
\arguments{
\item{target}{target matrix}
}
\description{
The RSS between a target matrix \eqn{V} and its estimate \eqn{v} is computed as:
\deqn{RSS = \sum_{i,j} (v_{ij} - V_{ij})^2}
}
\details{
Internally, the computation is performed using an optimised C++ implementation,
that is light in memory usage.
}
\examples{
# RSS bewteeen random matrices
x <- rmatrix(20,10, max=50)
y <- rmatrix(20,10, max=50)
rss(x, y)
rss(x, x + rmatrix(x, max=0.1))

}
