% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\name{rss}
\alias{rss}
\alias{evar}
\title{Residual Sum of Squares and Explained Variance}
\usage{
rss(object, ...)

evar(object, ...)
}
\arguments{
\item{object}{an R object with a suitable \code{\link{fitted}}, \code{rss} or
\code{evar} method.}

\item{...}{extra arguments to allow extension, e.g. passed to \code{rss}
in \code{evar} calls.}
}
\value{
a single numeric value
}
\description{
\code{rss} and \code{evar} are S4 generic functions that respectively computes
the Residual Sum of Squares (RSS) and explained variance achieved by a model.

The explained variance for a target \eqn{V} is computed as:
\deqn{evar = 1 - \frac{RSS}{\sum_{i,j} v_{ij}^2} }{evar = 1 - RSS/sum v_{ij}^2},
}
\details{
where RSS is the residual sum of squares.

The explained variance is usefull to compare the performance of different
models and their ability to accurately reproduce the original target matrix.
Note, however, that a possible caveat is that some models explicitly aim at
minimizing the RSS (i.e. maximizing the explained variance), while others do not.
}
