% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{runtime,NMFfit-method}
\alias{runtime,NMFfit-method}
\title{Returns the CPU time required to compute a single NMF fit.}
\usage{
\S4method{runtime}{NMFfit}(object, ...)
}
\description{
Returns the CPU time required to compute a single NMF fit.
}
