% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{runtime.all,NMFfit-method}
\alias{runtime.all,NMFfit-method}
\title{Identical to \code{runtime}, since their is a single fit.}
\usage{
\S4method{runtime.all}{NMFfit}(object, ...)
}
\description{
Identical to \code{runtime}, since their is a single fit.
}
