% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{runtime.all,NMFfitX-method}
\alias{runtime.all,NMFfitX-method}
\title{Returns the CPU time required to compute all the NMF runs.
It returns \code{NULL} if no CPU data is available.}
\usage{
\S4method{runtime.all}{NMFfitX}(object)
}
\description{
Returns the CPU time required to compute all the NMF runs.
It returns \code{NULL} if no CPU data is available.
}
