% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{featureScore}
\alias{featureScore}
\alias{extractFeatures}
\title{Feature Selection in NMF Models}
\usage{
featureScore(object, ...)

extractFeatures(object, ...)
}
\arguments{
\item{object}{an object from which scores/features are computed/extracted}

\item{...}{extra arguments to allow extension}
}
\value{
\code{featureScore} returns a numeric vector of the length the number 
of rows in \code{object} (i.e. one score per feature).

\code{extractFeatures} returns the selected features as a list of indexes,
a single integer vector or an object of the same class as \code{object} 
that only contains the selected features.
}
\description{
The function \code{featureScore} implements different methods to computes 
basis-specificity scores for each feature in the data.

The function \code{extractFeatures} implements different methods to select the 
most basis-specific features of each basis component.
}
\details{
One of the properties of Nonnegative Matrix Factorization is that is tend to 
produce sparse representation of the observed data, leading to a natural 
application to bi-clustering, that characterises groups of samples by 
a small number of features.

In NMF models, samples are grouped according to the basis 
components that contributes the most to each sample, i.e. the basis 
components that have the greatest coefficient in each column of the coefficient 
matrix (see \code{\link{predict,NMF-method}}).
Each group of samples is then characterised by a set of features selected
based on basis-specifity scores that are computed on the basis matrix.
}
\section{Feature scores}{
 
The function \code{featureScore} can compute basis-specificity scores using 
the following methods:

\describe{

\item{\sQuote{kim}}{ Method defined by \cite{KimH2007}.

The score for feature \eqn{i} is defined as: 
\deqn{S_i = 1 + \frac{1}{\log_2 k} \sum_{q=1}^k p(i,q) \log_2 p(i,q)}{
S_i = 1 + 1/log2(k) sum_q [ p(i,q) log2( p(i,q) ) ] },

where \eqn{p(i,q)} is the probability that the \eqn{i}-th feature contributes 
to basis \eqn{q}: \deqn{p(i,q) = \frac{W(i,q)}{\sum_{r=1}^k W(i,r)} }{
p(i,q) = W(i,q) / (sum_r W(i,r)) }

The feature scores are real values within the range [0,1].
The higher the feature score the more basis-specific the corresponding feature.
}

\item{\sQuote{max}}{Method defined by \cite{Carmona-Saez2006}.

The feature scores are defined as the row maximums.   
}

}
}

\section{Feature selection}{

The function \code{extractFeatures} can select features using the following 
methods:
\describe{
\item{\sQuote{kim}}{ uses \cite{KimH2007} scoring schema and
feature selection method.

The features are first scored using the function
\code{featureScore} with method \sQuote{kim}.
Then only the features that fulfil both following criteria are retained:

\itemize{
\item score greater than \eqn{\hat{\mu} + 3 \hat{\sigma}}, where \eqn{\hat{\mu}}
and \eqn{\hat{\sigma}} are the median and the median absolute deviation
(MAD) of the scores respectively;

\item the maximum contribution to a basis component is greater than the median
of all contributions (i.e. of all elements of W).
}

}

\item{\sQuote{max}}{ uses the selection method used in the \code{bioNMF} 
software package and described in \cite{Carmona-Saez2006}.

For each basis component, the features are first sorted by decreasing 
contribution.
Then, one selects only the first consecutive features whose highest 
contribution in the basis matrix is effectively on the considered basis. 
}

}
}

