% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{seed,ANY,ANY,character-method}
\alias{seed,ANY,ANY,character-method}
\title{Use the registered seeding method whose access key is \code{method}.}
\usage{
\S4method{seed}{ANY,ANY,character}(x, model, method, ...)
}
\description{
Use the registered seeding method whose access key is \code{method}.
}
