% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{seed,ANY,ANY,function-method}
\alias{seed,ANY,ANY,function-method}
\title{Seeds an NMF model using a custom seeding strategy, defined by a function.}
\usage{
\S4method{seed}{ANY,ANY,`function`}(x, model, method, name, ...)
}
\arguments{
\item{name}{optional name of the seeding method for custom seeding strategies.}
}
\description{
\code{method} must have signature \code{(x='NMFfit', y='matrix', ...)}, where 
\code{x} is the unseeded NMF model and \code{y} is the target matrix to fit.
It must return an \code{\linkS4class{NMF}} object, that contains the seeded 
NMF model.
}
