% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{seed,ANY,ANY,missing-method}
\alias{seed,ANY,ANY,missing-method}
\title{Seeds the model with the default seeding method given by 
\code{nmf.getOption('default.seed')}}
\usage{
\S4method{seed}{ANY,ANY,missing}(x, model, method, ...)
}
\description{
Seeds the model with the default seeding method given by 
\code{nmf.getOption('default.seed')}
}
