% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{seed,ANY,ANY,NULL-method}
\alias{seed,ANY,ANY,NULL-method}
\title{Use NMF method \code{'none'}.}
\usage{
\S4method{seed}{ANY,ANY,`NULL`}(x, model, method, ...)
}
\description{
Use NMF method \code{'none'}.
}
