% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{seed,ANY,list,NMFSeed-method}
\alias{seed,ANY,list,NMFSeed-method}
\title{Seed a model using the elements in \code{model} to instantiate it with 
\code{\link{nmfModel}}.}
\usage{
\S4method{seed}{ANY,list,NMFSeed}(x, model, method, ...)
}
\description{
Seed a model using the elements in \code{model} to instantiate it with 
\code{\link{nmfModel}}.
}
