% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{seed,ANY,numeric,NMFSeed-method}
\alias{seed,ANY,numeric,NMFSeed-method}
\title{Seeds a standard NMF model (i.e. of class \code{\linkS4class{NMFstd}}) of rank 
\code{model}.}
\usage{
\S4method{seed}{ANY,numeric,NMFSeed}(x, model, method, ...)
}
\description{
Seeds a standard NMF model (i.e. of class \code{\linkS4class{NMFstd}}) of rank 
\code{model}.
}
