% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{methods}
\name{seed,matrix,NMF,NMFSeed-method}
\alias{seed,matrix,NMF,NMFSeed-method}
\title{This is the workhorse method that seeds an NMF model object using a given 
seeding strategy defined by an \code{NMFSeed} object, to fit a given 
target matrix.}
\usage{
\S4method{seed}{matrix,NMF,NMFSeed}(x, model, method, rng, ...)
}
\arguments{
\item{rng}{rng setting to use. 
If not missing the RNG settings are set and restored on exit using 
\code{\link{setRNG}}. 

All arguments in \code{...} are passed to teh seeding strategy.}
}
\description{
This is the workhorse method that seeds an NMF model object using a given 
seeding strategy defined by an \code{NMFSeed} object, to fit a given 
target matrix.
}
