% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{seeding,NMFfit-method}
\alias{seeding,NMFfit-method}
\title{Returns the name of the seeding method that generated the starting point
for the NMF algorithm that fitted the NMF model \code{object}.}
\usage{
\S4method{seeding}{NMFfit}(object)
}
\description{
Returns the name of the seeding method that generated the starting point
for the NMF algorithm that fitted the NMF model \code{object}.
}
