% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{seeding,NMFfitXn-method}
\alias{seeding,NMFfitXn-method}
\title{Returns the name of the common seeding method used the computation of all fits 
stored in \code{object}}
\usage{
\S4method{seeding}{NMFfitXn}(object)
}
\description{
Since all fits are seeded using the same method, this method returns the 
name of the seeding method used for the first fit.
It returns \code{NULL} if the object is empty.
}
