% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{seqtime,NMFList-method}
\alias{seqtime,NMFList-method}
\title{Returns the CPU time that would be required to sequentially compute all NMF 
fits stored in \code{object}.}
\usage{
\S4method{seqtime}{NMFList}(object)
}
\description{
This method calls the function \code{runtime} on each fit and sum up the 
results.
It returns \code{NULL} on an empty object.
}
