% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{setupBackend}
\alias{setupBackend}
\alias{setupSharedMemory}
\alias{setupTempDirectory}
\alias{setupLibPaths}
\alias{setupRNG}
\title{Computational Setup Functions}
\usage{
setupBackend(spec, backend, optional = FALSE, verbose = FALSE)

setupSharedMemory(verbose)

setupTempDirectory(verbose, dir = getwd())

setupLibPaths(pkg = "NMF", verbose = FALSE)

setupRNG(seed, n, verbose = FALSE)
}
\arguments{
\item{spec}{target parallel specification: either \code{TRUE} or \code{FALSE},
or a single numeric value that specifies the number of cores to setup.}

\item{backend}{value from argument \code{.pbackend} of \code{nmf}.}

\item{optional}{a logical that indicates if the specification must be fully
satisfied, throwing an error if it is not, or if one can switch back to
sequential, only outputting a verbose message.}

\item{verbose}{logical or integer level of verbosity for message outputs.}

\item{dir}{path where to create the temporary directory}

\item{pkg}{package name whose path should be exported the workers.}

\item{seed}{initial RNG seed specification}

\item{n}{number of RNG seeds to generate}
}
\value{
Returns \code{FALSE} if no foreach backend is to be used, \code{NA} if the currently
registered backend is to be used, or, if this function call registered a new backend,
the previously registered backend as a \code{foreach} object, so that it can be restored
after the computation is over.
}
\description{
Functions used internally to setup the computational environment.

\code{setupBackend} sets up a foreach backend given some specifications.
}
\keyword{internals}
