\docType{methods}
\name{show,NMFarray-method}
\alias{show,NMFarray-method}
\title{Show method for objects of class \code{NMFarray}}
\usage{
  \S4method{show}{NMFarray}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
  Show method for objects of class \code{NMFarray}
}
\keyword{methods}

