\docType{methods}
\name{show,NMFfitX-method}
\alias{show,NMFfitX-method}
\title{Show method for objects of class \code{NMFfitX}}
\usage{
  \S4method{show}{NMFfitX}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
  Show method for objects of class \code{NMFfitX}
}
\keyword{methods}

