\docType{methods}
\name{show,NMFfitX1-method}
\alias{show,NMFfitX1-method}
\title{Show method for objects of class \code{NMFfitX1}}
\usage{
  \S4method{show}{NMFfitX1}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
  Show method for objects of class \code{NMFfitX1}
}
\keyword{methods}

