\docType{methods}
\name{show,NMFfitXn-method}
\alias{show,NMFfitXn-method}
\title{Show method for objects of class \code{NMFfitXn}}
\usage{
  \S4method{show}{NMFfitXn}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
  Show method for objects of class \code{NMFfitXn}
}
\keyword{methods}

