\docType{methods}
\name{show,NMFList-method}
\alias{show,NMFList-method}
\title{Show method for objects of class \code{NMFList}}
\usage{
  \S4method{show}{NMFList}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
  Show method for objects of class \code{NMFList}
}
\keyword{methods}

