\docType{methods}
\name{show,NMFOffset-method}
\alias{show,NMFOffset-method}
\title{Show method for objects of class \code{NMFOffset}}
\usage{
  \S4method{show}{NMFOffset}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
  Show method for objects of class \code{NMFOffset}
}
\keyword{methods}

