\docType{methods}
\name{show,NMFSeed-method}
\alias{show,NMFSeed-method}
\title{Show method for objects of class \code{NMFSeed}}
\usage{
  \S4method{show}{NMFSeed}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
  Show method for objects of class \code{NMFSeed}
}
\keyword{methods}

