\docType{methods}
\name{show,NMFStrategyIterative-method}
\alias{show,NMFStrategyIterative-method}
\title{Show method for objects of class \code{NMFStrategyIterative}}
\usage{
  \S4method{show}{NMFStrategyIterative}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
  Show method for objects of class
  \code{NMFStrategyIterative}
}
\keyword{methods}

