% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{show,NMF-method}
\alias{show,NMF-method}
\title{Show method for objects of class \code{NMF}}
\usage{
\S4method{show}{NMF}(object)
}
\description{
Show method for objects of class \code{NMF}
}
