% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{show,NMFfit-method}
\alias{show,NMFfit-method}
\title{Show method for objects of class \code{NMFfit}}
\usage{
\S4method{show}{NMFfit}(object)
}
\description{
Show method for objects of class \code{NMFfit}
}
