% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{show,NMFfitX-method}
\alias{show,NMFfitX-method}
\title{Show method for objects of class \code{NMFfitX}}
\usage{
\S4method{show}{NMFfitX}(object)
}
\description{
Show method for objects of class \code{NMFfitX}
}
