% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{show,NMFfitX1-method}
\alias{show,NMFfitX1-method}
\title{Show method for objects of class \code{NMFfitX1}}
\usage{
\S4method{show}{NMFfitX1}(object)
}
\description{
Show method for objects of class \code{NMFfitX1}
}
