% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{show,NMFfitXn-method}
\alias{show,NMFfitXn-method}
\title{Show method for objects of class \code{NMFfitXn}}
\usage{
\S4method{show}{NMFfitXn}(object)
}
\description{
Show method for objects of class \code{NMFfitXn}
}
