% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{show,NMFList-method}
\alias{show,NMFList-method}
\title{Show method for objects of class \code{NMFList}}
\usage{
\S4method{show}{NMFList}(object)
}
\description{
Show method for objects of class \code{NMFList}
}
