% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFns-class.R
\docType{methods}
\name{show,NMFns-method}
\alias{show,NMFns-method}
\title{Show method for objects of class \code{NMFns}}
\usage{
\S4method{show}{NMFns}(object)
}
\description{
Show method for objects of class \code{NMFns}
}
