% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFOffset-class.R
\docType{methods}
\name{show,NMFOffset-method}
\alias{show,NMFOffset-method}
\title{Show method for objects of class \code{NMFOffset}}
\usage{
\S4method{show}{NMFOffset}(object)
}
\description{
Show method for objects of class \code{NMFOffset}
}
