% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSeed-class.R
\docType{methods}
\name{show,NMFSeed-method}
\alias{show,NMFSeed-method}
\title{Show method for objects of class \code{NMFSeed}}
\usage{
\S4method{show}{NMFSeed}(object)
}
\description{
Show method for objects of class \code{NMFSeed}
}
