% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategyIterative-class.R
\docType{methods}
\name{show,NMFStrategyIterative-method}
\alias{show,NMFStrategyIterative-method}
\title{Show method for objects of class \code{NMFStrategyIterative}}
\usage{
\S4method{show}{NMFStrategyIterative}(object)
}
\description{
Show method for objects of class \code{NMFStrategyIterative}
}
