% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{sparseness,matrix-method}
\alias{sparseness,matrix-method}
\title{Computes the sparseness of a matrix as the mean sparseness of its column vectors.
It returns a single numeric value.}
\usage{
\S4method{sparseness}{matrix}(x)
}
\description{
Computes the sparseness of a matrix as the mean sparseness of its column vectors.
It returns a single numeric value.
}
