% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{sparseness,NMF-method}
\alias{sparseness,NMF-method}
\title{Compute the sparseness of an object of class \code{NMF}, as the sparseness of
the basis and coefficient matrices computed separately.}
\usage{
\S4method{sparseness}{NMF}(x)
}
\description{
It returns the two values in a numeric vector with names \sQuote{basis} and \sQuote{coef}.
}
