% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\docType{methods}
\name{sparseness,numeric-method}
\alias{sparseness,numeric-method}
\title{Base method that computes the sparseness of a numeric vector.}
\usage{
\S4method{sparseness}{numeric}(x)
}
\description{
It returns a single numeric value, computed following the definition
given in section \emph{Description}.
}
