% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawnNmfJobs.R
\name{spawnNmfJobs}
\alias{spawnNmfJobs}
\title{Non-negative matrix factorization}
\usage{
spawnNmfJobs(
  A,
  parent.dir,
  k.range = 1:10,
  repeats = 10,
  impute.prop = 0.1,
  seed = 1,
  verbose = T,
  func.scripts.path = MAIN_NMF_FUNCTIONS_PATH,
  time = "0:45:00",
  mem = "2G",
  rm.path.prefix = "/Users/lnguyen",
  overwrite.jobs = T
)
}
\arguments{
\item{A}{A numeric matrix (rows=samples, columns=features)}

\item{k.range}{Integer vector specifying the range of ranks}

\item{repeats}{Number of repeats per rank}

\item{impute.prop}{Fraction of values in `A` to set to NA for calculating performance}

\item{seed}{Random seed}

\item{verbose}{Show messages? Can be 0,1,2}
}
\value{
A list containing the factorized matrices and MSE values from rank search
}
\description{
Spawn SLURM jobs to run NNLM::nnmf() on the HPC
}
