% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFfit-class.R
\docType{methods}
\name{summary,NMFfit-method}
\alias{summary,NMFfit-method}
\title{Computes summary measures for a single fit from \code{\link{nmf}}.}
\usage{
\S4method{summary}{NMFfit}(object, ...)
}
\description{
This method adds the following measures to the measures computed by the method 
\code{summary,NMF}:
}
\details{
\describe{
\item{residuals}{Residual error as measured by the objective function associated
to the algorithm used to fit the model.}
\item{niter}{Number of iterations performed to achieve convergence of the algorithm.}
\item{cpu}{Total CPU time required for the fit.}
\item{cpu.all}{Total CPU time required for the fit. For \code{NMFfit} objects, this element is 
always equal to the value in \dQuote{cpu}, but will be different for multiple-run fits.}
\item{nrun}{Number of runs performed to fit the model. This is always equal to 1 for 
\code{NMFfit} objects, but will vary for multiple-run fits.}
}
}
\examples{
# generate a synthetic dataset with known classes: 50 features, 18 samples (5+5+8)
n <- 50; counts <- c(5, 5, 8);
V <- syntheticNMF(n, counts)
cl <- unlist(mapply(rep, 1:3, counts))

# perform default NMF with rank=2
x2 <- nmf(V, 2)
summary(x2, cl, V)
# perform default NMF with rank=2
x3 <- nmf(V, 3)
summary(x2, cl, V)

}
