% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{methods}
\name{summary,NMFfitX-method}
\alias{summary,NMFfitX-method}
\title{Computes a set of measures to help evaluate the quality of the \emph{best
fit} of the set. 
The result is similar to the result from the \code{summary} method of 
\code{NMFfit} objects. 
See \code{\linkS4class{NMF}} for details on the computed measures.  
In addition, the cophenetic correlation (\code{\link{cophcor}}) and 
\code{\link{dispersion}} coefficients of the consensus matrix are returned, 
as well as the total CPU time (\code{\link{runtime.all}}).}
\usage{
\S4method{summary}{NMFfitX}(object, ...)
}
\description{
Computes a set of measures to help evaluate the quality of the \emph{best
fit} of the set. 
The result is similar to the result from the \code{summary} method of 
\code{NMFfit} objects. 
See \code{\linkS4class{NMF}} for details on the computed measures.  
In addition, the cophenetic correlation (\code{\link{cophcor}}) and 
\code{\link{dispersion}} coefficients of the consensus matrix are returned, 
as well as the total CPU time (\code{\link{runtime.all}}).
}
