% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-terms.R
\name{bterms<-}
\alias{bterms<-}
\alias{cterms<-}
\title{Fixed Terms in NMF Models}
\usage{
bterms(object) <- value

cterms(object) <- value
}
\arguments{
\item{object}{NMF object to be updated.}

\item{value}{specification of the replacement value for fixed-terms.}
}
\description{
These functions are for internal use and should not be called by the end-user.
}
\details{
They use \code{\link{model.matrix}(~ -1 + ., data=value)} to generate suitable
term matrices.
}
\keyword{internal}
\section{Methods (by generic)}{
\itemize{
\item \code{bterms<-(object = NMFstd)}: Default method tries to coerce \code{value} into a \code{data.frame} with
\code{\link{as.data.frame}}.

\item \code{cterms<-(object = NMFstd)}: Default method tries to coerce \code{value} into a \code{data.frame} with
\code{\link{as.data.frame}}.
}}

