\name{bioc-NMF}
\alias{.atrack,ExpressionSet-method}
\alias{bioc-NMF}
\alias{featureNames,NMFfitX-method}
\alias{featureNames,NMF-method}
\alias{metagenes}
\alias{metagenes<-}
\alias{metaprofiles}
\alias{metaprofiles<-}
\alias{nmeta}
\alias{nmf,ExpressionSet,ANY,ANY-method}
\alias{nmf,matrix,ExpressionSet,ANY-method}
\alias{nmfModel,ANY,ExpressionSet-method}
\alias{nmfModel,ExpressionSet,ANY-method}
\alias{rnmf,ANY,ExpressionSet-method}
\alias{run,NMFStrategy,ExpressionSet,ANY-method}
\alias{sampleNames,NMFfitX-method}
\alias{sampleNames,NMF-method}
\alias{seed,ExpressionSet,ANY,ANY-method}
\title{Specific NMF Layer for Bioconductor}
\description{
  The package NMF provides an optional layer for working
  with common objects and functions defined in the
  Bioconductor platform.
}
\details{
  It provides: \itemize{ \item computation functions that
  support \code{ExpressionSet} objects as inputs. \item
  aliases and methods for generic functions defined and
  widely used by Bioconductor base packages. \item
  specialised visualisation methods that adapt the titles
  and legend using bioinformatics terminology. \item
  functions to link the results with annotations, etc... }
}

