\name{checkErrors}
\alias{checkErrors}
\title{Error Checks in NMF Runs}
\usage{
  checkErrors(object, element = NULL)
}
\arguments{
  \item{object}{a list of lists}

  \item{element}{name of an element of the inner lists}
}
\description{
  Auxiliary function for internal error checks in nmf results.
}
\keyword{internal}

