% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R, R/NMFSet-class.R
\name{connectivity}
\alias{connectivity}
\alias{consensus}
\title{Clustering Connectivity and Consensus Matrices}
\usage{
connectivity(object, ...)

consensus(object, ...)
}
\arguments{
\item{object}{an object with a suitable \code{\link{predict}} method.}

\item{...}{extra arguments to allow extension.
They are passed to \code{\link{predict}}, except for the \code{vector} and
\code{factor} methods.}
}
\value{
a square matrix of dimension the number of samples in the model, full
of 0s or 1s.
}
\description{
\code{connectivity} is an S4 generic that computes the connectivity matrix
based on the clustering of samples obtained from a model's \code{\link{predict}}
method.

The consensus matrix has been proposed by \cite{Brunet2004} to help 
visualising and measuring the stability of the clusters obtained by 
NMF approaches.
For objects of class \code{NMF} (e.g. results of a single NMF run, or NMF
models), the consensus matrix reduces to the connectivity matrix.
}
\details{
The connectivity matrix of a given partition of a set of samples (e.g. given
as  a cluster membership index) is the matrix \eqn{C} containing only 0 or 1
entries such that:
\deqn{C_{ij} = \left\{\begin{array}{l}
1\mbox{ if sample }i\mbox{ belongs to the same cluster as sample }j\\
0\mbox{ otherwise}
\end{array}\right..}{
C_{ij} = 1 if sample i belongs to the same cluster as sample j, 0 otherwise}
}
\seealso{
\code{\link{predict}}
}
