% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\name{consensushc}
\alias{consensushc}
\title{Hierarchical Clustering of a Consensus Matrix}
\usage{
consensushc(object, ...)
}
\arguments{
\item{object}{a matrix or an \code{NMFfitX} object, as returned by multiple 
NMF runs.}

\item{...}{extra arguments passed to next method calls}
}
\value{
an object of class \code{dendrogram} or \code{hclust} depending on the  
value of argument \code{dendrogram}.
}
\description{
The function \code{consensushc} computes the hierarchical clustering of 
a consensus matrix, using the matrix itself as a similarity matrix and 
average linkage.
It is
}
