% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\name{cophcor}
\alias{cophcor}
\title{Cophenetic Correlation Coefficient}
\usage{
cophcor(object, ...)
}
\arguments{
\item{object}{an object from which is extracted a consensus matrix.}

\item{...}{extra arguments to allow extension and passed to subsequent calls.}
}
\description{
The function \code{cophcor} computes the cophenetic correlation coefficient 
from consensus matrix \code{object}, e.g. as obtained from multiple NMF runs.
}
\details{
The cophenetic correlation coeffificient is based on the consensus matrix
(i.e. the average of connectivity matrices) and was proposed by 
\cite{Brunet2004} to measure the stability of the clusters obtained from NMF.

It is defined as the Pearson correlation between the samples' distances
induced by the consensus matrix (seen as a similarity matrix) and their
cophenetic distances from a hierachical clustering based on these very
distances (by default an average linkage is used).  
See \cite{Brunet2004}.
}
\seealso{
\code{\link{cophenetic}}
}
