% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\name{deviance}
\alias{deviance}
\alias{nmfDistance}
\title{Distances and Objective Functions}
\usage{
deviance(object, ...)

nmfDistance(method = c("", "KL", "euclidean"))
}
\value{
\code{deviance} returns a nonnegative numerical value

\code{nmfDistance} returns a function with least two arguments:
an NMF model and a matrix.
}
\description{
The NMF package defines methods for the generic \code{deviance} from the package \code{stats},
to compute approximation errors between NMF models and matrices, using a variety of
objective functions.

\code{nmfDistance} returns a function that computes the distance between an NMF model and a
compatible matrix.
}
\seealso{
Other stats: \code{\link{deviance,NMF-method}},
  \code{\link{residuals}}
}
