% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\name{dispersion}
\alias{dispersion}
\title{Dispersion of a Matrix}
\usage{
dispersion(object, ...)
}
\arguments{
\item{object}{an object from which the dispersion is computed}

\item{...}{extra arguments to allow extension}
}
\description{
Computes the dispersion coefficient of a -- consensus -- matrix
\code{object}, generally obtained from multiple NMF runs.
}
\details{
The dispersion coefficient is based on the consensus matrix (i.e. the
average of connectivity matrices) and was proposed by \cite{KimH2007} to 
measure the reproducibility of the clusters obtained from NMF.

It is defined as: 
\deqn{\rho = \sum_{i,j=1}^n 4 (C_{ij} - \frac{1}{2})^2 , }
where \eqn{n} is the total number of samples.

By construction, \eqn{0 \leq \rho \leq 1} and \eqn{\rho = 1} only for a perfect
consensus matrix, where all entries 0 or 1. 
A perfect consensus matrix is obtained only when all the connectivity matrices 
are the same, meaning that the algorithm gave the same clusters at each run.  
See \cite{KimH2007}.
}
