% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\name{fitted}
\alias{fitted}
\title{Fitted Matrix in NMF Models}
\usage{
fitted(object, ...)
}
\arguments{
\item{object}{an object that inherit from class \code{NMF}}

\item{...}{extra arguments to allow extension}
}
\value{
the target matrix estimate as fitted by the model \code{object}
}
\description{
Computes the estimated target matrix based on a given \emph{NMF} model.
The estimation depends on the underlying NMF model.
For example in the standard model \eqn{V \equiv W H}{V ~ W H}, the target matrix is
estimated by the matrix product \eqn{W H}.
In other models, the estimate may depend on extra parameters/matrix
(cf. Non-smooth NMF in \code{\link{NMFns-class}}).
}
\details{
This function is a S4 generic function imported from \link[stats]{fitted} in
the package \emph{stats}.
It is implemented as a pure virtual method for objects of class
\code{NMF}, meaning that concrete NMF models must provide a
definition for their corresponding class (i.e. sub-classes of
class \code{NMF}).
See \code{\linkS4class{NMF}} for more details.
}
