% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{registerDoBackend}
\alias{registerDoBackend}
\alias{getDoBackend}
\alias{setDoBackend}
\alias{register}
\alias{ForeachBackend}
\alias{getDoParHosts}
\alias{getDoParNHosts}
\title{Utilities and Extensions for Foreach Loops}
\usage{
registerDoBackend(object, ...)

getDoBackend()

setDoBackend(data, cleanup = FALSE)

register(x, ...)

ForeachBackend(object, ...)

getDoParHosts(object, ...)

getDoParNHosts(object)
}
\arguments{
\item{object}{specification of a foreach backend, e.g. \sQuote{SEQ}, 
\sQuote{PAR} (for doParallel), \sQuote{MPI}, etc\ldots}

\item{...}{extra arguments passed to the backend own registration function.}

\item{data}{internal data of a foreach \%dopar\% backend.}

\item{cleanup}{logical that indicates if the previous
backend's cleanup procedure should be run, \strong{before} 
setting the new backend.}

\item{x}{specification of a foreach backend}
}
\description{
\code{registerDoBackend} is a unified register function for foreach backends.

\code{getDoBackend} returns the internal data of the currently registered foreach \%dopar\% backend.

\code{setDoBackend} is identical to \code{\link[foreach]{setDoPar}}, but 
returns the internal of the previously registered backend.

\code{register} is a generic function that register objects.
It is used to as a unified interface to register foreach backends.

\code{ForeachBackend} is a factory method for foreach backend objects.

\code{getDoParHosts} is a generic function that returns the hostname of the worker nodes used by a backend.

\code{getDoParNHosts} returns the number of hosts used by a backend.
}
\keyword{internal}
