% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf-package.R, R/heatmaps.R
\name{NMF-defunct}
\alias{NMF-defunct}
\alias{metaHeatmap}
\title{Defunct Functions and Classes in the NMF Package}
\usage{
metaHeatmap(object, ...)
}
\arguments{
\item{object}{an R object}

\item{...}{other arguments}
}
\description{
Defunct Functions and Classes in the NMF Package
}
