% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\name{nmf.equal}
\alias{nmf.equal}
\title{Testing Equality of NMF Models}
\usage{
nmf.equal(x, y, ...)
}
\arguments{
\item{x}{an NMF model or an object that is associated with an NMF model, e.g.
the result from a fit with \code{\link{nmf}}.}

\item{y}{an NMF model or an object that is associated with an NMF model, e.g.
the result from a fit with \code{\link{nmf}}.}

\item{...}{extra arguments to allow extension, and passed to subsequent calls}

\item{identical}{a logical that indicates if the comparison should be made
using the function \code{\link{identical}} (\code{TRUE}) or \code{\link{all.equal}}
(\code{FALSE}). See description for method \code{nmf.equal,NMF,NMF}.}
}
\description{
The function \code{nmf.equal} tests if two NMF models are the same, i.e. they
contain -- almost -- identical data: same basis and coefficient matrices, as
well as same extra parameters.
}
\details{
\code{nmf.equal} compares two NMF models, and return \code{TRUE} iff they are
identical acording to the function \code{\link{identical}} when \code{identical=TRUE},
or equal up to some tolerance acording to the function \code{\link{all.equal}}.
This means that all data contained in the objects are compared, which includes
at least the basis and coefficient matrices, as well as the extra parameters
stored in slot \sQuote{misc}.

If extra arguments are specified in \code{...}, then the comparison is performed
using \code{\link{all.equal}}, irrespective of the value of argument \code{identical}.
}
