% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{class}
\name{NMFfitX-class}
\alias{NMFfitX-class}
\title{Virtual Class to Handle Results from Multiple Runs of NMF Algorithms}
\description{
This class defines a common interface to handle the results from multiple
runs of a single NMF algorithm, performed with the \code{\link{nmf}} method.
}
\details{
Currently, this interface is implemented by two classes,
\code{\linkS4class{NMFfitX1}} and \code{\linkS4class{NMFfitXn}}, which
respectively handle the case where only the best fit is kept, and the case
where the list of all the fits is returned.

See \code{\link{nmf}} for more details on the method arguments.
}
\section{Slots}{

\describe{
\item{\code{runtime.all}}{Object of class \code{\link[=proc.time]{proc_time}} that 
contains CPU times required to perform all the runs.}
}}

\examples{

# generate a synthetic dataset with known classes
n <- 20; counts <- c(5, 2, 3);
V <- syntheticNMF(n, counts)

# perform multiple runs of one algorithm (default is to keep only best fit)
res <- nmf(V, 3, nrun=3)
res

# plot a heatmap of the consensus matrix
\dontrun{ consensusmap(res) }

# perform multiple runs of one algorithm (keep all the fits)
res <- nmf(V, 3, nrun=3, .options='k')
res
 
}
\seealso{
Other multipleNMF: \code{\link{NMFfitX1-class}},
  \code{\link{NMFfitXn-class}}
}
