% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\name{NMFfitX}
\alias{NMFfitX}
\title{Factory Method for Multiple NMF Run Objects}
\usage{
NMFfitX(object, ...)
}
\arguments{
\item{object}{an object from which is created an \code{NMFfitX} object}

\item{...}{extra arguments used to pass values for slots}
}
\description{
Factory Method for Multiple NMF Run Objects
}
\keyword{internal}
