% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{class}
\name{NMFfitXn-class}
\alias{NMFfitXn-class}
\title{Structure for Storing All Fits from Multiple NMF Runs}
\description{
This class is used to return the result from a multiple run of a single NMF
algorithm performed with function \code{nmf} with option
\code{keep.all=TRUE} (cf. \code{\link{nmf}}).
}
\details{
It extends both classes \code{\linkS4class{NMFfitX}} and \code{list}, and
stores the result of each run (i.e. a \code{NMFfit} object) in its
\code{list} structure.

IMPORTANT NOTE: This class is designed to be \strong{read-only}, even though
all the \code{list}-methods can be used on its instances. Adding or removing
elements would most probably lead to incorrect results in subsequent calls.
Capability for concatenating and merging NMF results is for the moment only
used internally, and should be included and supported in the next release of
the package.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{standard slot that contains the S3 \code{list} object data.
See R documentation on S3/S4 classes for more details (e.g., \code{\link{setOldClass}}).}
}}

\examples{

# generate a synthetic dataset with known classes
n <- 20; counts <- c(5, 2, 3);
V <- syntheticNMF(n, counts)

# get the class factor
groups <- V$pData$Group

# perform multiple runs of one algorithm, keeping all the fits
res <- nmf(V, 3, nrun=3, .options='k') # .options=list(keep.all=TRUE) also works
res
 
summary(res)
# get more info
summary(res, target=V, class=groups)

# compute/show computational times
runtime.all(res)
seqtime(res)

# plot the consensus matrix, computed on the fly
\dontrun{ consensusmap(res, annCol=groups) }

}
\seealso{
Other multipleNMF: \code{\link{NMFfitX-class}},
  \code{\link{NMFfitX1-class}}
}
