% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\name{nmfFormals}
\alias{nmfFormals}
\alias{nmfArgs}
\title{Showing Arguments of NMF Algorithms}
\usage{
nmfFormals(x, ...)

nmfArgs(x)
}
\arguments{
\item{x}{algorithm specification}

\item{...}{extra argument to allow extension}
}
\description{
This function returns the extra arguments that can be passed
to a given NMF algorithm in call to \code{\link{nmf}}.

\code{nmfArgs} is a shortcut for \code{args(nmfWrapper(x))}, to 
display the arguments of a given NMF algorithm.
}
\examples{

# show arguments of an NMF algorithm
nmfArgs('brunet')
nmfArgs('snmf/r')
}
