% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R,
%   R/NMFStrategyFunction-class.R, R/NMFStrategyIterative-class.R
\name{nmfFormals}
\alias{nmfFormals}
\alias{nmfFormals.character}
\alias{nmfFormals.NMFStrategy}
\alias{nmfArgs}
\alias{nmfFormals.NMFStrategyFunction}
\alias{nmfFormals.NMFStrategyIterative}
\title{Showing Arguments of NMF Algorithms}
\usage{
nmfFormals(x, ...)

\method{nmfFormals}{character}(x, ...)

\method{nmfFormals}{NMFStrategy}(x, ...)

nmfArgs(x)

\method{nmfFormals}{NMFStrategyFunction}(x, ...)

\method{nmfFormals}{NMFStrategyIterative}(x, runtime = FALSE, ...)
}
\arguments{
\item{x}{algorithm specification}

\item{...}{extra argument to allow extension and passed to a suitable method.}

\item{runtime}{(internal use only) single logical that indicates if arguments should be split
between those passed internally and of passed down from the top call.}
}
\description{
This function returns the extra arguments that can be passed
to a given NMF algorithm in a call to \code{\link[=nmf]{nmf()}}.
}
\examples{

# show arguments of an NMF algorithm
nmfArgs('brunet')
nmfArgs('snmf/r')
}
\seealso{
\link{nmfWrapper}
}
\section{Methods (by generic)}{
\itemize{
\item \code{nmfFormals(character)}: Returns the extra arguments corresponfing to the \code{NMF} algorithm specified by its name
(a single string).
It is a shorcut for \code{nmfFormals(nmfAlgorithm(x), ...)}

\item \code{nmfFormals(NMFStrategy)}: Returns the extra arguments that can be passed to an algorithm encapsulated in an \code{NMFStrategy} object.

Arguments that have default values defined by the strategy are set accordingly.

\item \code{nmfFormals(NMFStrategyIterative)}: Returns the extra arguments that can be passed to the algorithm that is composed
by an \link[=NMFStrategyIterative-class]{NMFStrategyIterative} object.

These are the union of the formal arguments of all the functional steps in the algorithm:
\code{onInit}, \code{Update}, \code{Stop}, \code{onReturn}.
}}

