\name{nmfFormals}
\alias{nmfFormals}
\title{Showing Arguments of NMF Algorithms}
\usage{
  nmfFormals(x, ...)
}
\arguments{
  \item{x}{algorithm specification}

  \item{...}{extra argument to allow extension}
}
\description{
  This function returns the extra arguments that can be
  passed to a given NMF algorithm in call to
  \code{\link{nmf}}.
}

