% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\docType{class}
\name{NMFList-class}
\alias{NMFList-class}
\title{Class for Storing Heterogeneous NMF fits}
\description{
This class wraps a list of NMF fit objects, which may come from different 
runs of the function \code{\link{nmf}}, using different parameters, methods, etc..
These can be either from a single run (NMFfit) or multiple runs (NMFfitX).

Note that its definition/interface is very likely to change in the future.
}
