% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfModel.R
\name{nmfModel}
\alias{nmfModel}
\alias{nmfModels}
\title{Factory Methods NMF Models}
\usage{
nmfModel(rank, target = 0L, ...)

nmfModels(builtin.only = FALSE)
}
\arguments{
\item{rank}{specification of the target factorization rank 
(i.e. the number of components).}

\item{target}{an object that specifies the dimension of the estimated target matrix.}

\item{...}{extra arguments to allow extension, that are passed down to the 
workhorse method \code{nmfModel,numeric.numeric}, where they are used to 
initialise slots specific to the instantiating NMF model class.}

\item{builtin.only}{logical that indicates whether only built-in NMF models, 
i.e. defined within the NMF package, should be listed.}
}
\value{
an object that inherits from class \code{\linkS4class{NMF}}.

a list
}
\description{
\code{nmfModel} is a S4 generic function which provides a convenient way to 
build NMF models. 
It implements a unified interface for creating \code{NMF} objects from any 
NMF models, which is designed to resolve potential dimensions inconsistencies.

\code{nmfModels} lists all available NMF models currently defined that can be 
used to create NMF objects, i.e. -- more or less -- all S4 classes that 
inherit from class \code{\linkS4class{NMF}}.
}
\details{
All \code{nmfModel} methods return an object that inherits from class \code{NMF}, 
that is suitable for seeding NMF algorithms via arguments \code{rank} or 
\code{seed} of the \code{\link{nmf}} method, in which case the factorisation 
rank is implicitly set by the number of basis components in the seeding 
model (see \code{\link{nmf}}).

For convenience, shortcut methods and internal conversions for working on 
\code{data.frame} objects directly are implemented.
However, note that conversion of a \code{data.frame} into a \code{matrix} 
object may take some non-negligible time, for large datasets. 
If using this method or other NMF-related methods several times, consider 
converting your data \code{data.frame} object into a matrix once for good, 
when first loaded.
}
\examples{

# show all the NMF models available (i.e. the classes that inherit from class NMF)
nmfModels()
# show all the built-in NMF models available
nmfModels(builtin.only=TRUE)

}
\seealso{
Other NMF-interface: \code{\link{NMF-class}},
  \code{\link{basis}}, \code{\link{rnmf}}

Other NMF-interface: \code{\link{NMF-class}},
  \code{\link{basis}}, \code{\link{rnmf}}
}
