% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFOffset-class.R
\docType{class}
\name{NMFOffset-class}
\alias{NMFOffset-class}
\alias{show,NMFOffset-method}
\alias{initialize,NMFOffset-method}
\alias{offset,NMFOffset-method}
\alias{fitted,NMFOffset-method}
\alias{rnmf,NMFOffset,numeric-method}
\title{NMF Model - Nonnegative Matrix Factorization with Offset}
\usage{
\S4method{show}{NMFOffset}(object)

\S4method{initialize}{NMFOffset}(.Object, ..., offset)

\S4method{offset}{NMFOffset}(object)

\S4method{fitted}{NMFOffset}(object, W, H, offset = object@offset)

\S4method{rnmf}{NMFOffset,numeric}(x, target, ...)
}
\arguments{
\item{object}{an instance of class \code{NMFOffset}.}

\item{.Object}{an object of class \code{NMFOffset} that serves as a template to initialize
the object to be created.}

\item{...}{other arguments passed as object slot values.}

\item{offset}{offset vector}

\item{W}{a optional matrix to use in the computation as the basis matrix in place of
\code{basis(object)}.
It must be compatible with the coefficient matrix used in the computation.
That is, number of columns in \code{W} = number of rows in the mixture coefficient matrix \code{coef(object)} or \code{H} (if provided).}

\item{H}{a optional matrix to use in the computation as the coefficient matrix in place of
\code{coef(object)}.
It must be compatible with the basis matrix used in the computation.
That is, number of rows in \code{H} = number of columns in the basis matrix \code{basis(object)} or \code{W} (if provided).}
}
\description{
This class implements the \emph{Nonnegative Matrix Factorization with
Offset} model, required by the NMF with Offset algorithm.
}
\details{
The NMF with Offset algorithm is defined by Badea (2008) as a modification
of the euclidean based NMF algorithm from \code{Lee2001} (see section Details and
references below).
It aims at obtaining 'cleaner' factor matrices, by the introduction of an
offset matrix, explicitly modelling a feature specific baseline
-- constant across samples.
}
\section{Creating objects from the Class}{


Object of class \code{NMFOffset} can be created using the standard way with
operator \code{\link{new}}

However, as for all NMF model classes -- that extend class
\code{\linkS4class{NMF}}, objects of class \code{NMFOffset} should be
created using factory method \code{\link{nmfModel}} :

\code{new('NMFOffset')}

\code{nmfModel(model='NMFOffset')}

\code{nmfModel(model='NMFOffset', W=w, offset=rep(1, nrow(w)))}

See \code{\link{nmfModel}} for more details on how to use the factory
method.
}

\section{Initialize method}{

The initialize method for \code{NMFOffset} objects tries to correct the initial
value passed for slot \code{offset}, so that it is consistent with the dimensions
of the \code{NMF} model:
it will pad the offset vector with NA values to get the length equal to the
number of rows in the basis matrix.
}

\examples{
 
# create a completely empty NMF object
new('NMFOffset')

# create a NMF object based on random (compatible) matrices
n <- 50; r <- 3; p <- 20
w <- rmatrix(n, r) 
h <- rmatrix(r, p)
nmfModel(model='NMFOffset', W=w, H=h, offset=rep(0.5, nrow(w)))

# apply Nonsmooth NMF algorithm to a random target matrix
V <- rmatrix(n, p)
\dontrun{nmf(V, r, 'offset')}

# random NMF model with offset  
rnmf(3, 10, 5, model='NMFOffset')


# random NMF model with offset
x <- rnmf(2, 3, model='NMFOffset')
x
offset(x)
# from a matrix
x <- rnmf(2, rmatrix(5,3, max=10), model='NMFOffset')
offset(x)

}
\references{
Badea L (2008). “Extracting gene expression profiles common to colon and pancreatic adenocarcinoma using simultaneous nonnegative matrix factorization.” _Pacific Symposium on Biocomputing. Pacific Symposium on
Biocomputing_, *290*, 267-78. ISSN 1793-5091, <URL: http://www.ncbi.nlm.nih.gov/pubmed/18229692>.
}
\seealso{
Other NMF-model: 
\code{\link{NMFns-class}},
\code{\link{NMFstd-class}}
}
\concept{NMF-model}
\section{Methods (by generic)}{
\itemize{
\item \code{show(object = NMFOffset)}: Show method for objects of class \code{NMFOffset}

\item \code{offset(object = NMFOffset)}: Offsets in NMF Models with Offset

The function \code{offset} returns the offset vector from an NMF model
that has an offset, e.g. an \code{NMFOffset} model.

\item \code{fitted(object = NMFOffset)}: Computes the target matrix estimate for an NMFOffset object.

The estimate is computed as:
\deqn{ W H + offset }

\item \code{rnmf(x = NMFOffset,target = numeric)}: Generates a random NMF model with offset, from class \code{NMFOffset}.

The offset values are drawn from a uniform distribution between 0 and
the maximum entry of the basis and coefficient matrices, which are drawn
by the next suitable \code{\link{rnmf}} method, which is the workhorse
method \code{rnmf,NMF,numeric}.
}}

