% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSeed-class.R
\docType{class}
\name{NMFSeed-class}
\alias{NMFSeed-class}
\title{Base class that defines the interface for NMF seeding methods.}
\description{
This class implements a simple wrapper strategy object that defines a unified
interface to seeding methods, that are used to initialise NMF models before 
fitting them with any NMF algorithm.
}
\section{Slots}{

\describe{
\item{\code{name}}{character string giving the name of the seeding strategy}

\item{\code{method}}{workhorse function that implements the seeding strategy.
It must have signature \code{(object="NMF", x="matrix", ...)} and initialise
the NMF model \code{object} with suitable values for fitting the target 
matrix \code{x}.}
}}

