% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{class}
\name{NMFStrategy-class}
\alias{NMFStrategy-class}
\alias{show,NMFStrategy-method}
\alias{objective,NMFStrategy-method}
\alias{objective<-,NMFStrategy,character-method}
\alias{objective<-,NMFStrategy,function-method}
\alias{is.mixed}
\title{Virtual Interface for NMF Algorithms}
\usage{
\S4method{show}{NMFStrategy}(object)

\S4method{objective}{NMFStrategy}(object)

\S4method{objective}{NMFStrategy,character}(object) <- value

\S4method{objective}{NMFStrategy,`function`}(object) <- value

is.mixed(object)
}
\description{
This class partially implements the generic interface defined for general 
algorithms defined in the \pkg{NMF} package (see \code{\link{algorithmic-NMF}}).

Gets the objective function associated with an NMF algorithm.
 
It is used in \code{\link[=deviance,NMFStrategy-method]{deviance}} 
to compute the objective value for an NMF model with respect to 
a given target matrix.

Sets the objective function associated with an NMF algorithm, with a character string
that must be a registered objective function.

Sets the objective function associated with an NMF algorithm, with a function
that computes the approximation error between an NMF model and a target matrix.

\code{is.mixed} tells if an NMF algorithm works on mixed-sign data.
}
\section{Slots}{

\describe{
\item{\code{objective}}{the objective function associated with the algorithm (Frobenius, Kullback-Leibler, etc...). 
It is either an access key of a registered objective function or a function definition. 
In the latter case, the given function must have the following signature \code{(x="NMF", y="matrix")}
and return a nonnegative real value.}

\item{\code{model}}{a character string giving either the (sub)class name of the NMF-class instance used 
and returned by the strategy, or a function name.}

\item{\code{mixed}}{a logical that indicates if the algorithm works on mixed-sign data.}
}}

\keyword{internal}
