% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R,
%   R/NMFStrategyFunction-class.R, R/NMFStrategyIterative-class.R,
%   R/NMFStrategyOctave-class.R
\docType{methods}
\name{NMFStrategy}
\alias{NMFStrategy}
\alias{run,NMFStrategy,matrix,NMFfit-method}
\alias{run,NMFStrategy,matrix,NMF-method}
\alias{run,NMFStrategyFunction,matrix,NMFfit-method}
\alias{run,NMFStrategyIterative,matrix,NMFfit-method}
\alias{run,NMFStrategyIterativeX,matrix,NMFfit-method}
\alias{run,NMFStrategyOctave,matrix,NMFfit-method}
\title{Factory Method for NMFStrategy Objects}
\usage{
NMFStrategy(name, method, ...)

\S4method{run}{NMFStrategy,matrix,NMFfit}(object, y, x, ...)

\S4method{run}{NMFStrategy,matrix,NMF}(object, y, x, ...)

\S4method{run}{NMFStrategyFunction,matrix,NMFfit}(object, y, x, ...)

\S4method{run}{NMFStrategyIterative,matrix,NMFfit}(object, y, x, .stop = NULL,
  maxIter = nmf.getOption("maxIter") \%||\% 2000L, ...)

\S4method{run}{NMFStrategyIterativeX,matrix,NMFfit}(object, y, x, maxIter, ...)

\S4method{run}{NMFStrategyOctave,matrix,NMFfit}(object, y, x, ...)
}
\arguments{
\item{name}{name/key of an NMF algorithm.}

\item{method}{definition of the algorithm}

\item{...}{extra arguments passed to \code{\link{new}}.}

\item{.stop}{specification of a stopping criterion, that is used instead of the 
one associated to the NMF algorithm.
It may be specified as:
\itemize{
\item the access key of a registered stopping criterion;
\item a single integer that specifies the exact number of iterations to perform, which will 
be honoured unless a lower value is explicitly passed in argument \code{maxIter}.
\item a single numeric value that specifies the stationnarity threshold for the 
objective function, used in with \code{\link{nmf.stop.stationary}}; 
\item a function with signature \code{(object="NMFStrategy", i="integer", y="matrix", x="NMF", ...)}, 
where \code{object} is the \code{NMFStrategy} object that describes the algorithm being run, 
\code{i} is the current iteration, \code{y} is the target matrix and \code{x} is the current value of 
the NMF model.  
}}

\item{maxIter}{maximum number of iterations to perform.}
}
\description{
Creates NMFStrategy objects that wraps implementation of NMF algorithms into 
a unified interface.

It throws an error if called directly.

Method to run an NMF algorithm directly starting from a given NMF model.

This method is usually not called directly, but only via the function \code{\link{nmf}}, which 
takes care of many other details such as seeding the computation, handling RNG settings, or setting up 
parallelisation.

Runs an NMF iterative algorithm on a target matrix \code{y}.

This method is usually not called directly, but only via the function \code{\link{nmf}}, which 
takes care of many other details such as seeding the computation, handling RNG settings, 
or setting up parallel computations.
}
