% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategyFunction-class.R
\docType{class}
\name{NMFStrategyFunction-class}
\alias{NMFStrategyFunction-class}
\alias{algorithm,NMFStrategyFunction-method}
\alias{algorithm<-,NMFStrategyFunction,function-method}
\title{Interface for Single Function NMF Strategies}
\usage{
\S4method{algorithm}{NMFStrategyFunction}(object)

\S4method{algorithm}{NMFStrategyFunction,`function`}(object) <- value
}
\arguments{
\item{value}{a function that implements an \code{NMF} algorithm.
See details of its specifications in the desccription of class slot \code{algorithm}.}
}
\description{
This class implements the virtual interface \code{\link{NMFStrategy}} for
NMF algorithms that are implemented by a single workhorse R function.
}
\section{Slots}{

\describe{
\item{\code{algorithm}}{a function that implements an NMF algorithm.
It must have signature \code{(y='matrix', x='NMFfit')}, where \code{y} is the
target matrix to approximate and \code{x} is the NMF model assumed to be
seeded with an appropriate initial value -- as it is done internally by
function \code{\link{nmf}}.

Note that argument names currently do not matter, but it is recommended to
name them as specified above.

The function must return an object that is compatible with the fitted NMF problem.
Dependning on the purpose of the algorithm, this can be:
\itemize{
\item an \linkS4class{NMF} object for an algorithm that fits estimates both the basis and
mixture coefficient matrices
\item a matrix-like object that has the dimensions of either the basis or mixture coefficient matrix, for an algorithm
that performs a partial fit.
}}
}}

\section{Methods (by generic)}{
\itemize{
\item \code{run(object = NMFStrategyFunction,y = matrix,x = NMFfit)}: Runs the NMF algorithms implemented by the single R function -- and stored in slot \code{'algorithm'}
of \code{object}, on the data object \code{y}, using \code{x} as starting point.
It is equivalent to calling \code{object@algorithm(y, x, ...)}.

This method is usually not called directly, but only via the function \code{\link{nmf}}, which
takes care of many other details such as seeding the computation, handling RNG settings, or setting up
parallelisation.

\item \code{algorithm(object = NMFStrategyFunction)}: Returns the single R function that implements the NMF algorithm -- as stored in
slot \code{algorithm}.

\item \code{algorithm<-(object = NMFStrategyFunction,value = function)}: Sets the function that implements the NMF algorithm, stored in slot \code{algorithm}.
}}

