% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSet-class.R
\name{getRNG1}
\alias{getRNG1}
\alias{.getRNG}
\title{Extracting RNG Data from NMF Objects}
\usage{
getRNG1(object, ...)

.getRNG(object, ...)
}
\arguments{
\item{object}{an R object from which RNG settings can be extracted, e.g. an
integer vector containing a suitable value for \code{.Random.seed} or embedded
RNG data, e.g., in S3/S4 slot \code{rng} or \code{rng$noise}.}

\item{...}{extra arguments to allow extension and passed to a suitable S4 method
\code{.getRNG} or \code{.setRNG}.}
}
\description{
The \code{\link{nmf}} function returns objects that contain embedded RNG data, 
that can be used to exactly reproduce any computation.
These data can be extracted using dedicated methods for the S4 generics 
\code{\link[rngtools]{getRNG}} and \code{\link[rngtools]{getRNG1}}.
}
