\name{RNG-Internals}
\alias{RNG-internals}

\alias{initialize,rstream2-method}
\alias{initialize,rstream.user-method}
\alias{RNGinfo}
%\alias{RNGhooks}
\alias{RNGlib}
\alias{RNGlibs}
\alias{RNGprovider}
\alias{RNGrecovery}
\alias{RNGseed}
\alias{RNGseq}
%\alias{RNGwrap}

\alias{setRNG}
\alias{setRNG-methods}
\alias{setRNG,ANY-method}
\alias{setRNG,character-method}
\alias{setRNG,numeric-method}
%\alias{setRNG,rstream-method}

\title{Undocumented Internal Methods and Functions for RNG Handling}

\description{
The functions documented here used internally by the package NMF to manage RNG 
settings.
}

\seealso{
	\code{\link[NMF]{getRNG-methods}} and the package's vignette
}

\author{Renaud Gaujoux}
\keyword{internal}
