% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{seed}
\alias{seed}
\title{Interface for NMF Seeding Methods}
\usage{
seed(x, model, method, ...)
}
\arguments{
\item{x}{target matrix one wants to approximate with NMF}

\item{model}{specification of the NMF model, e.g., the factorization rank.}

\item{method}{specification of a seeding method.
See each method for details on the supported formats.}

\item{...}{extra to allow extensions and passed down to the actual seeding method.}
}
\value{
an \code{\linkS4class{NMFfit}} object.
}
\description{
The function \code{seed} provides a single interface for calling all seeding 
methods used to initialise NMF computations.
These methods at least set the basis and coefficient matrices of the initial 
\code{object} to valid nonnegative matrices.
They will be used as a starting point by any NMF algorithm that accept 
initialisation.

IMPORTANT: this interface is still considered experimental and is subject 
to changes in future release.
}
