% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSeed-class.R, R/registry-seed.R
\name{NMFSeed}
\alias{NMFSeed}
\alias{setNMFSeed}
\alias{removeNMFSeed}
\title{\code{NMFSeed} is a constructor method that instantiate 
\code{\linkS4class{NMFSeed}} objects.}
\usage{
NMFSeed(key, method, ...)

setNMFSeed(..., overwrite = isLoadingNamespace(), verbose = TRUE)

removeNMFSeed(name, ...)
}
\arguments{
\item{key}{access key as a single character string}

\item{method}{specification of the seeding method, as a function that takes 
at least the following arguments:
\describe{
\item{object}{uninitialised/empty NMF model, i.e. that it has 0 rows and 
columns, but has already the rank requested in the call to \code{\link{nmf}} 
or \code{\link{seed}}.}
\item{x}{target matrix}
\item{...}{extra arguments}
}}

\item{...}{arguments passed to \code{NMFSeed} and used to initialise slots
in the \code{\linkS4class{NMFSeed}} object, or to \code{\link[pkgmaker]{pkgreg_remove}}.}

\item{overwrite}{logical that indicates if any existing NMF method with the 
same name should be overwritten (\code{TRUE}) or not (\code{FALSE}), 
in which case an error is thrown.}

\item{verbose}{a logical that indicates if information about the registration 
should be printed (\code{TRUE}) or not (\code{FALSE}).}

\item{name}{name of the seeding method.}
}
\description{
\code{NMFSeed} is a constructor method that instantiate 
\code{\linkS4class{NMFSeed}} objects.

NMF seeding methods are registered via the function \code{setNMFSeed}, which
stores them as \code{\linkS4class{NMFSeed}} objects in a dedicated registry.

\code{removeNMFSeed} removes an NMF seeding method from the registry.
}
