% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF-class.R
\name{sparseness}
\alias{sparseness}
\title{Sparseness}
\usage{
sparseness(x, ...)
}
\arguments{
\item{x}{an object whose sparseness is computed.}

\item{...}{extra arguments to allow extension}
}
\value{
usually a single numeric value -- in [0,1], or a numeric vector.
See each method for more details.
}
\description{
Generic function that computes the \emph{sparseness} of an object, as defined
by \cite{Hoyer2004}.
The sparseness quantifies how much energy of a vector is packed into only few components.
}
\details{
In \cite{Hoyer2004}, the sparseness is defined for a real vector \eqn{x} as:
\deqn{Sparseness(x) = \frac{\sqrt{n} - \frac{\sum |x_i|}{\sqrt{\sum x_i^2}}}{\sqrt{n}-1}}{
(srqt(n) - ||x||_1 / ||x||_2) / (sqrt(n) - 1)}

, where \eqn{n} is the length of \eqn{x}.

The sparseness is a real number in \eqn{[0,1]}.
It is equal to 1 if and only if \code{x} contains a single nonzero component,
and is equal to 0 if and only if all components of \code{x} are equal.
It interpolates smoothly between these two extreme values.
The closer to 1 is the sparseness the sparser is the vector.

The basic definition is for a \code{numeric} vector, and is extended for matrices as the
mean sparseness of its column vectors.
}
\seealso{
Other assess: \code{\link{purity}}
}
