% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFStrategy-class.R
\docType{class}
\name{Strategy-class}
\alias{Strategy-class}
\alias{name}
\alias{name<-}
\title{Generic Strategy Class}
\usage{
name(object, ...)

name(object, ...) <- value
}
\arguments{
\item{object}{an R object with a defined \code{name} method}

\item{...}{extra arguments to allow extension}

\item{value}{replacement value}
}
\description{
This class defines a common interface for generic algorithm strategies 
(e.g., \code{\linkS4class{NMFStrategy}}).

\code{name} and \code{name<-} gets and sets the name associated with an object.
In the case of \code{Strategy} objects it is the the name of the algorithm.
}
\section{Slots}{

\describe{
\item{\code{name}}{character string giving the name of the algorithm}

\item{\code{package}}{name of the package that defined the strategy.}

\item{\code{defaults}}{default values for some of the algorithm's arguments.}
}}

\keyword{internal}
