% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-terms.R
\name{bterms<-}
\alias{bterms<-}
\alias{cterms<-}
\title{Fixed Terms in NMF Models}
\usage{
bterms(object) <- value

cterms(object) <- value
}
\arguments{
\item{object}{NMF object to be updated.}

\item{value}{specification of the replacement value for fixed-terms.}
}
\description{
These functions are for internal use and should not be called by the end-user.

\code{cterms<-} sets fixed coefficient terms or indexes and should only be 
called on a newly created NMF object, i.e. in the constructor/factory generic 
\code{\link{nmfModel}}.
}
\details{
They use \code{\link{model.matrix}(~ -1 + ., data=value)} to generate suitable
term matrices.
}
\keyword{internal}
