% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-terms.R, R/NMFfit-class.R,
%   R/NMFSet-class.R
\name{ibterms}
\alias{ibterms}
\alias{ibterms,NMF-method}
\alias{ibterms,NMFstd-method}
\alias{icterms}
\alias{icterms,NMF-method}
\alias{icterms,NMFstd-method}
\alias{iterms}
\alias{nterms}
\alias{nbterms}
\alias{ncterms}
\alias{bterms}
\alias{cterms}
\alias{ibasis}
\alias{icoef}
\alias{ibterms,NMFfit-method}
\alias{icterms,NMFfit-method}
\alias{ibterms,NMFfitX-method}
\title{Fixed Terms in NMF Models}
\usage{
ibterms(object, ...)

\S4method{ibterms}{NMF}(object, ...)

\S4method{ibterms}{NMFstd}(object)

icterms(object, ...)

\S4method{icterms}{NMF}(object, ...)

\S4method{icterms}{NMFstd}(object)

iterms(object, ...)

nterms(object)

nbterms(object)

ncterms(object)

bterms(object)

cterms(object)

ibasis(object, ...)

icoef(object, ...)

\S4method{ibterms}{NMFfit}(object)

\S4method{icterms}{NMFfit}(object)

\S4method{ibterms}{NMFfitX}(object)

\S4method{icterms}{NMFfit}(object)
}
\arguments{
\item{object}{NMF object}

\item{...}{extra parameters to allow extension (currently not used)}
}
\description{
Formula-based NMF models may contain fixed basis and/or coefficient terms.
The functions documented here provide access to these data, which are
read-only and defined when the model object is instantiated
(e.g., see \code{\link[=nmfModel,formula,ANY-method]{nmfModel,formula-method}}).

\code{ibterms}, \code{icterms} and \code{iterms} respectively return the
indexes of the fixed basis terms, the fixed coefficient terms and all fixed
terms, within the basis and/or coefficient matrix of an NMF model.
}
\section{Methods (by generic)}{
\bold{ibterms:}

\itemize{
\item \code{ibterms(object = NMF)}: Default pure virtual method that ensure a method is defined for concrete
NMF model classes.

\item \code{ibterms(object = NMFstd)}: Method for standard NMF models, which returns the integer vector that is
stored in slot \code{ibterms} when a formula-based NMF model is instantiated.

\item \code{ibterms(object = NMFfit)}: Method for single NMF fit objects, which returns the indexes of fixed
basis terms from the fitted model.

\item \code{ibterms(object = NMFfitX)}: Method for multiple NMF fit objects, which returns the indexes of fixed basis
terms from the best fitted model.
}


\bold{icterms:}

\itemize{
\item \code{icterms(object = NMF)}: Default pure virtual method that ensure a method is defined for concrete
NMF model classes.

\item \code{icterms(object = NMFstd)}: Method for standard NMF models, which returns the integer vector that is
stored in slot \code{icterms} when a formula-based NMF model is instantiated.

\item \code{icterms(object = NMFfit)}: Method for single NMF fit objects, which returns the indexes of fixed
coefficient terms from the fitted model.

\item \code{icterms(object = NMFfit)}: Method for multiple NMF fit objects, which returns the indexes of fixed
coefficient terms from the best fitted model.
}

}

